Bug 1349416 - Hide HTMLTooltip on mouseup rather than on click;r=birtles
A click event is only fired if the element on which mousedown and mouseup
happened are the same. If this is not the case, the HTMLTooltip will not
be able to hide itself.

Listening to mouseup makes the behavior more consistent but forces the
hide() method to be always asynchronous, while before it was only
asynchronous for tooltips using the XULWrapper.

Depends on D2660

Differential Revision: https://phabricator.services.mozilla.com/D2661

UltraBlame original commit: c4f8eccf14dcdc14dee9aa597c8a42ed31c75555