Bug 1479454 [wpt PR 12229] - Create a dictionary of {item_type: ManifestItemClass}, a=testonly
Automatic update from web-platform-testsCreate a dictionary of {item_type: ManifestItemClass}

Sometimes we need to know all the types of test that exist. Typically
this has either been done by hardcoding a list or using some custom
approach to finding all the subclasses of the original. Instead it
makes sense to have the classes register themselves on creation so
that we can simply look in the registry for the types.

To do this we use a cusom metaclass that first constucts the class
object and then adds it to the registry.

--

wpt-commits: 76d1afcb887b5911e0151ff0be991a267d70d0a1
wpt-pr: 12229

UltraBlame original commit: e54c63c48323172fed44132f960b9f997787ea62