Bug 1479521 - Remove browser_jsterm_no_autocompletion_on_defined_variables.js; r=Honza.
The test wasn't doing what it was supposed to do (there was definitely
an autocomplete text displayed, we were not waiting properly for it.).
Also, the original issue with this was that that an unwanted token was
entered instead of the one the user wanted (user generated). Maybe at
that time the autocomplete did not contained user defined variables,
but now we do, and we are testing that in other tests.

Depends on D2824

Differential Revision: https://phabricator.services.mozilla.com/D2825

UltraBlame original commit: 67fc7d23d665f175f6f61448a5afbbb9f712395b