Bug 1482826 - Fix DAMP autocomplete test; r=Honza.
We can now simply call setInputValue and the autocompletion
will happen. Since we are forcing the call to jsterm.complete,
there were 2 calls made to the server, making the measurements
in the test erroneous.
This also revealed a race in setInputValue: the text was
set by codeMirror before the cursor was actually moved. Which
means we were sending an erroneous autocompletion query to the
server.
We use codeMirror.operation to tell codeMirror to both set the
text and the cursor in a single operation.

Differential Revision: https://phabricator.services.mozilla.com/D3192

UltraBlame original commit: d2d335ca743b25de0e9b6a103828d4ce0a20aab1