Bug 1484501 - Fix negated disabled selector in extension.css; r=bwinton
Everywhere else in this file we match on .disabled or :not(.disabled) but in two
instances we have :not(disabled). That's never going to match since it's already
combined with an `input` or `textarea` type selector (and there's no <disabled>
element).

Differential Revision: https://phabricator.services.mozilla.com/D3743

UltraBlame original commit: feaff184f7758a2aecd7ba36d274759dfac49208