Bug 1484773 - Containing shadow could be null after unlink. r=smaug
Bug 1483882 removed the line that used to remove the SHADOW_TREE bit during
unbind. I missed this caller which still assumed that `IsInShadowTree() ->
GetContainingShadow()`.

That means that before that patch it was going through
nsIDocument::RemoveStyleSheet, which looks pretty wrong to me.

Fix it so that we don't remove the stylesheet (we weren't doing it before
anyway), but we don't crash.

Differential Revision: https://phabricator.services.mozilla.com/D3831

UltraBlame original commit: 1598dc35d16af3bcb77e855190b47df1992dc6e3