Bug 1481670 - Initialize some stack variables to make valgrind happier. r=jandem
clang likes to transform conditions like:

  if (A && B)

into

  if (B && A)

where B might be undefined when A is true (or variants of the above with
! and ||, or split between multiple ifs). This doesn't go well with
valgrind, although it's valid (albeit seemingly convoluted).

In both cases fixed here, the condition A involves calling a function
with a pointer to a variable on the stack that may or may not be filled
by the function call, and the condition B checking the value of said
variable.

UltraBlame original commit: 0765d3f09d9a40a1569b0179fa5af297a3f8512a