Bug 1486166 - Set the tooltip text on the parent of the urlbar input. r=dao
The reason bug 1440682 broke this was because I moved all the code to
nsXULElement. However, there was a way for non-XUL elements to get XUL tooltips
before that, which was via the RestyleManager mechanism to handle attribute
mutations.

So the behavior before that patch was that non-XUL elements that got the
attribute dynamically added or removed before that patch got their tooltips,
like the HTML input in the toolbar, but if you specified the attributes
statically in the markup, or while the element was somehow outside of the
document or what not, it would never work.

Given that, this looks completely unintentional, and the fact that this ever
worked was a bit lucky.

Chances are we eventually want tooltip support for HTML elements in chrome
documents, but it is pretty likely that we want to use the HTML tooltip
infrastructure instead of nsXULTooltipListener, which is kind of an odd thing.

Thus, for now patch the code so that it sets it on the container of the <input>,
which is a XUL box that takes the same space as the <input>, instead of moving
all the XUL tooltip support to work on non-XUL elements.

Also, while at it, remove references to inputtooltiptext, since I didn't find a
single reference in the code that would set this attribute ever.

Differential Revision: https://phabricator.services.mozilla.com/D4294

UltraBlame original commit: 20a43066e367897769f5c155b1daba955cfaeea7