Bug 1487027 - Don't care about the placeholder overflow events. r=dao
Filed bug 1487036 for a fix that would avoid triggering events for the
placeholder when it's not showing, though realistically I'm not going to have
cycles to fix it.

Probably moving the <input> to Shadow DOM would allow such a solution, but it's
not clear we want to do that just yet at least.

Differential Revision: https://phabricator.services.mozilla.com/D4525

UltraBlame original commit: 380cf66149e2247d83a19d2ec66175b0da648146