Bug 1462850 [wpt PR 11082] - [css-fonts] Avoid race condition, a=testonly
Automatic update from web-platform-tests[css-fonts] Avoid race condition

The modified test primes the document's font source by declaring a
`font-face` rule and deferring tests until the font has been loaded. The
sub-tests are expressed in terms of a FontFace which is dynamically
generated via a `<style>` tag but which shares the same URL.

Prior to this patch, the test assumed that the declaration of the second
FontFace would be reflected in a synchronously-triggered reflow (via
access of the `offsetWidth` property). This is not guaranteed because
each parse of a `<style>` element creates a new FontFace object [1] which
must be loaded [2] with the potentially CORS-enabled fetch method of the
HTML specification [3].

Practically speaking, this caused the Firefox and Chrome browsers to
consistently fail the sub-tests.

Update the test to define a unique FontFace for every subtest and to
explicitly wait for it to be available for rendering. Annotate the test
as "long" to accomodate the additional time spent waiting.

[1] > A CSS font-face rule automatically defines a corresponding
    > FontFace object, which is automatically placed in the document’s
    > font source when the rule is parsed.

    https://www.w3.org/TR/css-font-loading-3/#font-face-css-connection

[2] > When a user-agent needs to load a font face, it must do so by
    > calling the load() method of the corresponding FontFace object.

    https://www.w3.org/TR/css-font-loading-3/#font-face-set-css

[3] > For font loads, user agents must use the potentially CORS-enabled
    > fetch method defined by the [HTML5] specification for URL's
    > defined within font-face rules.

    https://www.w3.org/TR/css-fonts-3/#font-fetching-requirements

--
[css-fonts] Correct typo

--

wpt-commits: 3559441249defd6b9045d10de0453e47a144d034, d829f205b7ac265c78a45ba75863baa404d4ce72
wpt-pr: 11082

UltraBlame original commit: f2058e8755b321e90c1e9c87fe1f4f5eda2e595f