Bug 1480144: Let Windows Restart Manager ride the trains r=jmathies
The plan is to study the "restart manager" in 62 with a pref flip,
and if that goes well enable it on release in 63. If it is to be
enabled for the whole population in 63 it needs to ride the trains,
so we'll let it be enabled by default as 63 goes to beta. There will
be time to uplift a backout so it won't reach the release population,
if the 62 study indicates problems.

Differential Revision: https://phabricator.services.mozilla.com/D2620

UltraBlame original commit: fafc1900d5dfeb7e94d7a39982d0603fd3c8f4fb