Bug 1487327 - Custom section for opting in to GC feature work.  r=luke
We introduce a new numbered section which must be the first section in
the .wasm file if it is present at all.  This section, GcFeatureOptIn,
carries the version number of the GC feature the module wants to opt
in to.

During validation we signal an error if we can't satisfy the
requirement; and we signal errors for all the GC feature aspects
(anyref, ref, struct, and all the ref-type instructions) if no opt-in
section was present.

This patch does not affect how we generate code for stubs; that code
is controlled by the --wasm-gc flag, as it is module-independent.

This patch also does not change how we choose the compiler; that is
still based on --wasm-gc.  But once --wasm-gc disappears we will be
using the opt-in section to control compiler selection until Ion can
support the GC feature.

Currently the only supported GC feature version is 1.  As the engine
evolves to accomodate new versions we will accept old version numbers
provided the engine remains completely backward compatible.  (We can
also provide version-triggered backward compatibility but I question
the utility.)

UltraBlame original commit: c91d997687bf454984288ff865e05aa6a4e680ea