Bug 1485151 - Part 2 - Switch our whitelisting of StrictMode violations to try-with-resources. r=jchen
In my opinion this is an altogether nicer way of setting and resetting temporary
StrictMode exceptions. The only fly in the ointment is that this causes a
warning about a technically unused variable that subsequently needs suppressing.

Differential Revision: https://phabricator.services.mozilla.com/D4889

UltraBlame original commit: 9f53dc79984c400b17a8e3ddfa23df67ff7299ca