Bug 1478582 - Bug 14785825 - Move address bar model tokenization to a module. r=adw
This is a first stab at the new tokenizer.
It's not expected to be perfect yet, but good enough to be modified and replace the existing code in unifiedComplete with just a few modifications.
It's mostly intended to start setting up a code and tests structure.

Differential Revision: https://phabricator.services.mozilla.com/D2838

UltraBlame original commit: c291d6d950222a55e9c717993c95daad52c02bd8