Bug 1488520 - Continue searching for a moof until a valid one is found. r=alwu
It is possible for audio and videos moofs to be interleaved in the stream. If we were to stop on an invalid moof or a moof whose track ID doesn't match, the track would end up not playing.

Differential Revision: https://phabricator.services.mozilla.com/D5157

UltraBlame original commit: b28c1da957075c15f306c05ff94332210b09b277