Bug 1024913 - Move jsterm auto-focus handling into App.js; r=bgrins.
The App component takes the whole document space, so we can move the
code that listen for every click in the console UI to decide if the
jsterm should be focused or not.
This feels more like the React way, and will also provide us better
access to the store and the props when we need them in this focus
handling (like this will be the case for hiding the reverse search UI).

Depends on D5150

Differential Revision: https://phabricator.services.mozilla.com/D5152

UltraBlame original commit: 28becf26d84b6e6f767d71b3252d7643a75a4ef7