Bug 1282098 - Pull probe parsers out of m-c and into a separate package. r=chutten
This patch moves the parsers' code to python_mozparsers, which is
the definitive name of the package on pypi. The data generation scripts
are adjusted accordingly. The README and setup files are added to
enable publishing on pypi.

Differential Revision: https://phabricator.services.mozilla.com/D5043

UltraBlame original commit: 28fb503aacbc3d30c3c35bf687c3d158a04a125b