Bug 1487085 - Allow to set cookie with a null channel in child process, r=valentin
In the current implmentation of CookieServiceChild::SetCookieString, pass a null channel will crash the child process. This is because we call aChannel->GetURI() without checking if aChannel is null.
However, set cookie with a null channel is possible in non-e10s mode. To make sure the behavior to be consistent in both non-e10s and e10s mode, we have to pass an empty URIParams in child process.

Differential Revision: https://phabricator.services.mozilla.com/D5432

UltraBlame original commit: 0d5b78130661fd02d9906e41f30c69ffdfca650b