Bug 1035774 - Introduce functions to get the capabilities for the primary pointer and for all available pointers on Android. r=jchen
There is no API that we can tell which input device is the primary pointer
device on Android, so we assume that the first pointer-type input device
in the list of InputDevice.getDevice is the primary one.

Test results on L-01F (a variant of LG-G2):
 - Normal boot
   The touchscreen is the primary device
   - Plugged a mouse
      The touchscreen is the primary device

 - Boot with a plugged mouse
   The mouse is the primary device
   - Plugged off the mouse
      The touchscreen is now the primary device
   - Plugged the mouse again
      The touchscreen is the primary device

Classification in getPointerCapablities refered to chrome impelementation [1].
In addition to the classification, we add SOURCE_JOYSTICK as a 'COARSE' and
'HOVER_CAPABLE' device.

[1] https://chromium.googlesource.com/chromium/src/+/d4c8983688edd2725a5bfb4a9538047f7a2efa27/ui/android/java/src/org/chromium/ui/base/TouchDevice.java#72

Differential Revision: https://phabricator.services.mozilla.com/D3302

UltraBlame original commit: ad515d6f27b365b2907a7c3fb1f04dcd070a9d1e