Bug 1404992 - Recreate the VideoConduit's video adapter when setting the send codec config. r=dminor
This enables us to configure it for cropping to a certain resolution alignment in a future patch.
For instance with simulcast, so we don't have to skip a layer because it is impossible to scale
the highest layer any further without losing the aspect ratio.

UltraBlame original commit: 9f177888ae65d03e7e0bc281180e4efe1d8ac93b