bug 1490674: audioipc: relax log crate dependency version; r=kamidphish
By relaxing the version dependency on the log crate we are able to
get rid of duplicate vendored crates.

This particular version number, combined with rsdparsa depending
on "*" (anything) meant that a change to this version number in
audioipc would have reprecussions on the dependencies of rsdparsa.
I will address the wildcard version in a separate patch on rsdparsa.

UltraBlame original commit: 7b01b1ef8e8a8f7354210d5d285b0110313a0dc2