Bug 1478505 - Run test_mq_prefers_reduced_motion_dynamic.html on Android. r=jchen
To make the test work properly, there needs two functions.  One is just setting
a value for testing and sending the notificaiton for the value changes, the
other one is to reset the state for testing.

The test itself will be introduced in bug 1486971, especially in
https://phabricator.services.mozilla.com/D5004

Depends on D5503

Differential Revision: https://phabricator.services.mozilla.com/D5504

UltraBlame original commit: 759d828c5c2d519540fd484b745b6cf70e0ab91a