bug 1488593 - move family safety root detection to background thread r=dipen,jcj
A previous patch in this bug made the incorrect assumption that we had disabled
the family safety root detection/importing feature by default. In reality, we
enabled it by default in bug 1282871.

In bug 1487258 we moved enterprise root loading to a background thread so as to
not block the main thread. This patch does the same with the family safety
feature.

Differential Revision: https://phabricator.services.mozilla.com/D5484

UltraBlame original commit: 83ca9766b8fe320bbbad6ea0e5b87aee24e9f78a