Bug 1491253 - Link the UrlbarController with the input and the view. r=dao
This links together the controller with the input and view. It moves the fake results generation to the controller. As a result, there is now an input -> controller -> view round-trip in place.

Also hooks up the input with determining if we're in private browsing mode or not.

Differential Revision: https://phabricator.services.mozilla.com/D6261

UltraBlame original commit: 861fdf8eabf24b6b7acc616b544ef5db3f31eaa0