Bug 1491787 - Use document rather than event target to check fullscreen state in Android's browser.js. r=snorp
The content of e.target has been changed in bug 1375319 from document
to the fullscreen element, so the current code makes it never dispatch
DOMFullScreen:Start, so the Java side would not be aware of such state.

This code was initially added in bug 859683 for supporting scrolling
when the fullscreen element is the document element. Since then, it
seems that the related code had been refactored, and the only usage
of rootElement was removed in bug 1416330.

Because of that, it is no longer necessary to use event target, and
we can just use document directly, which would fix this issue.

Differential Revision: https://phabricator.services.mozilla.com/D6221

UltraBlame original commit: 9d576eb300e650b9e50fc197c3d2027291d10c29