Bug 1491864 - Fix preserve-3d in presence of anon boxes. r=mattwoodrow
In the button case we have a ::-moz-button-content pseudo-element, but this is
also an issue for tables and such.

These are supposed to be implementation details, so avoid looking at them for
preserve-3d.

I don't know how I didn't think of this on the regressing bug.

Differential Revision: https://phabricator.services.mozilla.com/D6131

UltraBlame original commit: 236d827dc1c5a6a25afa6486f86e50ad2725fb26