Bug 1490371 - Remove function names in generated gcda files r=marco
Function names in gcda are just here to check that they match the ones we've in gcno: this is an extra check since it already exists a function checksum for this purpose.
So the size of gcda and the time to generate them will decrease.

Differential Revision: https://phabricator.services.mozilla.com/D6413

UltraBlame original commit: 4d3cd0ab72776e5940619b2589cf14e3b29c4f5c