Bug 1486959 - Remove bogus assertion in ContentClient. r=nical
It's possible to trigger this assertion with asyncPaint = true and
bufferRotation != (0, 0). In this case though everything should work fine,
because we didn't start a new paint task, and we should be able to copy over
buffers. So we should be able to remove this assertion.

UltraBlame original commit: 1330a9eb3da56d783aefabe063764eb55e9adadc