Bug 1405370 - [marionette] Add missing keys to virtual keyboard lookup table. r=ato
Until now we only supported the internal Unicode to virtual
key mapping at the standard and left location of the keyboard.
This patch adds all the remaining special keys which are
located at the right side and in the numpad area.

This mapping table is somewhat equivalent to the normalized
key translation, but is using the "VK_" prefix.

UltraBlame original commit: 3125cc183560de4287dd3e2cdf4cb9edcc39eb7f