Bug 1127855 - Handle title highlighting properly in all locales r=nalexander
This patch uses a case-insensitive matcher to highlight the title of a history
entry that's been typed by the user. Previously the matching substring was
calculated manually as lowercase assuming that it's representation would have
the same number of characters as the original mixed case. In some locales
howerver this assumption is wrong leading to out-of-bound exceptions when
highlighting part of the title.

Differential Revision: https://phabricator.services.mozilla.com/D6661

UltraBlame original commit: fbc27c2a3b88e1a66b7936cc7b0d59d9e9aee0ed