Bug 1485676 - Adapt gDevTools API to new async forTab and also adapt its callsites. r=yulia
Summary:
When switching to async, it is important to catch exception or register a rejection handler
so that errors keep being logged.
So in this patch I'm catching exception in a couple of important codepath.

Depends On D4541

Reviewers: yulia!

Tags: #secure-revision

Bug #: 1485676

Differential Revision: https://phabricator.services.mozilla.com/D4542

MozReview-Commit-ID: IDPJVkAPbTs

UltraBlame original commit: cd4a03a48a863c37e930e65440260357f22add58