Bug 1491724 - Search engine keyword should be highlighted only when the search will actually be performed on the corresponding search engine r=mak
My reasoning in https://bugzilla.mozilla.org/show_bug.cgi?id=1484737#c3 was wrong. I assumed that if a result started with a search alias then it had to be a search alias result, but that's clearly not true.

This is such a pain to get right, and all the messy logic in this patch reflects that.

Differential Revision: https://phabricator.services.mozilla.com/D6528

UltraBlame original commit: 7e5bf501404a0e6d8f28e7255a1ee49b0d05bacb