Bug 1389390 - Add a test to verify a persisted origin won't be evicted; r=janv
Bug 1298392 introduced persist() to QuotaManager. This patch intends to verify
that normally the oldest origin will be evicted if the global limit is reached,
but if the oldest origin is persisted, then it won't be evicted.

UltraBlame original commit: 3ed04fa44d07ed2a7c8734f24e0d198c6002b743