Bug 1489834 - Fix the Send Tab to Device page action title and tooltip r=jaws
Bug 1470555 moved the title from a dtd to a properties file, so we can and should now set the action's title directly when the action is created in PageActions.jsm.  But it's a PluralForm string, so we need to jump through that hoop.

Differential Revision: https://phabricator.services.mozilla.com/D5721

UltraBlame original commit: 2fab03252db7e57d93b5c38f53d9aa3db0b1c16e