Bug 1496055 - Part 1: Remove outdated checks for StyleRuleActor.modifySelectorUnmatched; r=gl
Since Firefox 41 it has been possible to add a rule with a selector
which does not match an element. The checks of the existence of this
functionality are outdated and can be safely removed.

No supported old versions of Firefox make use of the previous
functionality anymore.

MozReview-Commit-ID: D5ZP09Ckeqd

Differential Revision: https://phabricator.services.mozilla.com/D7597

UltraBlame original commit: 6dbab52dffb7bf83935c878b20f06621c5327eeb