Bug 1495814 - Remove pointless do_QueryInterface() functions from nsBaseChannel.h r=mayhemer
The first QI is never used. The second one is the same as the one in
nsCOMPtr.h, so we should be able to call that instead, which can be
done simply by deleting the method. The motivation is that I'm
changing how do_QueryInterface works, and I'd like to avoid changing
this place given that it isn't actually needed.

Differential Revision: https://phabricator.services.mozilla.com/D7528

UltraBlame original commit: fc4e7aa8f23ec6048fea85e69703deebe2b7477c