Bug 1496128 - Fix perma fail in browser/components/newtab/test/browser for search shortcuts r=k88hudson
Each of the tests account for search shortcuts being there, so turning them off was leading to failures. Instead, we'll leave them on, since that will be the default case, and let the tests decide on the behaviour for themselves.

Differential Revision: https://phabricator.services.mozilla.com/D7731

UltraBlame original commit: a9d0b65595c1cfd92c7e78747244066e6bbf4d47