Bug 1433905 - [mozprocess] Existence of _handle on Windows doesn't mean the process is still alive. r=gbrown
The assumption that when a handle is present for the process handler
on Windows doesn't mean that the process is still alive. It could
have already been externally killed, crashed, or closed itself.

The patch makes sure to check the process exit code, and run
clean-up steps if the process is already gone.

Depends on D7395

Differential Revision: https://phabricator.services.mozilla.com/D7396

UltraBlame original commit: 8793e332890e9dab0c931d92e295965028ad4bef