Bug 1494162 - Part 52: Don't show the empty message in the CssRuleView constructor. r=pbro
We will eventually call showEmpty() when an element is selected if there are no rules.
This improves the perceive performance of the rules view because it doesn't have
a flicker of the empty text to the css rules.

UltraBlame original commit: 465b5d63240a5e7bc8c47f79132081fde520b6df