Bug 1496529 - P8. Add media.navigator.mediadatadecoder_h264_enabled preference. r=drno
For now, the H264 decoding support is rather sturdy. It handles change of resolution and content smoothly thanks to the H264Converter.
The VP8/VP9 one however, not so much.

So we make a preference to only enable H264 for now.

Depends on D7895

Differential Revision: https://phabricator.services.mozilla.com/D7908

UltraBlame original commit: 357d4762fa8fda3470020886958c300aae51cb22