Bug 1470280 - Part 1: Make browser_preferences_usage.js less brittle. r=kmag
browser_preferences_usage.js hardcodes a max value of 15 accesses which doesn't scale well as we increase the number of processes. Instead we'll base the max on a multiplier of the default process count which should avoid bustage in the future but still catch egregious accesses.

UltraBlame original commit: 0e32136949201c16777740ec69999caa8541657b