Bug 1484989 - Prevent JsTerm to steal focus when reloading a page; r=bgrins.
JsTerm's focus function was called in clearOutput, which
we call when navigating to a new page (if Persist Logs is
not checked).
This means that we were forcing the JsTerm to be focused
each time the user navigated while having the console open.
This behavior, can be annoying, or at worst, if you're
debugging a focus issue in your content page, completely maddening.
The fix is striaghtforward: do not call focus in clearOutput.
A test is added to make sure we don't regress this.

Differential Revision: https://phabricator.services.mozilla.com/D8701

UltraBlame original commit: 57c85fc5a5bc6a1c25fc0050aa4350bbe82f6ecf