Bug 1424708 - trigger a shortcut even when the number pad is used for the key r=mstriemer,mixedpuppy
Prior to this patch, if a user attempted to use a keyboard shortcut that used a numeric key, the action would not be triggered if they used the numeric key from the number pad. This patch triggers the shortcut regardless of the source.

Differential Revision: https://phabricator.services.mozilla.com/D8535

UltraBlame original commit: 900eabf8fceeec876fe998b1b91d5ddd09bb20f2