Bug 1499153 - Remove obsolete manifest check and NO_JS_MANIFEST r=glandium
Bug 1438688 made it so that XPT information is compiled directly into
the binary instead of being shipped separately in interface
files. This means that manifests are no longer necessary for JS
components, which means the manifest check in emitter.py can be
removed.

That check is the only use of NO_JS_MANIFEST, so that can in turn be
removed entirely.

Differential Revision: https://phabricator.services.mozilla.com/D8885

UltraBlame original commit: 55bb23a27ec7e39516d9d7f4254b312220f2ea34