Bug 1473841 - Transform property access on a dot-notation invalid name into an element access; r=bgrins.
This patch turns property access that would result in
Syntax error (e.g. `x.data-test`) into element access
(e.g. `x["data-test"]`) when accepting a completion
value in the console input.
In order to do that, we use Reflect to parse  a custom
expression where we try to define the property the user
is going to accept. If this throws, this means we need to
modify the input into an element access.
A test is added to make sure this works as expected.

Differential Revision: https://phabricator.services.mozilla.com/D8952

UltraBlame original commit: 0e8706e5d2ffc8a43f45ae6a7f6e6b286e44e3ca