Bug 1499727 - make nsHtml5StreamParser timers slightly more efficient; r=hsivonen
Instead of creating a timer and then setting the timer's target, we can
determine the timer's target and pass it in directly when the timer is
created.  This reordering of steps is slightly more efficient, since
SetTarget() is both a virtual call and requires locking, both of which
can be skipped if we know the target at timer creation time.

UltraBlame original commit: fd914881db604175bdd129d62ec1cb3c8d0a0386