Bug 1499849 - make FileReader timers slightly more efficient; r=baku
Instead of creating a timer and then setting the timer's target, we can
determine the timer's target and pass it in directly when the timer is
created.  This reordering of steps is slightly more efficient, since
SetTarget() is both a virtual call and requires locking, both of which
can be skipped if we know the target at timer creation time.

UltraBlame original commit: 8b037400fdbed6b9fd14c36a8b32acd78a742e1d