Bug 1499070 - Fix timestamp for evaluation result; r=Honza.
When sending a command to the server, a timestamp is
computed before evaluating the string, and is then
sent back to the client in the packet.
However, if top-level await, or somme :commands, the
evaluation takes more time, which means the timestamp
is now innacurate.
For those cases, we update the timestamp before sending
the packet to the client.

Differential Revision: https://phabricator.services.mozilla.com/D8734

UltraBlame original commit: c645265aa5b827fac52c3692991cbd55558d5821