Bug 1486980 - Fetch tippytop collection in one bulk with getAll() r=asuth,nanj
Iterating an IDB cursor generates a lot of overhead. In this patch, we upgrade the kinto-offline to its latest version that uses `getAll()` when no filter is specified. We leverage this change in tippytop, by omitting the filter and filtering the whole list there instead. This allowed us to go from ~1sec on a 1000 entries to ~70ms.

Differential Revision: https://phabricator.services.mozilla.com/D9076

UltraBlame original commit: 9a05c09af3770321c9d3acd89204ac62f3fb1878