Bug 1486888 - Support XUL broadcasters in non-XUL documents. r=smaug
The majority of the XUL broadcaster logic is moved out of XULDocument and
into a separate class (XULBroadcastManager). The hookup points for when
listeners need to be created and listeners need to be notified is now
handled by the XULElement itself and nsDocument. To avoid any overhead,
the XULBroadcastManager is only ever created when a document uses a
listener.

The new approach does have the disadvantage that broadcasting can now only
work with XULElements, but going forward we'd like to discontinue this
feature and rely on MutationObservers to implement similar things.

One test had to be modified to use XUL elements instead of HTML elements
because of the reason noted above.

Differential Revision: https://phabricator.services.mozilla.com/D8888

UltraBlame original commit: 394a1e5544f65e59ea1e12e19f788f50078ddf20