Bug 1465619 - Part 10. Re-add support for discarding animated image frames. r=tnikkel
AnimatedFrameDiscardingQueue subclasses AnimationFrameBuffer to allow a
cleaner abstraction over the behaviour change when we cross the
threshold of too high a memory footprint for an animated image. The next
patch will build on top of this to provide an abstraction to reuse the
discarded frames.

Differential Revision: https://phabricator.services.mozilla.com/D7515

UltraBlame original commit: 453d21db2a306f4800e8be4d77f6d61000cb9a0c