bug 1499766 - rework the threading model of OSKeyStore r=jcj
The original threading model of OSKeyStore could lead to a deadlock if an
asynchronous event were dispatched and then the isNSSKeyStore attribute were
queried. This patch removes that pitfall by moving the determination of the
attribute to OSKeyStore rather than the underlying implementation.

Additionally, the original threading model was inefficient in that it created
and destroyed a thread per asynchronous operation. This patch reworks this to
only ever create one worker thread.

Differential Revision: https://phabricator.services.mozilla.com/D9299

UltraBlame original commit: f3273aa382dca7f8149d9839603635d183dc2cf5