Bug 1499554 - Don't blacklist the GPU process for vendors not on the whitelist. r=jrmuizel
We currently allow the GPU process if we are not hardware accelerated. One of the
reasons we might not use hardware acceleration is because the device vendor is
not in the gfxInfo whitelist. In this case, we should be consistent and still
use the GPU process.

Differential Revision: https://phabricator.services.mozilla.com/D9446

UltraBlame original commit: b7b09fca2cc5c9b06ecf3a4fb4494b9a867d4541