Bug 1438979 - Fix browser_webconsole_network_messages_expand.js intermittent; r=Honza.
This is an attempt to fix the intermittent on this test.
It looks like we were setting the listeners on some events
after the request message was received, which might have
made us missed some events.
We take this bug as an opportunity to do some cleanup
on the test.

Differential Revision: https://phabricator.services.mozilla.com/D9075

UltraBlame original commit: f266aa4cd4b04f7cec1fffeec4fccd1baafdf244