Bug 1419091: Make TextInputListener handle non-native events for input and textarea too. r=masayuki
platformHTMLBindings attaches event handlers directly to the <input> and
<textarea> elements. This matches that by making TextInputListener look up and
call the <input> and <textarea> event handlers from the static C array.
There is a slight difference in that the event handlers are added to the system
bubbling phase as opposed to the regular bubbling phase but in tests this does
not seem to cause problems.

Differential Revision: https://phabricator.services.mozilla.com/D8931

UltraBlame original commit: 40371ad8146476cc3eb65fd1389e1a4fa9cbc91e