Bug 1480550 - add ctypes support for aarch64 windows; r=dmajor
The bulk of this patch is the new win64.asm, which is a more-or-less
direct copy of aarch64's sysv.S file, with modifications for armasm64's
peculiarities.  The changes to ffi.c were minimal, mostly so that
arithmetic on `void*` conforms to the C standard.

UltraBlame original commit: 94c9a52a4f1da95c6a260cffe391cb024ede58d2