bug 1492499: marionette: warn on TimedPromise bailing; r=automatedtester
We often use TimedPromise to ensure Marionette does not unexpectedly
block on a promise that, for whatever reason, does not resolve.
It can however be useful to be alerted when they don't, as it quite
often means there is an underlying problem.

Depends on D8406

Differential Revision: https://phabricator.services.mozilla.com/D8407

UltraBlame original commit: 9ca2c179e63e292808918158eb778b16bf18e67f