Bug 1502087 - Don't reuse session for AccessibilityTest.testScroll. r=snorp
It looks like a previous test has an input focus and this test starts
when the scheduled animated ZoomToRect starts (via GeckoView:ZoomToInput). Then it interferes with the expected scroll events in this test.

There might be a deeper bug here (the unload from a previous page should not affect the current APZC). There might also be a way to do better cleanup on the text edit tests that mess with this one. But it's tricky because we don't get a callback when ZoomToInput is done.. (TextInputDelegate.restartInput won't do).

UltraBlame original commit: 8a49abaefd77f365465d0e1f73631b4e2f5964fc