Bug 1503877 - Handle PauseComposition correctly with webrender enabled. r=kats
On android with webrender enabled, minimising the app results in an
assertion failure: CompositorBridgeParent::DidComposite asserts that
it should not be called with a webrender compositor.

This makes it so that that in CompositorBridgeParent::PauseComposition
we call NotifyPipelineRendered instead of DidComposite for webrender
compositors.

Differential Revision: https://phabricator.services.mozilla.com/D10563

UltraBlame original commit: b23c55fef2bee35ffa2ea5e6006bd57251273247