Bug 1428558 - Part 3. Release our reference to an external image ID as soon as possible. r=nical
Animated images will work by changing the external image ID that an
ImageKey points to. We cannot allow the old external image to be
released and potentially unmapped until we have produced a new frame
with the new external image ID. We currently wait until the epoch has
advanced, but in the future when we don't rebuild the scene to animate
an image, the epoch will remain the same. This could cause us to hold
onto no longer used surfaces for much longer than expected. As such, in
this patch we switch to waiting for a FrameRendered notification from
WebRender, which works even if the scene rebuild was avoided.

Differential Revision: https://phabricator.services.mozilla.com/D7500

UltraBlame original commit: 92fde8cff06837f53d5c85a447b6eb66d5ffea95