bug 1504174: marionette: delete unused test_window_fullscreen.py test; r=automatedtester
The test_window_fullscreen.py test is not run because it is not
part of the test manifest.  Since it is duplicated by WPT test,
we will not try to resurrect it by fixing the test.

Differential Revision: https://phabricator.services.mozilla.com/D10710

UltraBlame original commit: a39f7a592250fd777f574aca3ac5910abd3e2f1e