Bug 1503006 - Part 3: Trivial slot accessor methods for class ReadableStreamReader. r=tcampbell
This commit changes ReadableStreamReaderGenericRelease to use
UnwrapInternalSlot when accessing reader.[[closedPromise]]. Before this commit,
the code assumed that there would be a promise in the slot, not a wrapper.
I think that was actually correct, because it's impossible for the stream to
still be readable and the reader to have a [[closedPromise]] that's a wrapper.
Still, it's a bit precarious (and I'm still not 100% sure about that), so I just
changed it to unwrap.

Depends on D10374

Differential Revision: https://phabricator.services.mozilla.com/D10428

UltraBlame original commit: 992f76656e1187fb7a33caa7682000d76da717d6