Bug 1446307 - Compute toggled call size properly; fix constant pool header. r=sstangl
ToggledCallSize() on ARM64 did not properly account for all the
possible places where a constant pool could be found; also, it can be
updated to use better abstractions.

WritePoolHeader() on ARM64 did not generate a correct header because
it used sizeof(Instructions), which is meaningless on this platform,
as Instruction has no data fields.  Use kInstructionSize instead,
that's what it's for, and everyone else uses it.

UltraBlame original commit: 24f10efb4ce15935103f61a360969af89503a41d