Bug 1503082 - Clear CanvasShutdownObserver::mCanvas when the canvas goes away. r=nical
It is possible for the CanvasRenderingContext2D to be destroyed while
we're in the middle of the call to nsObserverService::NotifyObservers()
for shutdown. This leaves the shutdown observer with a dangling pointer
to the canvas, so this patch explicitly clears the pointer when the
context goes away.

Differential Revision: https://phabricator.services.mozilla.com/D10777

UltraBlame original commit: d3284735c7fa4bbe5cddc24f3905112f3490fa6d