Bug 1501226 - Support for CSS transforms in the Flexbox highlighter r=gl
In many places we were using bounds to calculate an objects dimensions but that obviously doesn't work when an object is rotated e.g.

{F588303}

Also, we were using `getBoxQuads()`, which gives the co-ordinates of the translated object. We were then applying the transform matrix to the canvas even though the coordinates came from the object **after** it was already transformed.

Anyhow, now we get the dimensions of objects as if they are not transformed and then apply the transformation matrix, which gives a great result every time.

Differential Revision: https://phabricator.services.mozilla.com/D9805

UltraBlame original commit: fe2f73a6db0e245944ff3db9b5f001b3c2db110e