Bug 1497917 - Extract client creation to RuntimeHelper module;r=ladybenko,daisuke
Depends on D10095

The initial intent was to extract the logic to create the USB clients
so that it could be mocked in our tests. However since USB clients
share most of their logic with network clients, it did not make sense
to extract USB without network. And once you have USB and network, there
is no good reason to keep local client out.

I don't think "runtime-helper.js" is a good module name here. The role
of this module is to build clients for runtimes, but at the same time
it is just a static helper. Could not find a good name, and we already
have inconsistent module names in our modules/ folder so I did not want
to introduce something like RuntimeClientBuilder of RuntimeClientProvider
etc...

One of the consequences is that once we start mocking this helper (in the
next patch) we will also have to mock the this-firefox runtime. It has
advantages and disadvantages but I think we can live with this for now.

Differential Revision: https://phabricator.services.mozilla.com/D10096

UltraBlame original commit: dfd527956c4b0df389286e62b51f28e87b64676d