Bug 1503776 - Save default value of WebRender rollout pref to user pref, check on startup. r=mattwoodrow
Normandy's Preference Rollout code sets default values on prefs, not user
values (see uses of PrefUtils.setPref() in PreferenceRolloutAction.jsm).
Default prefs are not persistent; unlike user prefs, changes to default pref
values are not stored on disk. Changes to default values are only made on the
in-memory copy of the pref's value, and thus don't survive a browser restart.
Normandy changes the rolled out prefs early on in the startup of the browser,
but *after* gfxPlatform::Init() runs. So that means gfx can't use Normandy
pref rollout to gradually rollout WebRender to release, as
gfxPlatform::InitWebRenderConfig() won't see the rolled out version of the
pref in time to turn on WebRender.

So to work around this, add a profile-before-change shutdown observer that
saves the default value of the gfx.webrender.all.qualified pref to a new user
pref, gfx.webrender.all.qualified.default. We check that on startup and
emulate the behavior that the pref system would have if that pref's default
value had already been set by Normandy.

Also add two prefs; the first overrides the default value for the rollout pref
stored in gfxPrefs.h. WebRender is default on in Nightly on Windows at the
moment, but the test needs to test the case where it's default off and we're
rolling out turning it on.

The second pref is to override whether the machine we're running on is
considered "qualifying hardware" or not. This ensures we can test the
"non-qualifying hardware" and "qualifying hardware" cases, independently
of whatever hardware the test machine is running.

Differential Revision: https://phabricator.services.mozilla.com/D10527

UltraBlame original commit: 838d9917f028af1982b0bf18133d824873bb5a38