Bug 1504418 - Access to nsIDateTimeInputArea interface implemented by XBL from chrome via wrappedJSObject r=mconley
Part I of bug 1496242 was not implemented correctly as it accesses the
nsIDateTimeInputArea methods directly on the <datetimebox> element object.
This won't work because the XBL binding is run in a non-chrome scope.
The methods it implemented are exposed thus had to be accessed via
wrappedJSObject.

Object pass to setValueFromPicker() therefore has to clone into the content scope.

We can be sure that the methods called are implemented by the XBL binding,
because the web content should not have access to <datetimebox>, which is a NAC.

A small clean-up is included in the patch also, removing the useless 3rd argument
to the CustomEvent constructor.

Tests related to the datetime input is duplicated and run with UA Widget disabled
to ensure the XBL binding continue to work.

Differential Revision: https://phabricator.services.mozilla.com/D10947

UltraBlame original commit: 97fa9c6849ff49a08ce0a55b568e43e9ac3cc7c7