Bug 1506920: [taskgraph] Track parent images in docker image digests; r=dustin
The digest for a docker image task did not include the digest for the parent
image in it, and so in particular did not depend on the versions of packages
included in a parent image.

If two branches have a docker image with identical docker files, but different
parents, this would lead to them both getting the same digest, leading to
unexpected interference between the branches.

This fixes things by including the digest of the parent image as input into the
digest of child images.

Differential Revision: https://phabricator.services.mozilla.com/D11807

UltraBlame original commit: 7aa761d6638c31c7a1db5bdbb846c802d50f0a7d