Bug 1506660 - Dispense with TODO comments about bytesWritten in ReadableStream implementation. r=till
The first comment appears in ReadableByteStreamControllerPullSteps. It's OK to
delete it because we just asserted that there are no pending read requests
(step 3.a.).

The second comment appears in JS::ReadableStreamUpdateDataAvailableFromSource.
It's OK to delete this one because it's in the hasDefaultReader branch. Default
readers don't have read requests that care about the number of bytes available.

Differential Revision: https://phabricator.services.mozilla.com/D12111

UltraBlame original commit: 8c5eaa1d4356ff8a4cd295f43d89f82d375bcc1b