Bug 1507730: Generate a temporary for negative power-of-two constants in mul64 on 32 bits platforms; r=lth
In the code generator function, we assume we have a temp if a known constant is
a non-negative power of two. But lowering only checked for power of twos
independently of their sign, so this was incorrect. This patch syncs them up on
both ARM32 and x86.

UltraBlame original commit: 98b78581ff763823e5eb20b0b559181e12bea5ea