Bug 1501878 - drop complete mar from repackage-signing. r=tomprince
We want to defer mar-signing to the second half of the promote phase,
but we want to keep Windows repackaged exe signing in the first half of
the promote phase. This means we need to stop signing the complete MAR
in repackage-signing, and defer it to another kind.

Differential Revision: https://phabricator.services.mozilla.com/D11730

UltraBlame original commit: 050c4bba5d710059175026e9f0185e795fcdfc28