Bug 1507615 - part1 : don't activate document if the 'pointdown' is originated from touch event r=smaug
For touch events, we use 'touchend' to activate document, because 'touchstart' can be the start of either draging
or touch. As 'pointdown' is a precedent event of the 'touchstart', we should not activate document if the source of
event is from touch event.

Differential Revision: https://phabricator.services.mozilla.com/D12333

UltraBlame original commit: f0879820fc44ee68410702f98024c2c841da382d