Bug 1508393 - Enable animated images producing full frames by default. r=tnikkel
Also, fix a minor bug where when we discard an animated image that is a
full frame animated image, we need to reset
AnimationState::mCompositedFrameInvalid to false, just like we do for
animated images blended by FrameAnimator. This is because it is used as
part of our state checking in FrameAnimator::GetCompositedFrame before
we are willing to yield frame data.

Differential Revision: https://phabricator.services.mozilla.com/D12362

UltraBlame original commit: 3c7a3fc19c85ebca0be908143148c388f9d36f04