Bug 1505878 [wpt PR 13984] - Add report-only mode to Feature Policy, a=testonly
Automatic update from web-platform-testsAdd report-only mode to Feature Policy

This change adds a "report-only" policy to each security context, which
is local to that context (not replicated across processes) and is used
to determine whether a report should be sent even if a feature is
enabled, when the feature is used.

Reports are now augmented with a "disposition" field, which is either
"enforce", if the feature usage was actually blocked, or "report", if
it was not.

Feature policy directives are placed in the report-only policy if the
feature name is suffixed with "-report-only", otherwise, they affect
the regular (enforcing) policy.

Explainer at
https://github.com/WICG/feature-policy/blob/master/reporting.md

Existing tests are updated, and new tests for report-only mode are
added to ensure that reports are sent even when the feature is used
successfully.

Bug: 904878
Change-Id: I27bc42729c5ab5560160f3d993431e606a8a3a47
Reviewed-on: https://chromium-review.googlesource.com/c/1178811
Commit-Queue: Ian Clelland <iclellandchromium.org>
Reviewed-by: Ken Buchanan <kenrbchromium.org>
Reviewed-by: Dmitry Gozman <dgozmanchromium.org>
Cr-Commit-Position: refs/heads/master{#608004}

--

wpt-commits: b39f5e0d0e0872cdd9b7592f0b9ff0b5c11bcef3
wpt-pr: 13984

UltraBlame original commit: 0475d6073c309df4fb00aa1f2adad320d6f0a276