bug 1507146 - make changesView have an async init method; r=ochameau
ChangesView was challenging because the async call was in the constructor. I moved it out of the
constructor to the init method, and do a check to see if its been initialized in the destroy method.

Depends on D11886

Differential Revision: https://phabricator.services.mozilla.com/D11887

UltraBlame original commit: 5990eabd10965f59d4a33c9fd2453a5cc96794dd