Bug 1352355 - Enable storing and updating mozleak metadata in wpt ini, r=ato
This adds two new properties to wpt metadata files:

mozleak-allowed - This is a list of the form [process-name:object
name], which indicates objects that may be leaked in that specific
process. Automatic updates that find a leak may add that object to
this list.

mozleak-threshold - This is a list (but conceptually a map) of
[process-name: threshold bytes], indicating a threshold below which
leaks will not cause a test failure. This number is updated by setting
it to the observed value for a process.

MozReview-Commit-ID: KA1oPl837a8

Depends on D12410

Differential Revision: https://phabricator.services.mozilla.com/D12411

UltraBlame original commit: 02e4dbfecbc426e118199e4ce140db13814aaf7a