Bug 1352355 - Fix use of wpt settings in firefox Browser, r=ato
The settngs() method is called to get a list of session-level settings
that should be applied when running a test. If those settings differ
from the ones applied to the previous test then we have to restart the
browser. But if we set the session on the class in the settings method
then at shutdown time we'll have incorrect settings. This is important
for things like leak checking where the values affect the shutdown
process. So instead ensure that we only set the settings on the class
during startup.

Depends on D12414

Differential Revision: https://phabricator.services.mozilla.com/D13512

UltraBlame original commit: 36c4c80ddc1a12665af8cbcaef716e384caab99e