Bug 1514104 - Remove vpx_mem_set_functions support. r=jya
vpx_mem_set_functions support is only enabled when
MOZ_VPX_NO_MEM_REPORTING is not set. It is currently set unconditionally
when building with the in-tree libvpx. When building with system libvpx,
it is set when the vpx_mem_set_functions can't be found in the system
libvpx library.

Upstream removed the vpx_mem_set_functions function in version 1.5, and
we require at least that version, meaning, in practice,
MOZ_VPX_NO_MEM_REPORTING is now always set.

We might as well remove the define and the code that's conditional to
not being defined.

Differential Revision: https://phabricator.services.mozilla.com/D14517

UltraBlame original commit: 8160b9f376eefdfba360fb8d1a851c6dcf4bcd93