Bug 1510934 - Ensure that pref changes during content process launch are delivered to the process. r=njn
Some prefs need to be available before IPC is started, so we serialize a
snapshot when we start launching the process, and then stream further
changes over IPC messages.  However, async launch introduces a window
between the snapshot and when the parent can start sending messages,
during which other code can run on the main thread and change prefs.

In order to not lose those updates, they're queued and sent when the
launch is complete.

Depends on D14089

Differential Revision: https://phabricator.services.mozilla.com/D14090

UltraBlame original commit: 74bfa63e922c5782da7f5a58979bbddec0e4fa45