Bug 1512029 part 3 - Use freshZone for debuggee sandbox created in devtools xpconnect tests. r=bzbarsky
The debuggee and debugger must be in separate compartments. I tried using a null
principal for the debuggee, but that doesn't work because some of these tests
use Components or do interesting principal-related things with the sandbox (see
test_objectgrips-17.js for an example of this).

Differential Revision: https://phabricator.services.mozilla.com/D14254

UltraBlame original commit: 871140f005370426034bc932d4f4e30820776266