Bug 1514844 - Fix notification events in service workers. r=asuth
The current code assumes that it can get the ServiceWorkerManager in the child
process to send a message to the proper service worker. That isn't true, we
need to ask the parent to do it for us.

Differential Revision: https://phabricator.services.mozilla.com/D14773

UltraBlame original commit: c30fafd98c6f13b0c925f8f1bcfe9f95fb092b18