Bug 1514251 part 1.  Stop relying on compartment-per-global in ReparentWrapper. r=peterv,jandem
This code dates back to when we had a concept of parent as distinct from the
concept of global.  It was comparing compartments back then because in the
same-compartment case it would just JS_SetParent and return.  When we got rid of
the concept of parents, the code was left as-is, even though at that point we
could just as easily compare the two globals.

I believe that in the same-compartment-different-globals case this is safe,
because in that case JS_TransplantObject will just keep using the original
object allocation but JSObject::swap it with the new object, so that it will
pick up the new global.

UltraBlame original commit: 8e52c32b787410c88a0e45c5e86ad217173dc891