Bug 1516025 part 3: Move SVG static helper function "ParseNumber" outside of the mozilla namespace to fix unified build bustage. r=longsonr
This function is invoked elsewhere in this file as "::ParseNumber(...)", which
only works if it's declared outside of any namespace.  So, it needs to be moved
outside of the mozilla namespace in order to be referenced that way (and to be
truly a static function that's private to this .cpp file).

I'm not sure why this builds at all in unified mode, but however it works, it's
cheating, apparently.

Also: this patch adds a 'using' statement for the mozilla namespace, for
e.g. the RangedPtr usages in this function definition (since it no longer lives
inside of a namespace mozilla{} block).

Depends on D15232

Differential Revision: https://phabricator.services.mozilla.com/D15233

UltraBlame original commit: 8fd83a79a594383e7a3bdf9c7c74a9262e97ee9a