Bug 1519184: Fix cache purging on windows. r=Gijs
When Firefox runs we cache a number of things, JS, XUL, XBL, CSS etc. Whenever
a new build of Firefox runs against a profile we clear the cache to rebuild with
the new information.

On the first run of a profile where compatibility.ini doesn't exist (presumably
to cover the case of upgrading from a very old versions of Firefox, but also
affects new profiles) we attempt to clear the cache as well.

If any attempt to clear the cache fails we set a flag in compatibility.ini
telling us to clear the cache on next startup.

Unfortunately nsIFile.remove returns different error codes on different
platforms when the file in question already doesn't exist. So an attempt to
remove a cache that does't exist is counted as a failure on Windows and so the
cache will be cleared again on second run.

This change counts the file not found return code as counting as a success when
clearing the cache.

Differential Revision: https://phabricator.services.mozilla.com/D16224

UltraBlame original commit: f747aad08cbcc77d2b93dbe2b30a04b2096e79b4