Bug 1520139: Make sure nsXREDirProvider can find commonupdatedir.h regardless of whether the updater is enabled. r=froydnj
When the updater is disabled toolkit/mozapps/update/common is not traversed as
part of the build. But toolkit/mozapps/update/common/commonupdatedir.cpp is
included in the toolkit/xre build regardless and GetInstallHash is used. This
makes nsXREDirProvider.cpp able to find the header that defines it.

Differential Revision: https://phabricator.services.mozilla.com/D16582

UltraBlame original commit: 81873890e8b7d43fa291b54fe506f9b1e4fb4ee2