Bug 1510139 - Part 1. Ensure we honor invalidations if using vector image containers. r=tnikkel
If a vector image has an image container, it is unlikely the caller will
call VectorImage::Draw (and thus Show indirectly) to display the image.
As such, WebRender was missing subsequent invalidations and not
regenerating the rasterized surface as expected. Thus we now resume
honoring the invalidations if we updated the image container.

Differential Revision: https://phabricator.services.mozilla.com/D15667

UltraBlame original commit: 325d06d3ba7341035f01ae14bc882643d0235083