Fix Bug 1419109 - Empty default icon shows blank icon in toolbar r=mixedpuppy
If the extension had either default_icon or one of it's property as an empty string, it would show a black icon in the toolbar. With this patch, it checks if any of default_icon property is empty and throws an error on extension load.

Differential Revision: https://phabricator.services.mozilla.com/D16070

UltraBlame original commit: d7bf0d21448d4c1564f3d5e512167d2b8f887264