Bug 1512436 Add new built-in addon location r=kmag
The big change here is breaking the previous invariant that every
active instance of XPIState contains a file and path.  With the change
in place, the implementation of the new "location" for built-in addons
is pretty straightforward.

Differential Revision: https://phabricator.services.mozilla.com/D15639

UltraBlame original commit: e476cb7c0c0b9baaabfa0b75851ee2a922ec3678