Bug 1494924. Redo the calculation of paintRect. r=mattwoodrow
This changes the calculation to match what we do for SVG blobs.
Notably, it takes the residual into account when computing the area
we want to paint into. This makes a difference if the fallback
is happening at a fractional offset.

Differential Revision: https://phabricator.services.mozilla.com/D17178

UltraBlame original commit: 4b841cc68f4d4ad817412290a95290915649773b