Bug 1519862 - Interpret the window being missing as a crash, r=ato
If we get a NoSuchWindow error from marionette, then something happend
that caused the window to disappear, which we can interpret as a
crash. This may not be true if we accidentially navigated the window
to file:// or something else that causes a remoteness change, but
those are basically marionette bugs so in the long term this should be
correct.

Differential Revision: https://phabricator.services.mozilla.com/D16830

UltraBlame original commit: a1544d1ae5be112f87dde5c4702549299c6caf5d