Bug 1521052 - Destroy DebuggerServer in the content process when the last connection drops. r=jdescottes
We never really tried to cleanup the DebuggerServer and so a few tests require some tweaks
to acknowledge that once the last connection drop (typically, we close the toolbox or target),
the server is destroyed and dynamically registered actors are also destroyed.

I think it is great to consider that everything is cleaned up as we may followup to destroy
the whole loader.

Depends on D16961

Differential Revision: https://phabricator.services.mozilla.com/D16962

UltraBlame original commit: c68ab257f8c8f0023ad3dbb436de6fb49bd54317