Bug 1522140 - Make .ycm_extra_conf.py handle headers better. r=Ehsan
Files with .h extension are built with the c compiler by default, which makes
the diagnostics useless. If we know this is a cpp file, hint it so that it
builds it with the right mode.

Differential Revision: https://phabricator.services.mozilla.com/D17372

UltraBlame original commit: 2a22490b8a06c9a5eef192ef3192386f8e7ff74f