Bug 1522051 - Stop giving singleton types to call objects. r=tcampbell
This optimization was added in bug 864218 to optimize asm.js-like code in Ion.

We're removing it now because:

* WebAssembly happened and we've been focusing more on page load than on hot
  loop performance.

* It complicates the interpreter work because JSOP_SETALIASEDVAR can't do the
  fast thing on singleton call objects and needs to fall back to an IC.

* JSOP_SETALIASEDVAR and JSOP_INITALIASEDLEXICAL are currently marked as JOF_IC
  so they always get a SetProp IC allocated for them. With this patch that no
  longer happens.

* The patch removes a lot of complexity.

* It doesn't seem to affect Speedometer, Octane, other benchmarks.

Note that this also removes the EnvironmentCoordinateNameCache because
EnvironmentCoordinateName is now only called in debug-only or error-handling
code. I renamed it to EnvironmentCoordinateNameSlow to emphasize this.

Differential Revision: https://phabricator.services.mozilla.com/D17339

UltraBlame original commit: 7259ee92e345359164baba67b035cc5d98b76ff4