Bug 1521510 - Keep a reference to the document in PaymentRequest. r=qdot
The current document of a window can change in between the time when a
PaymentRequest registers as an activity observer and when it attempts
to unregister, so keep a strong reference to the document. This is the
same issue as bug 1317805.

Differential Revision: https://phabricator.services.mozilla.com/D17540

UltraBlame original commit: 385938bebc4ed74106e3b3c9562aa6c493abdd90