Bug 1523245 - Fix stacktrace with frames group copy; r=bgrins.
This is a follow-up of Bug 1486870.
In this patch we make sure that copying a
stacktrace containing a group, i.e. frames
that were grouped together because they're
part of the same library/framework, are copied
as expected to the clipboard.
There was new lines added between each frames
of the group, because they're in divs.
To fix that, we remove the display: block that
was set for group frames (so they're display: contents,
like the regular frames), and place them in a
grid with only one column (to take the whole horizontal
space).
A br is added in the group description to add a new line
after it.
A test is added to ensure this works as expected.

Differential Revision: https://phabricator.services.mozilla.com/D17817

UltraBlame original commit: fa32d5cfc2232b12eec0c786cd76a34a563f9280