Bug 1523684: Set `--target` to be a 32 bit target for L10n builds on win32; r=nalexander,Callek a=tomprince
In Bug 1522354, the default for `--host` was changed to detect running on a 64bit host.
Since `--target` defaults to `--host`, this changed the default for `--target`
as well. This didn't affect builds, as they explictly set the `--target` in
`build/win32/mozconfig.vs2017`. Set it explicitly in L10n builds as well.

Differential Revision: https://phabricator.services.mozilla.com/D17967

UltraBlame original commit: 58b77413fd917d29a8975f87903309056e8affe4