Bug 1515774 - Store window's screen offset in APZ and use it for gesture interpretation. r=botond
In Android the embedded GeckoView can be in a scrolling parent, and move
along with the user's finger. In order to intepret touch movements for
gestures, we need to account for the device position of each touch.

Differential Revision: https://phabricator.services.mozilla.com/D17045

UltraBlame original commit: 45c6f07160d9c0b7a472c68ff63d4aaa9bd00304