Bug 1455707: Detect when running an older version than previously ran with the selected profile. r=froydnj, r=mconley, r=Gijs
Use the information in compatibility.ini to detect that the current running
application is an older version than previously ran with the profile and in
that case open a UI allowing the user to launch the profile manager, launch
the previous instance of the application or quit.

Also includes the patch from bug 1523725.

UltraBlame original commit: 21a801ca5f6d435509f93e1dee187cb6ca868c8f