Bug 1258897 - Ensure that both test_ext_sendmessage_reply2 tab scripts are ready to exchange messages. r=mixedpuppy
This test seems to be still failing intermittently, and its intermittency rate looks high enough.

Looking to the logs collected on the recent intermittent failures, it seems that the issue
is still related to the extension tabs opened in the test:

The two extension tabs (each one part of one of the two test extension) are going to exchange messages
between each other, and currently there is a chance (and apparently a very good chance) that one of the
two extension tabs is not yet ready to listen to the messages sent from the other extension tab.

This patch applies to the test the additional changes needed to ensure that both the extension tabs
are ready to exchange messages, before we let them exchange these messages over the extension messaging
API.

Differential Revision: https://phabricator.services.mozilla.com/D18230

UltraBlame original commit: c127e67367b9cfda6e170521a260c708b41c5a89