Bug 1511381 - Prevent forced layout flush when click-to-play UI loads on page load r=mconley
This patch attempts to bail out of probing into the layout when it is not ready.

It abuses the overflow event from layout a bit so that it could size and do the elementFromPoint() tests when there is a layout.

Differential Revision: https://phabricator.services.mozilla.com/D15905

UltraBlame original commit: 3e8eaa5b79372adec8622e9a486b64ba2f71208e