Bug 1523341 - Allow to only pass the CPU to --target and get the right thing for Fennec builds. r=nalexander
Also use armv7a as the default when no target is given at all.

Also change bootstrap to create a simpler mozconfig. The downside is
that the resulting mozconfig would not work when building older
revisions.

Add unit tests for this as well as the simplications added in bug
1523341.

Differential Revision: https://phabricator.services.mozilla.com/D17906

UltraBlame original commit: f59b1c1142590117a23f972183fbf34e558edfee