Bug 1522809 - Nodes with `name="attributes"` break the markup view r=rcaliman
This is another very specific issue.

If you have `<img name="attributes"/>` in the dom then `document.attributes` will return the `<img>` tag.

In the source we bail if `!this.rawNode.attributes` but if we are on the document node this returns the image tag. Because the image tag is not a `NamedNodeMap` trying to iterate over the tag throws the error.

There is a test file [here](https://bugzilla.mozilla.org/attachment.cgi?id=9040577).

Differential Revision: https://phabricator.services.mozilla.com/D18340

UltraBlame original commit: 85f55cb935f6b92943497373edf3b26c3442116a