Bug 1525327 - Don't crash wptrunner if mozversion fails, r=ato
On Android we aren't necessarily passing in the package path, so
mozversion can't extract the relevant metadata, and on other platforms
it could also fail if we pass in a script file or similar. In that
case we shouldn't fail.

Differential Revision: https://phabricator.services.mozilla.com/D18712

UltraBlame original commit: 9592b19c9b097e1a0136324cee3e8d54160a188e