Bug 1524702 - Make URL results use their domain names as their titles when they don't otherwise have titles. r=mak
There are a few potential places to do this:

(1) UrlbarProviderUnifiedComplete, when it creates results

(2) UrlbarResult.payloadAndSimpleHighlights(), which UrlbarProviderUnifiedComplete calls when it creates results (and presumably other providers would call it, too)

(3) UrlbarResult._titleAndHighlights(), when it figures out what to return as the title

It's hard to do it in (3) because in that location, we don't know the user's search string, so we don't know what to highlight. (1) is possible, but it would probably be nice for other providers and all result types to benefit from this? So I chose (2), and also because UrlbarResult.payloadAndSimpleHighlights() is just an easy place to do it since UrlbarProviderUnifiedComplete calls it for each result it creates.

The actual code to get the domain is copied from autocomplete.xml more or less.

Differential Revision: https://phabricator.services.mozilla.com/D18642

UltraBlame original commit: 9e5d27526a36a54f547622f348fd5e66ad00b3a5