Bug 1524325 Remove AddonEvent.needsRestart r=bzbarsky,kmag
We haven't supported addons that require restarts in some time
and a bunch of the underlying support for this flag is already gone
from the addons manager.  The mozAddonManager api is one of the
last vestiges, this patch rips it out.

Differential Revision: https://phabricator.services.mozilla.com/D18745

UltraBlame original commit: c7a0af90beb9d0a0bab586524501673d54d427ff