Bug 1525813 - Ensure ipc::ShmemCreated message serializes shmem size as uint32_t. r=jld
Shmem sizes serialized in an ipc::ShmemCreated message should be sent as an
uint32_t rather than a size_t, as size_t is defined as different sizes in 64
and 32 bit builds. If the size isn't consistent, we won't be able to reliably
send this message between cross architecture processes.

Also, Shmem's have a limit of 32bit for their size anyway:
https://searchfox.org/mozilla-central/rev/e00ea598e52bbb35f8c45abf9c2eade17962bb5e/ipc/glue/Shmem.cpp#127

So this patch should not affect allocations of shmems.

Differential Revision: https://phabricator.services.mozilla.com/D18909

UltraBlame original commit: 326199603b5dfa09155f0dbfd203dc477bcac768