Bug 1500257 part 2 - Add PRemoteFrame stub implementation. r=qdot
This commit adds a PRemoteFrame actor which is managed by PBrowser. It will
be created in a child process nsFrameLoader when loading a remote subframe.

This actor will mostly just bounce messages from a nsFrameLoader in the child
process to the actor in the parent process which will redirect the messages
to the TabParent of the remote subframe.

The piece in the parent actor to create the proxied PBrowser actors is
deferred to the next commit.

Differential Revision: https://phabricator.services.mozilla.com/D17442

UltraBlame original commit: e72d9d31a8bc15e0d3e17d3bdae0b5717465c4b9