Bug 1520478 - Wasm compilers: split MIRType::Pointer uses into MIRType::{Pointer or RefOrNull}.  r=lhansen.
Currently the baseline and via-Ion wasm compilers use MIRType::Pointer to
refer to objects allocated both on the C++ heap and the JS heap, and in fact
to pretty much any flavour of "address of something".

With the introduction of GC support for wasm we need to be much clearer about
which pointers denote a GCable object and which don't.  This patch introduces
a new type, MIRType::RefOrNull, to denote GCable pointers in wasm-land, and
converts most but not all existing MIRType::Pointer uses to MIRType::RefOrNull.

The patch was created by auditing all places in WasmBaselineCompile.cpp,
WasmIonCompile.cpp and WasmStubs.cpp, that contain the string
"MIRType::Pointer".  Once that was done, various other |switch| statements and
|if| sequences that failed during testing, were fixed up, either by

(1) adding MIRType::RefOrNull as a new case handled the same as the existing
    MIRType::Pointer case, or

(2) simply changing the MIRType::Pointer case into the MIRType::RefOrNull case
    when it seems that the ::Pointer case is no longer necessary.

UltraBlame original commit: ba0fe524ced9ca9f05dbb3d9b95e38440bc7261a