Bug 1525876: Export FILES_PER_UNIFIED_FILES in js/src/moz.build to avoid redefining it; r=froydnj
Before this patch, changing FILES_PER_UNIFIED_FILES in the directory would
require changes to all the files that defined the sandbox variable, which is
a bit misleading. Exporting the variable prevents this, and it is safe to use
because it doesn't escape the scope of the js/src build directory.

Differential Revision: https://phabricator.services.mozilla.com/D19647

UltraBlame original commit: 1bb3f6cdfa73a5e04688a8b3a79d3a620f72e0e9