Bug 1522253: Add a standard mechanism for requesting data from the other process. r=Gijs
As we add more specific methods to RemotePageManager some of them will have to
call the main process to get the results. This adds a standard mechanism for
doing so. Calling sendRequest will return a promise that is resolved or rejected
when the main process responds.

Differential Revision: https://phabricator.services.mozilla.com/D19412

UltraBlame original commit: a091f6b58aba85ea59443d5ae2f3756caca13ae9