Bug 1491445 - Wrap vibrancy and nondraggable views inside container views. r=spohl
NSView hierarchy before:

 - window contentView
   - ChildView
     - NonDraggableView 1
     - NonDraggableView 2
     - EffectViewWithoutForegroundVibrancy 1
     - EffectViewWithoutForegroundVibrancy 2

NSView hierarchy after:

 - window contentView
   - ChildView
     - ViewRegionContainerView
       - NonDraggableView 1
       - NonDraggableView 2
     - ViewRegionContainerView
       - EffectViewWithoutForegroundVibrancy 1
       - EffectViewWithoutForegroundVibrancy 2

This allows us to give those container views a new sibling view which stays
fixed in z-order with respect to the NSViews that get created by
mNonDraggableRegion and mVibrancyManager. More specifically, I'm going to add a
view for the drawing of our ChildView ("PixelHostingView") which is going to be
a direct child of the Gecko "ChildView" and a sibling of the
ViewRegionContainerViews; the PixelHostingView needs to always stay on top of
the vibrancy views.
Without the wrapper around the vibrancy views, whenever the vibrant region
changes, a vibrant view would be placed on top of the PixelHostingView and the
order would be wrong.

Differential Revision: https://phabricator.services.mozilla.com/D19598

UltraBlame original commit: 3507a9af7c90cd08ff79161156f1facffcb0b4f9