Bug 1518121 - Part 2: Update ExplicitChildIterator correctly when we are at the end of child list of a slot parent; r=smaug
The mIndexInInserted keeps pointing to the last assigned node of a slot parent
when we are at the end of child list. So when we calles GetPreviousChild when we
are at the end of child list, ExplicitChildIterator will skip last assigned node.

Differential Revision: https://phabricator.services.mozilla.com/D19651

UltraBlame original commit: db9ccc4f8c6ee8c425dfe4a234dbd6c30cac1c47