Bug 1512504 - Don't set __PRETTY_FUNCTION__ on clang-cl. r=froydnj
clang-cl defines it on its own, although the value is slightly different
from __FUNCSIG__ (it doesn't contain the ABI, which doesn't really
matter). We've only been setting it this was on clang-cl by extension of
setting it for msvc.

Depends on D19616

Differential Revision: https://phabricator.services.mozilla.com/D19617

UltraBlame original commit: b29c87add05f735b250612ca2444103652750091