Bug 1526486 - Don't muck with the RCD-RSF composition bounds clip rect during event delivery. r=mstange
In bug 1521644, we changed the clip rect computation so we don't divide by
the resolution when using containerless scrolling. The rationale for that
change was that this clip becomes a clip rect on the async zoom container
layer, which is applied after the resolution in the compositor.

However, when building a display list for event delivery, this rationale
does not apply as there are no layers and no compositor involved.

Differential Revision: https://phabricator.services.mozilla.com/D19221

UltraBlame original commit: 11133b5a9855c400b396312b6b11c3f00779560d