Bug 1528240 - Fix inconsistent return type use in BaseCompiler::emitInstanceCall.  r=lhansen.
emitInstanceCall() generates code to call a SymbolicAddress:: builtin.  One of
the parameters to emitInstanceCall() is the ABI-level return type of the call.
If that is non-void, the returned value is pushed on to the compiler's operand
stack.

There are some functions, for example Instance::memFill, which return a value
(an ExprType::I32) which is an implementation artefact, and which we don't
want to push on the operand stack.  As it stands, in such cases we lie to
emitInstanceCall() by telling it the return type is ExprType::Void, so as to
stop it pushing the return value.  This inconsistency is confusing, and more
importantly it gets in the way of ongoing work to move towards having
single-point-of-truth definitions of argument and result types for
SymbolicAddress:: builtin functions.

This patch:

* gives emitInstanceCall a new boolean parameter, saying whether or not the
  return value is to be pushed

* defaults that parameter to |true|, so most uses don't have to change

* for the few emitInstanceCall cases where the return type is non-void but we
  don't want to push it, passes |false| as required

* renames auxiliary function pushReturnedIfNonVoid to pushReturnValueOfCall,
  as that better describes its modified role

UltraBlame original commit: 55806c7e28b769963342000ee736246225f2b19d