Bug 1507493 - Update the gc feature version number to 3. r=jseward
Since the encoding of in-flight operations for the reftypes and GC
proposals have changed, we must update the gc-feature opt-in version
number (that's why it's there); this requires a small validator change.

Alas this means a large number of test cases must also be updated.

Differential Revision: https://phabricator.services.mozilla.com/D19795

UltraBlame original commit: 03c225b8ccfc781305c4facd922d1aca5752a873