Bug 1527570 - Support ref.null in passive element segments. r=jseward
The bulk memory proposal allows for null values in element segments;
in the absence of the reftypes proposal, these are only "anyfunc"
segments.  Null is a reasonable function value (it's the default value
for table slots).  A null value is an [Op::RefNull; Op::End]
initializer where Op::RefNull is an opcode that comes from the
reftypes proposal but is included also in the bulk memory proposal.

Thus we accept "ref.null" as text syntax to indicate a null value in
the text format, and we accept the [Op::RefNull; Op::End] initializer
expression in the passive segment to denote a null initializer value
for that element.

Differential Revision: https://phabricator.services.mozilla.com/D19754

UltraBlame original commit: 9e51f66c0b8175795618a9faf05909c271b5825a