Bug 1524779 - Use a separate button to show all preferences. r=mconley,flod
This allows interacting with the page without using the keyboard in the rare cases where this may be necessary. The button is ghosted because search is the preferred interaction.

Pressing ESC allows resetting the input field without the performance impact of showing all preferences.

Differential Revision: https://phabricator.services.mozilla.com/D19162

UltraBlame original commit: 329374773b81d379ada1d33abf01436310d35331