Bug 1529274 - Remove debugging code for bug 1153929 r=JuniorHsu
The code was added to diagnose an issue that was likely due to heap corruption.
The vtable of the objects was sometimes 0. More details in bug 1153929.
That crash doesn't reproduce anymore so it should be OK to remove.

Differential Revision: https://phabricator.services.mozilla.com/D20561

UltraBlame original commit: 6a72b418b277306c55bf2f6901827db407e23e38