Bug 1526257 - (part 1) Add a SafeZoneCellIter, making cellIter() GC-safe r=sfink,arai
ZoneCellIter can return cells that are about to be finalised in the next GC
slice.  These cells can be used briefly but cannot be kept around and the GC
needs this behaviour.

This change introduces ZoneCellIterSafe which will not return about-to-be
finalised cells.  Zone::cellIter() will use it now, the old behaviour is now
in cellIterUnsafe().

Update several uses of cellIter() to cellIterUnsafe() where that behaviour
is needed.

Differential Revision: https://phabricator.services.mozilla.com/D19890

UltraBlame original commit: d8ca9f5b4c24497b2d77fca5593f327615ced10e