Bug 1508559 - Reorganize Baldr globals for write barriers. r=jseward
We reorganize the nodes for handling global variable access so as to
allow for more shared code paths and to allow for write barriers on
ref stores.

For ref stores, the store is split in two: one part to compute the
address of the slot to be stored into (direct or indirect) and one
part to perform the store.  The address computation is different for
indirect and direct globals, but the ultimate loading and storing of
the global's value can then be handled by the same code in both cases.

For all loads, and for non-ref stores, the LIR and code generation
paths have been merged into WasmLoadSlot/WasmStoreSlot nodes.

Differential Revision: https://phabricator.services.mozilla.com/D20317

UltraBlame original commit: fbd22c49c63c6d582b910623e5663bf0ea50d63f