Bug 1436906 - Make Selection::NotifySelectionListeners() not scroll when it moves focus r=smaug
When Selection changed into an editing host,
Selection::NotifySelectionListeners() moves focus to the editing host.
In this case, we've scrolled to the focused element because it's our consistent
and traditional behavior.  However, Chrome does not behave so.  Therefore,
we should not scroll in this case for compatibility with Chrome.

Differential Revision: https://phabricator.services.mozilla.com/D19141

UltraBlame original commit: f5c2da8845b4bead4154f0946d4afbea87fcee19