Bug 1500941 - Add searchfox task for Android (ARMv7/API16). r=nalexander
The bulk of this is copy/pasted from a standard android-api-16 debug
build. The only changes are a few extra environment variables in the
taskcluster config, the subconfig file, and the mozconfig, as well as
the --enable-mozsearch-plugin flag in the mozconfig.

Depends on D20766

Differential Revision: https://phabricator.services.mozilla.com/D20767

UltraBlame original commit: 6e669b1b18355464ae76b94c4f1c89888f77c734