Bug 1529615 - Don't scroll to bottom if there's no need to; r=Honza.
We are trying to scroll to the bottom on componentDidMount and
on message addition, even if the console output does not overflow.
Since scrollToBottom can take a dozen ms (or more), we should
avoid to call it when there's no need to.

Differential Revision: https://phabricator.services.mozilla.com/D20688

UltraBlame original commit: 2a52513df092ecb5cb7bf51b33adbbbf0267da4f