Bug 1529642 - When using a search engine  alias, the only results that should be shown are search suggestion results, including in private contexts r=mak
We already have a check for stopping the search when using an  alais or restricting to suggestions. Move the check outside the block where we add suggestions so that it's used regardless of whether we added suggestions. Update the existing test to run in a private context.

Differential Revision: https://phabricator.services.mozilla.com/D21140

UltraBlame original commit: 5ad94d06e585bb9abcb38af273900e06f85a7a7f