Bug 1526604 [wpt PR 15194] - Fix BGPT backdrop-filter with opacity<1 parent, a=testonly
Automatic update from web-platform-tests
Fix BGPT backdrop-filter with opacity<1 parent

Previous to this CL, a layer with opacity<1 that contained a single compositing
child with backdrop-filter would not be properly rendered, due to an
optimization that removed the render surface for the opacity<1 layer. That
would cause the backdrop-filter to filter elements higher up the tree than it
should, due to the lack of isolation imposed by the opacity<1 layer. With this
CL, this situation is properly detected and a render surface is assigned to the
opacity<1 layer.

As part of this CL, backdrop-filter was moved from the Filter node to the
Effect node. This is required, because a single element with both opacity
and backdrop-filter need to be contained in the same node, so that the
render surface created for the backdrop filter also sees the opacity.

Bug: 497522, 836885
Change-Id: Ie6ad38231ddfda02fae8d574ca8f30d075c673e5
Reviewed-on: https://chromium-review.googlesource.com/c/1399467
Commit-Queue: Mason Freed <masonfreedchromium.org>
Reviewed-by: Chris Harrelson <chrishtrchromium.org>
Cr-Commit-Position: refs/heads/master{#629835}

--

wpt-commits: 7c61bbcbdaebae56e093017387b21b473070bd8d
wpt-pr: 15194

UltraBlame original commit: d2177892139f5df933bf005cf904f77e7f3fc5a2