Bug 1463128 - Expose setInputValue and getInputValue at hud level; r=Honza.
Now that the JsTerm is a React component, we shouldn't
let external consumers have to get the instance in order
to perform some actions on the console input.
We start with the most used actions, setInputValue and getInputValue,
and might extend that in the future.

Differential Revision: https://phabricator.services.mozilla.com/D20685

UltraBlame original commit: 63ca9c12292c99d7c54d9078d30ed195ae44c67d