Bug 1530220 - part4 : separate tesing web audio with gum to another test. r=cpearce
Testing web audio with GUM is not related with the original purpose of this test. In order to reduce the complexity of this test, separate this part as another new test.

Also modify the naming of functions, because Mozilla coding style is to use the camelCase for the name.

Differential Revision: https://phabricator.services.mozilla.com/D21257

UltraBlame original commit: 07c49703b70c4cd01985771942e84aa37a66663c