HeadersPanel: unecessary isRequired on PropTypes (Bug 1530323) r=Honza
Before this change, member was considered a required Prop on the file
HeadersPanel.js. The component itself wasn't using this prop, it was only passing to the
renderValue as a prop, and then renderValue uses member to render the data. So, the simpler
solution is remove isRequired from the PropTypes.

Differential Revision: https://phabricator.services.mozilla.com/D21539

UltraBlame original commit: ad6491a3248159b88e915c13de7e70af8babd568