Bug 1532285 - Handle zero-length operations at the edge of memory/table properly.  r=jseward
Our bounds checking is a little too stringent - we disallow
zero-length operations at the edge of the memory, table, or segment,
but the spec allows that, a behavior inherited (from active segments)
from the MVP.  It's also a natural consequence of how we handle bounds
checks for non-zero length operations.

So loosen the tests slightly and adapt the test cases, testing both at
the edge of the table/memory/segment and one past that.  In the former
case we allow the operation, in the latter not.

Differential Revision: https://phabricator.services.mozilla.com/D21929

UltraBlame original commit: 97b4af9b538e1e93da307dbcf4adeb757bc4cd87