Bug 1508241 - Change variable colors and set a new one for approved Zebra Colors for Network Tab in DevTools r=Honza
This changes the colors of the odd table cells and the hover state in rows in the Network tab in DevTools.

I have updated the //--table-zebra-background// in **variables.css** and inserted a new one, which is //--table-selection-background-hover//.

I have created the new variable, in order to prevent it affecting other table colors.

Differential Revision: https://phabricator.services.mozilla.com/D21777

UltraBlame original commit: a7c18fe7a05098c880d776c281cfd5c656f3ebdc