Bug 1534875 - Remove SearchWidgetConfigurationActivity from search's appwidget-provider; r=VladBaicu
This was a leftover from an initial implementation which needed to track
certain events related to the user adding the search widget.
Otherwise it is not needed as we don't actually expose any widget settings to
be configured by the user before adding it.
Turns out this Activity would actually mess with our PendingIntents which would
not fire for when tapping search widget's elements.

Differential Revision: https://phabricator.services.mozilla.com/D23543

UltraBlame original commit: 182f8f4a800c0c96c234f8c7c312d3c9da2514b3