Bug 1501665 Part 12: Update RDM viewport size reporting to make it consistent with zoomed presshells. r=gl
There are two viewports: the layout (or content) viewport, and the visual
viewport. This change ensures that we are both setting and checking the
visual viewport size, which is the size shown in the RDM control bar.
It is still possible to get the layout viewport with getContentSize().

Differential Revision: https://phabricator.services.mozilla.com/D22437

UltraBlame original commit: afaf26d7df4235daef3157d74b3bcd69761c6393