Bug 1534938 - Don't scroll to bottom when opening a group. r=Honza.
The heuristic to scroll the console output to the bottom was
to only look if new messages were visible. But it can happen
that such case is triggered by only opening a closed group.
This patch tweaks the heuristic a bit to exclude the case
where we open a group, while making sure that we still
scroll to bottom if there are new, opened, group messages.
A test is added to ensure this works as expected.

Differential Revision: https://phabricator.services.mozilla.com/D23996

UltraBlame original commit: 89c31db7b51ec20a9182531f16ef7d65dd97a937