Bug 1531790 - Reset DOMContentLoaded and load metrics on Reload while Persist Logs is enabled. r=Honza
Added a new state `requestsStarted` in the timing-marker.js reducer which is initialized with the time a new reload (when persist logs in enabled) is triggered. Used this parameter to calculate the `DOMContentLoaded` and `load` metrics in the StatusBar.

Differential Revision: https://phabricator.services.mozilla.com/D23848

UltraBlame original commit: c48b22d152f23fe1f8546e5d6a239df8e3ac52b9