Bug 1537669 - Use llvm-mt.exe instead of mt.exe when available. r=dmajor
MT is part of the Windows SDK, not MSVC, but an alternative exists for
it: llvm-mt. The official LLVM distribution for Windows doesn't include
it[1], though, so we can't just switch to it.

1. https://bugs.llvm.org/show_bug.cgi?id=40152

While here, remove the .exe suffix from the check, to future-proof for
cross-windows builds.

Differential Revision: https://phabricator.services.mozilla.com/D24326

UltraBlame original commit: 40832d18745436bf9521334193728de42f3d67d5