No bug: [diffoscope] Output an empty diff if the files do not differ; r=glandium
Currently, if the files match and you try to look at the diff, you get

   {
       "reason": "file-missing-on-worker",
       "message": "Artifact \"public/diff.html\" not found at \"/builds/worker/diff.html\""
   }

which makes it hard to tell if there was an error generating a diff, or if the files matched.
This changes things to ask diffoscope to always output a diff, to remedy that confusion.

Differential Revision: https://phabricator.services.mozilla.com/D24316

UltraBlame original commit: 5643338d4df0f6ee00cfcb99ad82c9687c07a59c