Bug 1538232 - Add gtests for TestVideoTrackEncoder handling MediaDecoder seek and pause. r=bryce
With bug 1423253 we need to handle resetting future frames directly, and that
bug made the signaling of when to reset a bit more explicit, so a null image
is sent when a MediaDecoder is paused, so we don't render frames that were
already buffered, but after the pause took effect. And an image where the
time stamp is earlier than the previous frame's is sent when a MediaDecoder
seeks, as it intends to discard any previously buffered frames and render
only the new ones.

Note that this hackish way of signaling a MediaDecoder's intention comes from
the fact that we have only an append-based API for pushing frames through a
track, but the MediaDecoder's behavior is tightly coupled to the ImageContainer
API of sending frames to the compositor; SetCurrentFrames.

We will be iterating towards a SetCurrentFrames-like API for video tracks in the
future, to have a cleaner solution to this problem.

Differential Revision: https://phabricator.services.mozilla.com/D24521

UltraBlame original commit: 2278142d75e8f5e02dc666b44b367bc2d7e3683c