Bug 1352113 - Shippable Builds - Make source-test depend on shippable instead of pgo. r=aki
This is useful in order to not have to run linux64/opt on push, especially on autoland/inbound
when we need a source test. It is also required if we remove the linux64-pgo build type entirely.

Differential Revision: https://phabricator.services.mozilla.com/D23124

UltraBlame original commit: b31cbe91bdf78a50a13d897356a8004b632b5fb9