Bug 1539204 part 1.  Wrap NonNullHelper() around subject principal args when the stack value is NonNull.  r=qdot
This is legwork for adding MOZ_KNOWN_LIVE, because it puts the conversion to
nsIPrincipal& inside the future MOZ_KnownLive.  Most of the changes are just
moving code around a bit and switching prepend() calls for append() calls as a
result.

I have verified that the only change in the generated code is the
addition of NonNullHelper() calls.  A diff of the generated code is available
at <https://phabricator.services.mozilla.com/F1241622>.

Differential Revision: https://phabricator.services.mozilla.com/D24926

UltraBlame original commit: a318c91c2cf0fb2ff891c8150dc32df08b05b150