Bug 1525762: Part 1i - Allow overriding some localized properties for built-in add-ons. r=aswan
We still don't have a great way to bundle localizations for built-in add-ons,
since our localization tools are not compatible with the WebExtension
localization format, and there's no way to fetch locales from language packs.

We used to have a fairly complicated mechanism for this which used localized
preference values, pointing to localized string bundle URLs, based on the
add-on ID and the property we wanted to localized. That seems needlessly
complicated at this point, so this patch just allows overrides in two specific
string bundles, one for toolkit, and one for app-specific extensions.

Differential Revision: https://phabricator.services.mozilla.com/D24623

UltraBlame original commit: 7c7df3ee8aaa9367dfc9715edd95586acfb2627a