Bug 1510569 - Filter TabChild nsIWebProgress events through a nsBrowserFilter r=kmag
The implementation of WebProgressChild.jsm filters all nsIWebProgress events
through an nsBrowserFilter. We need to do this in the TabChild as well, but
TabChild currently only handles onContentBlockingEvent, for which
nsBrowserFilter only calls the underlying listener.

We also unregister the progress listener much earlier in TabChild::Destroy as
this method causes OnStateChange event to be triggered later on, which we are
not interested in capturing.

Differential Revision: https://phabricator.services.mozilla.com/D24805

UltraBlame original commit: 1ba9a2cebf1252ffc1af89e3e978cf4b47fcb46e