Bug 1539820: Fix test_claim_locked.js to work with dev-edition. r=Gijs
The original name of the profile is irrelevant here since we aren't testing
a migration to profiles-per-install. On dev-edition DEDICATED_NAME is the same
as PROFILE_DEFAULT so we correctly make a new profile with a unique name.

Differential Revision: https://phabricator.services.mozilla.com/D26230

UltraBlame original commit: d086b1659894901f130fa9b89071f3d1d1bc2620