Bug 1541513 part 7.  Stop using AutoJSContext in XPConnect.  r=mccr8
The change to the second FindTearOff call in XPCWrappedNative::GetNewOrUsed is
fixing a longstanding bug that was introduced in bug 903891 when the sense of
the check was incorrectly reversed.  Luckily that code is unreached in
practice, because the two PreCreate hooks we have left never create the wrapper.

Differential Revision: https://phabricator.services.mozilla.com/D26007

UltraBlame original commit: aa44b9bb3b6ab95ee008065fdd3c51289a049702