Bug 1373835 - Add a test case for unreachable snap positions. r=botond
The snap alignment position of the target element is the top left of the target
and the position is located out of scroll port (top: -100px, left: -100px).
Even so we try to snap a position as much as possible.

From the spec [1];
 If a snap position is unreachable as specified, such that aligning to it would
 require scrolling the scroll container’s viewport past the edge of its
 scrollable overflow region, the used snap position for this snap area is the
 position resulting from scrolling as much as possible in each relevant axis
 toward the desired snap position.

[1] https://drafts.csswg.org/css-scroll-snap-1/#unreachabLe

Depends on D21630

Differential Revision: https://phabricator.services.mozilla.com/D21631

UltraBlame original commit: 29466eb6ee253cfd9eb21d6cf6e8d31a0bdd0c6d