Bug 1539694 - Part 2: Implement Promise.allSettled stage 3 proposal. r=jorendorff
Nightly-only for now because the proposal only just reached stage 3.

Promise.allSettled is similar to Promise.all, except that each element also has
an own reject handler. Apart from that most code from Promise.all can be reused
for Promise.allSettled, including calling the `CommonPerformPromiseAllRace`
helper function.

Because each element has an own reject handler and we need to track if either
function of the resolve/reject handler pair was already called, it's not
possible to reuse the same trick as in `PromiseAllResolveElementFunction` where
the data-holder slot is used to track if the handler was already called. Instead
`PromiseAllSettledElementFunction` uses the values array to check if the
current index position is still set to `undefined` as a mean to verify that the
resolving functions for each element weren't already called.

Differential Revision: https://phabricator.services.mozilla.com/D25209

UltraBlame original commit: 9378daa99e07d9b353473b9fd4c0b7857d1ccb8d