Bug 1543616 - Call TextureHost::PrepareForUse() when mCompositableCount becomes from 0 to 1 r=nical
By Bug 1529870, the PrepareForUse() is called in WebRenderImageHost::SetCurrentTextureHost(). It works with single buffer mode android SurfaceTexture for WebGL. But it does not work well with video's SurfaceTexture, since multiple TextureHosts are received and a TextureHost might be skipped. The timing of mCompositableCount becomes from 0 to 1 could be used for calling PrepareForUse().

Differential Revision: https://phabricator.services.mozilla.com/D27044

UltraBlame original commit: 02b69751e1c9144235a15ee368fe7667f814903a