Bug 1543549 - Only switch_to_window() if we have a handle to it. r=jgraham
On android wpt reftests, the window handles list is empty, so we can't use it to switch to a window. Not doing this in that case doesn't seem to break anything, and not doing it in that case prevents the IndexError from being logged to the error log.

Differential Revision: https://phabricator.services.mozilla.com/D27177

UltraBlame original commit: df13d5a937b5f2c94ebd90a41c0ad8134cbf577b