Bug 1543128 - Fire UAWidgetSetupOrChange for the no-controls <video> case on Desktop. r=jya
The Picture-in-Picture toggle buttons are now part of the video controls UAWidget
bindings, so we need to construct a UAWidget for the no-controls case for Desktop
to make that toggle available.

Up until now, we've never needed a no-controls UAWidget for Desktop, since we
never needed to show UI in that case.

Differential Revision: https://phabricator.services.mozilla.com/D26803

UltraBlame original commit: 1692fc6491a02f093482f6468a4056e1e4d65a9a