Bug 1260877 - Display a Show content messages checkbox in Browser Console. r=bgrins.
A preference is added to enable this feature, another one
to store the last value of the checkbox.
When unchecked, the console hides all the messages that
don't originate from a chrome contect, via the `chromeContext`
property on the message.

Since we already have a test to check that content messages are
displayed in the console output, we use it to assert the
effects of the "Show content messages" checkbox.

Differential Revision: https://phabricator.services.mozilla.com/D26337

UltraBlame original commit: 8186ed6d36ff70955f5ac0b6f8b2a79e71c55a8d