Bug 1541290 - Restart driver properly on revive method. r=padenot
Revive method of AudioCallbackDriver was wrong because it was re-initializing an already initialized driver. That was hitting an assert. Instead of that it should stop the drained stream and start it again. Also, mStarted member should be reset properly on the stop method.

Differential Revision: https://phabricator.services.mozilla.com/D26678

UltraBlame original commit: 510178f0dc32ee3c768c6c8f413bf16905c3e422