Bug 1545707 - Rework hidden-container-001.html CSS transitions tests; r=hiro
This behavior is not actually clearly specified yet.

See:

  https://github.com/w3c/csswg-drafts/issues/3790

But when it is specified it will most likely be specified in terms of HTML's
"being rendered" definition:

  https://html.spec.whatwg.org/#being-rendered

I had planned to specify this behavior before updating these tests but I need to
update them first in order to add the ::marker tests for this bug.

Differential Revision: https://phabricator.services.mozilla.com/D28174

UltraBlame original commit: 6ccf33c9e56e44937858a86890a65c4296861d98