Bug 1545707 - Add tests for animation of ::marker pseudo-elements; r=mats
In particular, this tests the composite order of ::marker pseudo-elements and
dispatching cancel events for these elements.

It adds one test that is known to fail in Gecko. This will be fixed in the next
patch in this series.

Differential Revision: https://phabricator.services.mozilla.com/D28175

UltraBlame original commit: bc8f94bb6bb6dad3f71b1ac955f1d89643548013