Bug 1523303 - [taskgraph] Define suite "categories" rather than flavours task configs, r=gbrown
Currently we have the concept of a "suite" and a "flavour" in our task
configuration.  Typically, the "suite" refers to the high-level test harness
like "mochitest" or "reftest", whereas the flavour is more specific, e.g
"browser-chrome-instrumentation" or "crashtest". However the line between suite
and flavour is not applied with any semblance of consistency which results in
inconsistent naming throughout the tree.

This patch gets rid of the concept of "flavours" entirely (at least when it
comes to task configuration). A suite is a type of test run, for example:

    - mochitest-plain
    - mochitest-devtools-chrome
    - mochitest-browser-chrome-instrumentation
    - jsreftest
    - reftest
    - firefox-ui-functional-remote
    etc

There is no confusion here between suites and flavours because flavours don't
exist. However, there are a couple of places where we *do* need to know what
"test harness" is used to run a suite. These cases are:

    1. For SCHEDULES moz.build rules
    2. For the desktop_unittest.py mozharness script which takes arguments like
       --mochitest-suite=browser (this is not a compelling use of this information
       and should be refactored to work more like the android_emulator_unittest.py
       script)

So to get this information, this patch introduces a new concept of a "category"
which is the overall "test harness" that runs the suite. For many suites, the
"category" is identical to the suite name. Unlike flavours, "categories" have
no bearing on how we call or refer to the suite.

Differential Revision: https://phabricator.services.mozilla.com/D27554

UltraBlame original commit: 00c2b2449639493d03dbed26e5765838c017db7d