Bug 1544539 - Collect GEOLOCATION_REQUEST_GRANTED through 2019 r=jdm
The GEOLOCATION_REQUEST_GRANTED probe expires in the next release.
We're using this probe to target a survey, as in bug 1499851;
new followup work that we're discussing now will also require the probe.
Since we've extended collection a few times already, let's turn collection on
for the rest of the year.

Differential Revision: https://phabricator.services.mozilla.com/D27567

UltraBlame original commit: 8ca248347c5b8f645d3fb286569d6536dddb42bf