bug 1546933: remote: drop unnecessary flags for ./mach test; r=me a=doc
The tests are instrumented to automatically set the remote.enabled
preference as part of their manifests. In the case of xpcshell,
it is also technically not need it as they do not trigger the
command-line-handler.js code path.

DONTBUILD

UltraBlame original commit: b79b875c82881c48b3529daa6da012a1095367bb