Bug 1544115 - Consolidate whitelisting of Rust allocation functions for LSan WPT. r=chmanchester
This whitelisting is not as mangled as the old one, so hopefully it
will be more resilient to changes in the Rust compiler. I moved them
to the top level white list, because like with the JS leaks we are
mostly focused on C++ leaks. Any interesting Rust leak is probably due
to a C++ leak.

Differential Revision: https://phabricator.services.mozilla.com/D28727

UltraBlame original commit: cde301c4dbfc2fe7e5816331952075aa2a7989dd