Bug 1440537 - Introduce an explicit flush in text_composition_querycontent.xul. r=masayuki
The iframe test runs on an cross-docgroup iframe, even though chrome JS can
observe it in this test.

This test is relying on the getBoundingClientRect() call below in order to flush
the parent document layout as well, but that's going to stop happening (see the
bug and patch).

Differential Revision: https://phabricator.services.mozilla.com/D28302

UltraBlame original commit: 40ccb621b5d137feb6d7f82eb2b7facd8d8f837f