Bug 1546498 - Split out MaybePrefetchMemory's check into two functions r=aklotz
Temporarily just sidestep the issue in bug 1546498 (crash with latest SDK
on startup in Windows 7) by just continuing to use the old method in
Windows 7. We saw no wins in telemetry for Windows 7 anyway, so we should
investigate why that is, and why we see a mysterious crash in the fallback
code, in a followup bug.

Differential Revision: https://phabricator.services.mozilla.com/D29239

UltraBlame original commit: 1c79adcd8483d038b94b6690c65a6f868132733d