Bug 1548375 - Remove redundant canvas log info. r=baku
Now that we're reporting to the page console, there's no need to include the top
document URI, since that's obvious from the tab.

In addition, the script file and line are already added by the console reporting
path (and displayed with nicer formatting), so we don't need to add them to the
message text.

Differential Revision: https://phabricator.services.mozilla.com/D29573

UltraBlame original commit: 69fb960771d254fde9b7567fd3f527a205a60555