Bug 1533890: Don't push callee unnecessarily for jit calls r=mgaudet
When pushing arguments for a JIT call, we push the callee value onto the stack, and then immediately pop it off. This is a remnant of the old implementation.

This patch removes that wart, with a small detour to rewrite/recomment alignJitStackBasedOnNArgs for clarity.

Differential Revision: https://phabricator.services.mozilla.com/D29532

UltraBlame original commit: 98c5fbe185df062d9459c003dab7ce0d64da6003