Bug 1549283: add missing calls to `SerializeNodeEnd` r=hsivonen
In order to have one call of `SerializeNodeEnd` for every `SerializeNodeStart`
call. In the current state of the code, the absence of these calls doesn't
cause bugs, but this can change anytime `SerializeNodeEnd` is changed. Moreover,
it makes it simpler to reason about the code.

Differential Revision: https://phabricator.services.mozilla.com/D29987

UltraBlame original commit: 59306f187449c5f35614f7e86293dd53ec1f5422