bug 1549249 - build bustage follow-up r=mt,dveditz a=tomprince
SECItem uses an unsigned int to indicate its length. We need to cast a size_t
down to the appropriate size. This is safe because what we're casting will
always fit in an unsigned int on the platforms we're using (it's just the size
of the intermediate certificate we added).

Differential Revision: https://phabricator.services.mozilla.com/D30144

UltraBlame original commit: ab7db57ad2ac2ebb1818346220f2045968b49415