Bug 1536762 - part1 : use unscaled bounding box'size as a cue box's size. r=baku,heycam
When adjusting cue's position, the goal is to place cue boxes fully inside the rendering area, and place in the right place.

In order to check whether the box is correctly inside the rendering area, we have to know the correct size of the cue box.

Therefore, we implement the new chrome-only APIs to acquire this information (box's height and width), and this returned value won't be affected by CSS transformation, which is good for positioning.

Differential Revision: https://phabricator.services.mozilla.com/D27178

UltraBlame original commit: 2d810269b72e198765af3d997f3b117d832afd38