Bug 1551167 - Move import of mochitest/helpers/context.js to test instead of head.js r=pbro
Importing scripts in shared head files is hard to maintain because it forces all browser ini that use the shared file to whitelist all its imports.
The debugger context helper is almost unused in the inspector, so it would be easier to just import it in the few tests that require it.
Also removed the whitelist entry in rules' browser.ini, no test in this suite used it.

Differential Revision: https://phabricator.services.mozilla.com/D30882

UltraBlame original commit: 87e01787d465cfcb89c0e593685c15d43a30f61e