Bug 1493361 - Add a `urlCropLimit` prop in StringRep. r=Honza.
This new prop allow us to define a maximum length for indivual
URLs (as opposed to `cropLimit`, which sets it for the whole text).
The URL regex is also modified to be able to linkify messages
wrapped in curly quotes.
Some tests are added to ensure we handle this prop as expected.

Differential Revision: https://phabricator.services.mozilla.com/D29006

UltraBlame original commit: 395e27a912b3b6aea3c27ccdb8a0028c23abf223