Bug 1493361 - Use StringRep to render PageError message. r=Honza.
This allows us to benefit from the linkification that
is done there. We crop the URL at 120 chars for now. We might
consider bumping this up depending on the feedback we get.

We add both a mocha and a mochitest to make sure this work as expected.

Differential Revision: https://phabricator.services.mozilla.com/D29007

UltraBlame original commit: b87cdb4550833374f77f1bc5163e554552fb35f9