Bug 1548369 - Reimplement token previews without onMouseLeave.
This implements editor previews in a way that doesn't rely on mouseLeave events. It does this actively by using the event loop to poll whether it should clear or not, and if it's not hovered on the token or the preview, it clears the preview.

Differential Revision: https://phabricator.services.mozilla.com/D31112

UltraBlame original commit: 03b1762fb0734dc1f58ce5453a8aef8770745538