Bug 1547351 - Use offset relative to reference frame for nsDisplayRenderRoot r=mattwoodrow
Please advise if there is a better alternative. After ensuring that
extension popups are wired up to create an nsDisplayRenderRoot, the
bottom bit of the popup's content is cut off. This fixes that issue,
but I'm not certain if it is the most robust option.

Differential Revision: https://phabricator.services.mozilla.com/D31007

UltraBlame original commit: cb6320b272f90d23a044a22d6ee738f7e6bf9d24