Bug 1547034: Migrate PreferenceExperiments to allow for multiple prefs r=mythmon,leplatrem
This is part 1 of the required changes. This just addresses the
storage mechanism and any place that uses experiments in their raw
form. This updates most callers to support studies with multiple
preferences.

We update about-studies to assume only one preference. This seems
counterproductive, but studies with multiple preferences will include
a description field that obviates the need for this.

Differential Revision: https://phabricator.services.mozilla.com/D29293

UltraBlame original commit: 137380671f01c6185f61a04b3fd55aac9617da4a