Bug 1550050 - Fix intermittent failures in test_same_base_domain_3.html forcing no cache via HTTP headers, r=valentin
At the end of each test, file_chromecommon.js deletes all the cookies, but it
doesn't invalidate the network cache. Some of the files are shared between
tests. When they are loaded from the cache, we don't set the cookies and,
because of this, the tests were reporting a wrong cookie number.

Differential Revision: https://phabricator.services.mozilla.com/D31596

UltraBlame original commit: 61f89ab0895f18aac004f37ec227997311f94879