Bug 1545079 - Execute revive logic in operation thread. r=padenot
`IsStarted` may not be updated at the time `Revive()` method is executed since the update of the flag happens on the async operation. This will be the case if `Revive` is executed right after `Start`. When that happens the revive method asserts that the stream is started.

Differential Revision: https://phabricator.services.mozilla.com/D30889

UltraBlame original commit: 70a4ba81a3960e7fdec38063e2608b56b61436b4