Bug 1552712 - Reserve capacity in IntervalSet::Intersection(). r=jya
Profiling shows we can spend a lot of time in IntervalSet::Intersection().

Much of this time is spent appending elements to a temporary buffer as we
loop over the two IntervalSets' arrays comparing them. As we append, the
storage of the temporary array is reallocated as the storage must grow to
hold more elements. These reallocations are expensive, as we must copy the
old elements to the new storage.

So before starting the loop, reserve enough capacity to store the upper
bound; the minimum of the two IntervalSets' lengths. This may waste memory,
but will be faster, as expensive reallocations are avoided.

Differential Revision: https://phabricator.services.mozilla.com/D31744

UltraBlame original commit: 40b9b95aa0e0a199a5510b71c955b0d8ebfdbd58