Bug 1552855 - [mozlog] Add new optional field to test actions for expected intermittents. r=jgraham
Added a new optional field on the test_end and test_status actions called expected_intermittents
which accepts a List of expected intermittent statuses. The default is None. As it is an optional
field, upon landing of D25081, it will only show if there is an expected intermittent.

A test was added to check that expected_intermittents are logged.

Differential Revision: https://phabricator.services.mozilla.com/D31809

UltraBlame original commit: 1dad417c28ad49a3fb16130ef3b9272dc424e2c9