Bug 1530190. Allow nsIconChannel::Open to work on Windows. r=aosmond
As stated in other patches in this bug, the system call has to happen off main thread, and it has to always be on the same thread. So we dispatch the task to that thread and synchronously wait on the main thread for it to finish.

Differential Revision: https://phabricator.services.mozilla.com/D29626

UltraBlame original commit: 642316b68e6a267ce07e99d5e7506119820b437b