Bug 1552528 - Enable & disable the structured spewer when the gecko profiler starts and stops r=mgaudet
The structured spewer does not exit gracefully when the browser is closed.  This leads to incomplete JSON since we never end up emitting the closing ']' on destruction.  Another approach that lets us do this gracefully is to have the structured spewer start and stop alongside the gecko profiler.  This also lets us focus spewing on the specific area of interest such as a page load, or some janky behaviour on a website.

Differential Revision: https://phabricator.services.mozilla.com/D32135

UltraBlame original commit: 136b9bb7ebee5455f5dcd3b564d790606675d819