Bug 1550633 - part11.1 - modify 'track-element-src-change-error.html'. r=jya
According to spec [1], if the track URL changes so that it is no longer equal to URL while fetching is ongoing, we need to change the text track readiness state to`failed to load` and dispatch `error`.

So we now only fire `error` if `src` is removed during loading, but if the track element has finished loading before removing `src`, we won't dispatch `error` for it.

[1] https://html.spec.whatwg.org/multipage/media.html#sourcing-out-of-band-text-tracks:text-track-failed-to-load-3

Differential Revision: https://phabricator.services.mozilla.com/D31915

UltraBlame original commit: 2aaaa42c409b7b56b3660326c4c7ffbdb2215635