Bug 1553198: Introduce PreferenceExperiments storage format v3 r=mythmon
This format includes an "actionName" for each experiment which
identifies the source of the experiment. This makes it possible for
each experiment type to identify which experiments it should clean up
vs. which it should leave alone because they don't belong to it.

Differential Revision: https://phabricator.services.mozilla.com/D32338

UltraBlame original commit: 81d97da2c8b05e734845244cdf8a8207622e3e87