Bug 1121623 part 10.  Use a more-accurate default value for 'concrete' in Web IDL bindings.  r=peterv
The idea is that we should only generate concreate-binding (wrap methods, etc)
machinery for an interface by default if we have reason to expect that the
interface is used as the primary interface for some objects.  Two clear signals
that would indicate that are the interface being a leaf interface (with no
descendants) and the interface having a constructor.  Other cases would require
a 'concrete' annotation in Bindings.conf.

Differential Revision: https://phabricator.services.mozilla.com/D32208

UltraBlame original commit: 1bec5b571ed1d86f15386e0fedb187319e199d4a