Bug 1552565 - Fix event listener for CustomizableUI when destroying widgets; r=Gijs
The CustomizableUI does not delete the _addedEventListeners property from
the view node when the widget is destroyed. This stops the widget from
correctly having events dispatched to it after recreating it, as the
initialization code assumes that it has already been set up.

Differential Revision: https://phabricator.services.mozilla.com/D31677

UltraBlame original commit: 92835da1844508ecf594a3e98afffd1f67615304