Bug 1550422 - P19. Convert gfxPrefs::LayersWindowRecordingPath to gfxVars. r=jrmuizel
StaticPrefs doesn't support nsCString type and the changes required to support this would be rather big. Seeing that there was only a single gfxPrefs using this, and this is a "Once" pref ; we move it to gfxVars instead.

Differential Revision: https://phabricator.services.mozilla.com/D31462

UltraBlame original commit: f855f9309c8bb1f35c0d5b79545f8abab47b62e0