Bug 1536744 - Add nsGIOProtocolHandler::GetSingleton(). Make sure nsIOService::GetProtocolHandler only returns it when it's supported. r=baku
Before, we'd try to create a URI using the GIO protocol handler, and if that
succeeded, we'd return the protocol handler.
Now we can't return it if NS_NewURI succeeds, because NS_NewURI doesn't check
the protocol handler anymore. So instead we just instantiate the handler,
check if the scheme is supported, and if so return it.

Differential Revision: https://phabricator.services.mozilla.com/D30706

UltraBlame original commit: 2dead78339e0a913d9bc9fae753bb42ac7ae9039