Bug 1524280 - Part 1. Ensure we always post when freeing SharedUserData. r=jrmuizel
We now also post the releasing of the shared surface image keys and
external image ID to the main thread. This allows the current
transaction to complete before freeing the surface, which guards against
cases where the surface is referenced and released somehow in the space
of the same transaction.


Differential Revision: https://phabricator.services.mozilla.com/D32861

UltraBlame original commit: 8b66f48247b621cadc22d15b315257fed1de4bc9