Bug 1395509 - Refactor malloc allocation tracking r=sfink
This refactors the malloc allocation tracking so that the MemoryTracker object is only present in debug builds and the count of malloc bytes is kept separately in the Zone. This makes things a little clearer I think and removes one level of inlining.

Differential Revision: https://phabricator.services.mozilla.com/D33179

UltraBlame original commit: 50769742e29584f3e653bea901a6f898833f7b35