Bug 1525618 - Add an option in the settings panel to toggle warning groups. r=Honza.
We also listen for the warningGroup preference change to
toggle warningGroups in the console output.
If warningGroups were disabled, we need to loop through
the state messages to create warningGroups when needed.
A test is added to ensure this works as expected.

Differential Revision: https://phabricator.services.mozilla.com/D32716

UltraBlame original commit: 41ead3af829120a34c76a1f01472dca7c004e0e1