Bug 1518730 - Wait for both A and AAAA responses to come back before notifying the listeners r=dragana
This way we preserve the behaviour of getaddrinfo, where both A and AAAA
responses come back at the same time.
Without this Firefox will always be biased, as the first request will usually
be resolved first. So if we requested IPv4 first, we'd mostly be using IPv4.
If we requested IPv6 first, normally we'll wait for the IPv4 response to come
back too, which is functionally equivalent to the new behaviour.
However, if the pref is set network.trr.early-AAAA;true then we'd use the IPv6
response immediately, possibly leading to a failed request if the IPv6
connection fails before we have an IPv4 address to fall back to.

A test for this patch was added in bug 1542561.

Depends on D33476

Differential Revision: https://phabricator.services.mozilla.com/D33477

UltraBlame original commit: bf3e5c1163b7aca1dff433e4704612116a6a4b74