Bug 1493369 - Fix and rewrite browser_storage_updates.js r=rcaliman
### Changes

In a nutshell, the old test produced the occasional orange and was badly written so it was very difficult to follow, especially when there were failures. The new test uses a snapshot system, which is much easier to understand.

### Try

https://treeherder.mozilla.org/#/jobs?repo=try&revision=39b92ebbce9ce1ae58f0466ccc1608314257bf81&group_state=expanded
https://treeherder.mozilla.org/#/jobs?repo=try&revision=f0fdf5e246550494e41b1ad88da1258e62a63ab7&group_state=expanded
https://treeherder.mozilla.org/#/jobs?repo=try&revision=79489426ffcf177663aceb9db43efabc190de613&group_state=expanded
https://treeherder.mozilla.org/#/jobs?repo=try&revision=6ee7fbfb846be73e08023caddd1baed2e6962d6b&group_state=expanded

That is 1040 successful test runs across all supported operating systems. There is just one window leak, which is common when running a single test in this way.

Differential Revision: https://phabricator.services.mozilla.com/D33766

UltraBlame original commit: cd7e247de34bd7fadf5b79b3e2927f764b9581e3