Bug 1556831: Have safe mode write out the normal compatibility version. r=froydnj
When launching into safe mode we write out a compatibility.ini with an invalid
compatibility version in it. This is to ensure that on the next startup we flush
the caches. This is unnecessary though, we already have a flag we can set in
compatibility.ini in order to request a flush on next startup. This patch
switches to doing that so the version check on startup continues to work.

Differential Revision: https://phabricator.services.mozilla.com/D33699

UltraBlame original commit: 7ca8fe32a5d993f0eef763e0f8e17268f2798aad