Bug 1547561 - Part 3: Additional CheckParseTree pass, enabled in debug builds only. r=khyperia
This is an experiment. We'll see if the fuzzers find anything. If nothing else,
it's good to have the invariants of the ParseNode data structure documented in
code, as assertions, separate from the parser and emitter code.

We can add more checks to this as we go. It may not be fast enough to leave on
in release builds, even in Nightly, and I don't want to push my luck. So
DEBUG-only for now.

Differential Revision: https://phabricator.services.mozilla.com/D32245

UltraBlame original commit: 6f6ebd68ccad4b93cb2238b029ab2476577d0128