Bug 1554075 - Don't store MP4 Box header; read it when needed. r=jya
Profiling Twitch we see a significant amount of time at the bottom of our call
stacks in the MoofParser creating Box objects. Most of the time spend therein
is allocating and copying into an nsTArray. This is us copying the Box's header
into Box::mHeader. The reallocation and append are expensive. It turns out we
only use the box header in one place; when reporting the PSSH boxes up to EME.
So in the non-EME case, the storage of the Box header is a complete waste of
time and space.

So we can change the Box class to instead read the header from its storage when
we need to retrieve the complete PSSH boxes.

Differential Revision: https://phabricator.services.mozilla.com/D33870

UltraBlame original commit: 82dcb120bc3fbc06289b307ff8e9af0818d88ed2