Bug 1557793 part 1.  Fix incorrect template arguments in AnimationPerformanceWarning.  r=birtles
It looks like bug 1321412 changed the number of params but didn't update the
template parameter, and then bug 1364221 followed its example of not matching
the template parameter to the actual number of params.  As a result we're
passing an array with one garbage pointer to the callee, and it just happens
that the callee doesn't examine that pointer.  But that's about to change.

Differential Revision: https://phabricator.services.mozilla.com/D34234

UltraBlame original commit: 17869bc540e5fde8961a19497c3cbb4a4de631f3