Bug 1557079 - Remove install_dep from wptrunner setup.py file, r=gsnedders
This was reading the confguration from requriements.txt which only
works in the simple case where that's a list of package names. In
practice the only thing that cares about this file is tox and it already
installs the requirements files directly.

Differential Revision: https://phabricator.services.mozilla.com/D33812

UltraBlame original commit: 7148d66e15231da32c5a7e9ec279ce668c8180a8