Bug 1559131 - Add skip-if=require_signing to test_recommendations.js r=rpl
test_recommendations.js has one test that verifies that an unsigned
add-on is not recommended. In order to load an unsigned add-on, it
needs to disable signing. Signing can however not be disabled when
MOZ_REQUIRE_SIGNING=1. So skip the test if require_signing is set.

Differential Revision: https://phabricator.services.mozilla.com/D34891

UltraBlame original commit: ffa978d11ec14051f97116c0a9bfce5983519d05