Bug 1494796 - get rid of ThreadClient specifics in DebuggerClient ; r=jdescottes
This is part one of removing threadClient specifics out of the debuggerClient. We were
managing messages from the thread client in a special way -- this was the "Unsolicited Pauses"
object that we had before. This patch updates the threadClient to use Front style events. This
required updating the spec for the threadClient, and several of the methods. What has not been fully
migrated here is the "resumed" event, as this is much more complex. This is taken care of in the
next patch.

Differential Revision: https://phabricator.services.mozilla.com/D32695

UltraBlame original commit: 25e69c21dc2e27e578db2d79f5564b0493e0e2a5