Bug 1494796 - use getFront to retrieve threadClient; r=jdescottes
In order for a front to be available to getFront on a given target, it must be first --
registered on the target scope, and second -- set on the target's targetForm. This makes that update
for both browsing context and worker targets. This works as part of a work around until we can get
the server into better shape.

Differential Revision: https://phabricator.services.mozilla.com/D32698

UltraBlame original commit: fe1559f5f1d4ca7e10be2004dcf36f36389e30eb