Bug 1557635 - Update the Sync Onboarding screen; r=VladBaicu
With this changes we must now support 3 Onboarding versions.

Latest Onboarding UX will use a new title for the "Sync screen", new subtext,
new image and new text for the signin button.
This will be presented only if all this new Strings are localized.

Refactored the existing OnboardingStringUtil to serve as a central point of
getting the right Strings to be used and querying the Onboarding UX version the
app should offer.

Applied a lossless compression for the new sync image which resulted in a 26.5%
size reduction.

Removed the lint suppression initially necessary for when first added the
updated Sync Strings which were not used at the moment.

Differential Revision: https://phabricator.services.mozilla.com/D35012

UltraBlame original commit: cc09b9a55c4825e055b32c2c2e812e555283cb56