Bug 1494796 - Removing threadClient specifics from DebuggerClient Special case resume; r=jdescottes,jlast
The resume case is much more complex than the other events, because we do an
unsafeSynchronize to send an unsolicited pause. In the old system, the resume response would have
been ignored, but that is no longer the case. With the new system, we do not want to send a response
to a resume action if it did not come from the UI. This also update the debugger panel code to
accept a resume.

Differential Revision: https://phabricator.services.mozilla.com/D32697

UltraBlame original commit: 9732099eeb5693bcffc0e0ec870bb47c5d7fefec