Bug 1552978 [wpt PR 16926] - Scrollable table cell child with % height may affect table height., a=testonly
Automatic update from web-platform-tests
Scrollable table cell child with % height may affect table height.

There's a rule in the table spec (which is marked as not ready for
implementation) [1] that percentage block-size children of a table cell
do not contribute to the block-size of the cell or table, if said child
has overflow 'auto' or 'scroll'. However, this shouldn't be applied if
the block-size of the table cell is unrestricted (auto), since the child
in question may very well be the only child of the cell, and the cell
may be the only child of its row, in which case we'd end up with zero
block-size for the content.

There's similar code in the legacy engine, in
LayoutBox::ContainingBlockLogicalHeightForPercentageResolution().

Just like in legacy, we'll just check the block-size of the cell and the
table, and ignore e.g. min-block-size.

[1] https://drafts.csswg.org/css-tables-3/#row-layout

Bug: 964282
Change-Id: I312f683282d848f0ae71a27c1128a0eb16ad20f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1621166
Commit-Queue: Morten Stenshorne <mstenshochromium.org>
Reviewed-by: Christian Biesinger <cbiesingerchromium.org>
Reviewed-by: Emil A Eklund <eaechromium.org>
Cr-Commit-Position: refs/heads/master{#661682}

--

wp5At-commits: 90af87b59f220ee2a4e618ff2a06ba42ee1bac19
wpt-pr: 16926

UltraBlame original commit: 407d74b99d7e75344b09ec872f41aacc8793e7c9