Bug 1555780 [wpt PR 17105] - [LayoutNG] Fix `vertical-align: top/bottom` not to include paddings, a=testonly
Automatic update from web-platform-tests
[LayoutNG] Fix `vertical-align: top/bottom` not to include paddings

This patch fixes the top and bottom of inline boxes for the
`top` and `bottom` values of `vertical-align` not to include
their paddings.

The spec[1] says align with the top/bottom of boxes. Line
box in CSS is special that paddings are included to draw
background and borders, but are not included when computing
their heights[2]. All impls agree that the top/bottom in this
case is the latter; i.e., not to include paddings.

[1]: https://drafts.csswg.org/css2/visudet.html#propdef-vertical-align
[2]: https://drafts.csswg.org/css2/visudet.html#inline-non-replaced

Bug: 968438
Change-Id: I034d0390e3695e8a3fcc737f94fc5561e16cd617
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1637006
Commit-Queue: Emil A Eklund <eaechromium.org>
Reviewed-by: Emil A Eklund <eaechromium.org>
Cr-Commit-Position: refs/heads/master{#664862}

--

wp5At-commits: 1f7dcd9d4364946e8d2cca4aef7e264d7278cb14
wpt-pr: 17105

UltraBlame original commit: 4841ed584343ac4e06377cb7b07be84a1af59521