Bug 1558849: Temporarily disable web-platform-test ruby-inlinize-blocks-002.html on more categories of Windows testers. (no review, test-annotation-only)
This test is known intermittent on windows. It's already annotated to be
disabled for some categories of Windows testers, but that annotation was overly
specific.

I've already fixed the "upstream" version of this test (the version in
layout/reftests), so the intermittent failure will stop once that fix has been
synchronized around.  But in the meantime, let's adjust the annotation so that
it includes Win10 64-bit testers, so we don't get intermittent failures for
those testers.

UltraBlame original commit: da96f37538d8a29743b268808d1f475e17ec372e