Bug 1498044 - Don't ask to reboot when a directory we try to uninstall isn't empty. r=agashlin
Our uninstaller currently calls the NSIS RMDir instruction on the `defaults`
directory with the /REBOOTOK flag set, which means that NSIS will register
the directory for deletion during the next reboot if it can't be removed
immediately for any reason. The problem here is that "any reason" includes
"the directory wasn't empty", and setting the reboot flag is not at all
the right thing to do in that situation. It won't help resolve the problem,
and those directories are expected to be empty at that point in the uninstall,
because we've already removed all the files that we know we put there, so
anything left is something we don't know about and shouldn't try to delete.
So the right thing to do here is just to remove the /REBOOTOK and allow the
RMDir to fail, so that any user files are never touched and we don't show
an unnecessary reboot prompt at the end of the wizard.

I also removed the FileExists checks, because the RMDir instruction does that.

Differential Revision: https://phabricator.services.mozilla.com/D34938

UltraBlame original commit: e8cc9c07e8b2574b95993619a8c3d6b90bb35f7a