Bug 1557343: Replace the sole use of DebuggerWeakMap::removeIf with its definition. r=jorendorff
Later in this patch series, we will be gathering up all the code to manage the
association between DebuggerFrame and AbstractGeneratorFrame objects into a pair
of functions, one to establish a relation and the other to tear it down. The
removeif method combines iteration and entry removal, but we would rather have
entry removal live next to the code that tears down the rest of the association.

In preparation for that, this changeset replaces the sole use of removeIf with
its (not very large) definition, so that the entry removal can be more readily
moved into another function.

Differential Revision: https://phabricator.services.mozilla.com/D35606

UltraBlame original commit: a9de28d699a73c1a84e403f3db7c0cc22f44547c