Bug 1559814 - Use a more similar representation in Rust and C++ for grid lines. r=heycam
Option<> is not FFI-safe, so if we want to use the same representation
everywhere we need to get rid of it. This also makes it take the same amount of
memory as the C++ representation, and it's not very complex, I'd think.

Differential Revision: https://phabricator.services.mozilla.com/D35195

UltraBlame original commit: 362d9435ca4e6d90ee16fe2c6a9d620a03c03e90