Bug 1560746 - Fall back to default locale if localized name is missing r=kmag
Legacy add-ons can have null values instead of localized strings.
This breaks code that expects names to always be a string.
Fix this by falling back to the default locale.

This code is not needed for WebExtensions, because the name for those is
always a string, even if the translation is missing.

Differential Revision: https://phabricator.services.mozilla.com/D36266

UltraBlame original commit: 7598ffdf3ca1bb91db4b0f941860dbf65eb01d8f