Bug 1549631 - Add new target app (browser) backgrounded, idle-power-use test. r=rwood,AlexandruIonescu
This patch adds idle browser (*-idle-bg)tests to Raptor which have the app running on about:blank in the background for 20 minutes. It adds the 'background_test' flag to 'scenario' type Raptor tests to make a test run in the background. Furthermore, a unit test is added to the Raptor control server tests as well as methods to check if the app was properly backgrounded or foregrounded.

Differential Revision: https://phabricator.services.mozilla.com/D32428

UltraBlame original commit: c532dce1d92a49744074fe3e300892ceb6f81f08