Bug 1562159. Restore second copy of test because pref is different second time.
I forgot that we run the tests twice in this file, once with the downscale pref on and once with it off. Restore the second time for the moz-icon tests. And adjust the fuzz because it can be lower now.

Differential Revision: https://phabricator.services.mozilla.com/D36415

UltraBlame original commit: 2f25d26ffa94129eae86d5b617883e1b41ca57bd