Bug 1558598 - Add the --enable-webrender option to android HW unittests. r=ahal
This is not used yet but will be eventually so I'm just going to
add it now.

As of this patch, all the tests that we currently run on android HW do
accept the --enable-webrender flag and explicitly disable WR if it is
not provided.

Differential Revision: https://phabricator.services.mozilla.com/D35868

UltraBlame original commit: 67f20ed19e3ee867e29a94bce1c30c293ad6ed87