Bug 1561832 - Separate setting the current chunk from poisoning it r=jonco
This allows us to run the poisoning code after resizing the nursery,
ensuring that the correct region of that chunk is poisoned, fixing the bug.

This also simplifies the logic around how much of the nursery to poison, we
always poison the valid region of the nursery regardless of how much was
used (removing an earlier optimisation).

Differential Revision: https://phabricator.services.mozilla.com/D36315

UltraBlame original commit: 75e28014a6bbaebf9d5226772ddd43689114089c