Bug 1563066, part 3 - Use refptrs in more places in XULContentSinkImpl. r=bzbarsky
There are a few places that use XUL prototype nodes without ref
pointers, which is dangerous. A benefit of this change is that it
removes a direct call to delete on a refcounted class. It does result
in some additional refcount traffic, but hopefully that doesn't
matter.

Differential Revision: https://phabricator.services.mozilla.com/D36790

UltraBlame original commit: fd915382080de872b69a6245d8e8477b0dd543ce