Bug 1562830 - Keep Baseline Interpreter bytecode pc in a register between VM/IC calls. r=djvj
This is both simpler and faster than the old scheme where the pc was stored in
a register but could be clobbered by R2.

On x64 this wins about 9-10%. On 32-bit x86 we don't have enough registers so
there we load the pc from the frame in more cases. That's about a 2-3%
regression and is a reasonable trade-off.

Differential Revision: https://phabricator.services.mozilla.com/D36583

UltraBlame original commit: 874fd6880afc7776b5d473cb186252a85a9c71bc