Bug 1562602 part 4 - Move jitSupportsFloatingPoint and jitSupportsUnalignedAccesses flags from JSRuntime to JitOptions. r=tcampbell
This is process-wide state and storing it in JSContext/JSRuntime requires
callers to pass a context/runtime for no good reason. This simplifies the
next patch.

Storing this in JitOptions seemed nicer than adding new globals for this and
maybe in the future we will have a testing function/flag to toggle these
flags.

Differential Revision: https://phabricator.services.mozilla.com/D36471

UltraBlame original commit: 8cbfd166cb6173c48dc3994399892046c983409e