Bug 1554208 - patch 4 - Make the FontNameCache that we use to store the font list in the startupCache use a more robust and extensible serialization. r=jwatt
I'll need to add a couple of extra fields to the cache records, and realized that the current format
looks really fragile; in theory, it'd break if a font name ever contained a comma or semicolon
(unlikely though that may be). So let's fix it to be a bit more robust before we add to it further.
Using control characters from the C0 range to delimit fields/records, instead of ASCII punctuation,
removes the (tiny) risk of conflict with characters that actually occur in a name, and using
distinct field and record separators means that we can better check that the records we're loading
from the cache actually match the expected format.

(Given that the startup cache gets re-created when the build ID is updated, a change in the format
between versions wasn't going to cause problems for users; their old cache just gets blown away
when upgrading. Still, a little more robustness seems like a good thing.)

Differential Revision: https://phabricator.services.mozilla.com/D36109

UltraBlame original commit: 3e6c11ef906c79d7073bf3f04a5277e565cf66a4