Bug 1553515 - Always send navigation events for about:blank. r=snorp
Some background information on this change:

DocShell loads about:blank when first starting up. This initial load can be
avoided by setting `nodefaultsrc` on the browser element, but this will still
cause some load events related to `about:blank` to fire anyway, as they are
specified in the DOM spec (see e.g. bz's Comment #2 Bug 1447406). In particular
`onSecurityChange` and `onLocationChange` are still fired, `onPageStart` and
onPageStop` are not.

These messages are unreliable and sometimes do not fire, to unrestand why it's
helpful to look at a timeline of the events when starting up a GeckoSession.

page           about:blank ---------------------------------------------------------

nsDocShell     --- startup --- onPageStart --- onLocationChange --- onPageStop -----

busyState      idle ---------- busy ------------------------------------------ idle-
                ^           ^               ^                    ^               ^
geckoview.js   (0)         (1)             (2)                  (3)             (4)

geckoview.js loads in parallel with nsDocShell startup process (and loading of
about:blank). This means that consumers of GeckoView might start receiving
navigation events at any of the points marked with (0) - (4). E.g. if geckoview
starts up at (4) a consumer waiting for onPageStop for `about:blank` will
actually wait forever (this is what happens before this change for our tests).

As it can be seen there's not really a way to detect in which of the (0) - (4)
state DocShell is when starting up geckoview.js. Checking for the busy state is
not enough as an `idle` busy state might mean that we're in (0) or (4).
Furthermore a consumer of geckoview has no way to know which onPageStop
messages to wait for after an initial loadUri as e.g. the following is what
would happen if geckoview.js happens to startup at (3):

- loadUri(hello.html)
- onPageStop (for about:blank)
- onLoadRequest (for hello.html)

which confuses any code that just waits for onPageStop.

Desktop deals with this in `TabProgressListener.onStateChange` where the
initial `about:blank` navigation is ignored and fake events are triggered at a
convenient time.

To patch implements a very similar behavior for geckoview, we ignore the
initial `about:blank` `onLoadRequest` call, set `nodefaultsrc` so `onPageStart`
and `onPageStop` events don't fire and fire the above calls when the
`GeckoViewProgress` module has finished loading.

This makes `about:blank` events deterministic with the exception of reloading
an empty `GeckoSession`, which will often not fire any extra events. To account
for that we load a dummy html page before the tests that used to rely on this
behavior (which would actually fail occasionally due to the startup race
condition explained above).

This makes the tests pass reliably on x86_64 (20/20 runs passed in try).

Differential Revision: https://phabricator.services.mozilla.com/D32586

UltraBlame original commit: cabf177300f4f47eddb7c7ae9b5113faf7cb1109