Bug 1563178 - Properly manage the renderRoot variable during the APZ tree walk. r=botond
The APZ tree walk is recursive but the render root was not being updated when
walking up out of a subtree with a different render root. This changes the
code to use a stack and push/pop the render root for subtrees as we enter
and exit the subtrees as part of the tree walk.

Differential Revision: https://phabricator.services.mozilla.com/D37614

UltraBlame original commit: dc058b3425f758eadb09e6c18d872d8635b4a23d