Bug 1564299: Fire a reorder event on the outer doc accessible when an OOP document gets added. r=yzen
For same-process iframe documents, this reorder event is fired by the content process.
This isn't possible when the document is in a different process to its embedder.
In this case, we need to fire it from the parent process when the embedded document accessible is added.

Differential Revision: https://phabricator.services.mozilla.com/D37352

UltraBlame original commit: bf25c3b58362c73e91a1587c1caaeb33f70de53b