Bug 1563797 - Vendor backports.shutil_which r=Callek
I couldn't use 'mach vendor python' to get this module in-tree because that
puts a 'backports' module on the PYTHONPATH. This is a problem because many
(most?) compatibility backports also use a 'backports' library. So in the
likely event that a user happens to have one of these installed in their
system Python, we'll search that module before the vendored one and fail to
find it.

This gets around the issue by foregoing 'backports' and just putting
'shutil_which' on the path

Differential Revision: https://phabricator.services.mozilla.com/D36838

UltraBlame original commit: e29cc4ee9158b4d1f9482628c5fb7bd689ee7858