Bug 1567108 - Fix bug 1527210 in a simpler way. r=hiro
This is IMO the right RestyleManager change for what bug 1527210 tried to fix.

We need to apply the animation hints to the primary frame, not the style frame.
The other non-RestyleManager bits of that bug still apply and look fine to me.

Differential Revision: https://phabricator.services.mozilla.com/D38598

UltraBlame original commit: 47ff979267bedeb347c2597068dc7e5009cb1845