Bug 1566332 part 5 - Remove support for C++ interpreter -> Baseline JIT OSR, always enter Baseline Interpreter. r=tcampbell
If the script already has a BaselineScript or the Interpreter/JIT thresholds are the same,
we will immediately switch to Baseline JIT code at the same LOOPENTRY op. This should be
unlikely for most loops and the extra overhead for this case is negligible.

Differential Revision: https://phabricator.services.mozilla.com/D38280

UltraBlame original commit: db5038b9b3d3b3d8d8296b7ee5f978cc1efec8ac