Bug 1567752 - Keep mInFontFaceset and mFontFaceSet in sync. r=heycam
When we unlink, we may null out the mFontFaceSet pointer, but still keep
mInFontFaceset as true. Which can cause null dereference bugs as described in
comment 0.

This is mostly a speculative fix since I think that's a plausible cause for a
bug like that, but it may not be the only cause for the crash.

Differential Revision: https://phabricator.services.mozilla.com/D38822

UltraBlame original commit: d7abe3bfd99331c88af78cc9cdb0b10184655d41