Bug 1528616 - Make an assertion fatal to figure out how this can happen. r=dholbert
We shouldn't have a frame and a null root frame. This assertion failing is the
only way to get this to happen.

So assert it a bit harder so that we can hopefully find a way to repro it and
thus figure out what the right thing to do is. If it legitimately fails, chances
are we shouldn't be putting this function on PresShell in the first place.

Differential Revision: https://phabricator.services.mozilla.com/D39124

UltraBlame original commit: bf824987ef9f41d0c2fc3dec516ff4d1f34b5c32