Bug 1568617 - Add a `newtab` option to browser.urlbar.topSites.get that returns the top sites exactly as shown on newtab r=mixedpuppy,Mardak
This is based on D39589, which moves the top sites API from toolkit to browser.

* Add a `newtab` option to `browser.urlbar.topSites.get` (similar to the abandoned D36200) that returns exactly the top sites shown on newtab
* Add a `AboutNewTab.getTopSites` function that returns those sites from activity stream
* Keep the changes made in bug 1547669

There are a couple of missing things related to the default top sites that this doesn't address but ideally we would have. I think we can come back to these if necessary.

* Actual favicons for the default top sites instead of using the bigger tile images since they don't seem to be the same for every site.
* Proper names names for the default sites. There's a `hostname` property, but it would be nice to have e.g. "YouTube" instead of "youtube".

Differential Revision: https://phabricator.services.mozilla.com/D39591

UltraBlame original commit: b33efbfec65ca65c509806ebf4225b4e4e365cad