Bug 1567438 part 1 - Replace the enter-monitor-chain bailout code with type monitoring in C++. r=tcampbell,iain
When we resumeAfter a JOF_TYPESET op, we had code to enter the type monitor
chain for the bailing op, passing it the top stack value. It's simpler to do
this monitoring in C++ in FinishBailoutToBaseline and then treat it as a regular
resumeAfter bailout.

The code for this will be simplified more later in the stack.

Differential Revision: https://phabricator.services.mozilla.com/D38860

UltraBlame original commit: 58ea17fc315ef8bc3a325a958953a50e047f4499