Bug 1533048 - Improve wpt expectation ini error messages, r=ato
Currently we don't always give a useful message and in particular
don't always include the file and line number of the place where the
exception occured in order to give a clue about where the error might
be.

To fix this, if we get an error that isn't a ParseError, wrap it in a
ParseError to give the file and line. Also improve one error message.

Differential Revision: https://phabricator.services.mozilla.com/D22312

UltraBlame original commit: 4b13ce92cd304082bc9ce84392bdb69ee7811e42