Bug 1564293 - Allow efficient slicewise upload of 3D textures r=jgilbert
Before this patch any partial upload to a texture would incur
a zeroing of the texture first to prevent leakage of information.
The texture now tracks, for each image not fully initialized,
which z-slices have been initialized, and only zeroes the rest
of the slices when the texture is used.

Differential Revision: https://phabricator.services.mozilla.com/D37843

UltraBlame original commit: 6019fc32fbf36e44a00248aa12f94c6f111a9213