Bug 1555231 - Part 1: Move the protection setting button into the footer section and change the style of the footer section. r=nhnt11
This patch makes changes to the protection setting button and the footer
section. First, it moves the protection setting button into the footer
section. Second, it updates styles of footer section. And then, it
changes the hover state of buttons.

In addition, it adds a new function in gProtectionsHandler in order to
open the 'about:protections' page when clicking the 'Show report' button.

Differential Revision: https://phabricator.services.mozilla.com/D39692

UltraBlame original commit: 35fe7ce1853452827dd27cc4b25f9f7ccbfb1235