Bug 1298602 - split condition in TestExpirationTracker; r=KrisWright
This condition is intermittently failing, but it's hard to tell from the
failure which part of the condition is failing.  Split the condition
apart so ideally we'll get some better error messages.  While we're
doing that, use a more explicit `EXPECT_LT` for the first half of the
condition for even better error messages.

Differential Revision: https://phabricator.services.mozilla.com/D40051

UltraBlame original commit: 925326628d3660e6f43edf78edfb8f66a1e35fa7