Bug 1563358 - Report pixel differences as part of message in wpt reftests, r=maja_zf
Instead of reporting the pixel differences in a logger info message,
put it in the text that's returned to the harness. This has a notable
advantage on android where this will cause it to be logged as part of
the harness logs rather than ending up in the logcat from the
device. It also makes these messages more accessible in other
consumers of the logs e.g. wpt.fyi.

Differential Revision: https://phabricator.services.mozilla.com/D40064

UltraBlame original commit: f5f5cbb377ac92a59c26a0f1de8b971ef878dee3