Bug 1388701 - Remove the special nursery size of 0 r=jonco
Rather than make re-enabling the nursery work correctly (because there's
various ways to disable it and we'd need to make sure we're enabling it for
the right reason).  It's simplier just to not disable it for a
gcMaxNurseryBytes of 0, and instead return an error.

Differential Revision: https://phabricator.services.mozilla.com/D39988

UltraBlame original commit: 99898d62af446cb6cff61223ae64e2b7e8ce4a61