Bug 1567954 - Part 4: Implement MozPower class in the Mozpower module. r=perftest-reviewers,rwood
This patch implements the MozPower class in the Mozpower module which is the interface for the user to use the power measurement tooling. It primarily detects the user's OS and CPU combination and provides the appropriate methods for them.

Differential Revision: https://phabricator.services.mozilla.com/D39194

UltraBlame original commit: ae743d95f2200a2d37f9e8b70c521b297b3dbd3b