Bug 1569135 Fix --screenshot r=kmag
This patch ressurects HiddenFrame.jsm and uses it when handling
the --screenshot command line argument to load the requested page
in a content process.  The actual logic for grabbing the image is
also ported to a JSWindowActor.  The test for this feature remains
suboptimal as described in the bug.

Differential Revision: https://phabricator.services.mozilla.com/D40148

UltraBlame original commit: 4ee4a9ba55938ef2bdae2dc6f77a7b1cbb1afc18