Bug 1562292: Part 1b.2 - Don't fall back to named lookup for inaccessible special name. r=nika
Currently, if a window with a special name is inaccessible to the caller, we
fall back to ordinary named lookup, which is not desirable.

This patch changes that behavior so that we never attempt fallback for special
names.

Differential Revision: https://phabricator.services.mozilla.com/D41260

UltraBlame original commit: a8ac7bcafc270b5be3eca84efebcde672fe8fba5