Bug 1573419 - Fix assertion failure finalizing JSObjects during shutdown r=bzbarsky
The problem is that if there is a leak at shutdown we can end up calling ClearWrapper on a wrapper that has already been cleared by the JS engine. The patch just relaxes the assertion to allow this. I tested this on try with a bunch of retriggers and it seems to fix the problem.

Differential Revision: https://phabricator.services.mozilla.com/D41938

UltraBlame original commit: 9a51c0c9a0f7b16b0fbaa675300e8b4b29add727