Bug 1491442 - Make reftest readback read from the current render target. r=mattwoodrow
This removes the assumption that there is a default framebuffer to read from,
in order to prepare for a world in which all rendering goes into non-default
framebuffers.
The current render target's framebuffer is already bound when this function is
called.

Differential Revision: https://phabricator.services.mozilla.com/D40513

UltraBlame original commit: fb949ba38402edbe1decc678c5a8ee65f446a5a5