Bug 1574521 - Make sure updateLayer is only called when the CoreAnimation pref is set. r=jrmuizel a=Aryx
updateLayer expects some objects to be non-null which are only initialized when the pref is set.
But in builds that were compiled with the 10.14 SDK, all NSViews are layer-backed by default, so
wantspdateLayer gets called, and returning YES from that unconditionally will cause updateLayer
to be called.

Differential Revision: https://phabricator.services.mozilla.com/D42333

UltraBlame original commit: 19cdc92b399fd0061f4a0af4adf0ec801a1395a1