Bug 1539979 - Fix eyedropper tests for DevTools in content frame r=ochameau
Depends on D40993

I don't fully understand the failures in this test on Linux.
The only way I could get it to pass for now was to force a window focus when the test runs in a window host.
Other OSes are fine, but without this, Linux is permafailing this.

I don't have a good linux env right now, so I would like to proceed with this and
fix it in a follow up.

Differential Revision: https://phabricator.services.mozilla.com/D40994

UltraBlame original commit: e7d974a6ba030b42858dc448dc02f6e09f63dcfb