Bug 1574977: restore `mach pastebin` and port to new service r=nalexander
This commit re-implements `mach pastebin` for use with the new Mozilla
pastebin tool (paste.mozilla.org). As with the old implementation, the
tool supports passing a file for upload as well as reading content to
paste from stdin.

`dpaste` (the software that runs `paste.mo`) is supposed to support
sniffing the correct syntax highlighter from the given filename, but
this feature does not seem to work (both on `paste.mo` nor `dpaste.de`).
Instead we have a mapping of known filename extensions to known available
highlighters on `paste.mo`. `mach pastebin` will attempt to guess a
highlighter from the file extension or the basename of the file if none
is available.

Differential Revision: https://phabricator.services.mozilla.com/D42902

UltraBlame original commit: 2120cea83d4c0d517eb7e12968e2c088a37fc12c