Bug 1576563 - Evaluate selected text if a selection exists in console input. r=Honza.
This is something Scratchpad allowed and that people
told me they were relying on. With Scratchpad being
deprecated, it makes sense to implement it in the
WebConsole as it's quite easy to do so.
A test is added to ensure this works as expected.

Differential Revision: https://phabricator.services.mozilla.com/D43394

UltraBlame original commit: 0690321bae187ea4f67eb59acef2484c3ccf8b46