Bug 1575711 - CallerSubsumes should return false for remote object proxies. r=peterv
This check is used to reject cross-origin objects from being passed in
as |any| or |object| parameters to WebIDL methods. Remote object
proxies are technically same-origin, but we want to make them behave
the same as when Fission is not enabled.

Differential Revision: https://phabricator.services.mozilla.com/D43105

UltraBlame original commit: 31f52909af179bbd4c8b2185b771c531254324a7