Bug 1576082 - Don't build jsapi-tests, gdb-tests and js tests when disabling js shell. r=sfink
For starters, js tests require the js shell, so recursing js/src/tests
make no sense.

As for the other, it feels like it's reasonably that if one didn't
opt-in to build the js shell for Gecko builds, they are not interested
in jsapi-tests and gdb-tests either.

The default for Spidermonkey standalone build being to build the shell,
nothing changes for them.

Building jsapi-tests and gdb-tests forces to build jsrust, which can be
a drag when building Firefox.

Differential Revision: https://phabricator.services.mozilla.com/D43537

UltraBlame original commit: c1835ddb51e2c60b6286a12baf3e6c35e13217ed