Bug 1575417 [wpt PR 18574] - Support percentage comparison resolution against negative reference values, a=testonly
Automatic update from web-platform-tests
Support percentage comparison resolution against negative reference values

It's a very edge case that a percentage may be resolved against a
negative value, which is allowed only in 'background-position' property.

To support this without massively changing existing code, this patch
adds an indicator flag to |CSSMathFunctionValue|, so that when the flag
is set, the resolution of calculated percentage values is delayed to when
a reference value is available. When the flag is not set, we still use
the existing code path where a calculated percentage value can be
resolved immediately.

Together with this patch is a test suite on parsing and evaluation of
comparisons on percentages.

Bug: 825895, 978682
Change-Id: I2a6aabd48313ccda82b50f6dc00fd62c4548bc84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1762970
Commit-Queue: Xiaocheng Hu <xiaochenghchromium.org>
Reviewed-by: Rune Lillesveen <futharkchromium.org>
Reviewed-by: Emil A Eklund <eaechromium.org>
Cr-Commit-Position: refs/heads/master{#689198}

--

wpt-commits: 8f929221b0664b735209dd1b44594dd21d643520
wpt-pr: 18574

UltraBlame original commit: e5e59e33b55fdcdbf3eed414a6a70ee9b2cdae4a