Bug 1473498 - [mozbuild.base] Fix Python 3 incompatibilities in mozconfig error handling, r=glandium
Use of 'BaseException.message` was deprecated in Python 2.6 (and removed in
Python 3). We should rely on the exception's '__str__' instead.

We also need to ensure the mozconfig subprocess' output is text when formatting
it into the message with Py3.

Differential Revision: https://phabricator.services.mozilla.com/D42843

UltraBlame original commit: 983a86c020c9ac14397855a5415a751bb23d897c