Bug 1575975 - Enable the maintenance service for MinGW r=firefox-build-system-reviewers,mshal
We don't need to disable it. The problem with this commit is that
we want to keep the --disable-maintenance-service flag working, so
we might want a separate build for that (and other unusual --disable
flags) - but such a build can be clearly labeled so people understand
why things broke.

Differential Revision: https://phabricator.services.mozilla.com/D43145

UltraBlame original commit: 663d87d66315567920932d793885182fcda014fe