Bug 1576991 - Ensure expected is correctly inserted merging wpt ini files - r=maja_zf
We can't parse files where the key-value pairs are after the subtest
data, so ensure that we add it before. Also fix a bug when metadata is
added on one side of the merge but unchanged on the other; we were
accidentially inserting the entire test rather than just the expected
value.

Differential Revision: https://phabricator.services.mozilla.com/D43658

UltraBlame original commit: a5f33d0358d3d308433110112d47fa804936373d