Bug 1534445 - Use same colors as webconsole warning messages in settings panel deprecation label. r=rcaliman .
In order to do this properly, we create new variable in
variables.css that hold the same values as the console-warning-*
variables we have in webconsole.css.
We take this as an opportunity to replace the box shadow on
the deprecation label by an outline. I think this was done
so the border doesn't take additional height, and outline
seems perfect for this job.

Differential Revision: https://phabricator.services.mozilla.com/D43933

UltraBlame original commit: 6b78b362900bc8e0a5f4289cae027f4d8a47f0b5