Bug 1553571. Pass the size of the display item to the StackingContextHelper constructor in nsDisplayTransform::CreateWebRenderCommands. r=jrmuizel
In FrameLayerBuilder::ChooseScale we hit this line https://searchfox.org/mozilla-central/rev/325c1a707819602feff736f129cb36055ba6d94f/layout/painting/FrameLayerBuilder.cpp#6124 and aVisibleRect is empty (because every call site except for nsDisplayMasksAndClipPaths::CreateWebRenderCommands (which doesn't need to) passes empty for the size of the bounds) and so the maxScale stays at 4, and we clamp to 4 instead of something 30-50.

The call to ChooseScale in StackingContextHelper::StackingContextHelper is the only place the size of aBounds is ever looked at. And we only ever call ChooseScale if we are passed a mBoundTransform which only nsDisplayTransform does. So this change should be quite safe.

Differential Revision: https://phabricator.services.mozilla.com/D43546

UltraBlame original commit: 5395a79f7a0b00cc89d3a4d567d542789808900e