Bug 1481951 late-breaking followup: Remove mistaken comment with a reference to this (closed) bug. r=TYLin
DONTBUILD because this is a comment-only change.

Per bug 1481951 comment 6, it seems our behavior was in fact correct even when
the comment was added, and this "// XXX" comment was just due to a
misunderstanding in what was going on in the testcase.

Differential Revision: https://phabricator.services.mozilla.com/D44229

UltraBlame original commit: 8f159b8c631d3e2d05a81644ad435ca751a8aefe