Bug 1576288 - Fix dom/tests/mochitest/storageevent/ tests to work with Fission. r=asuth
These tests compute the origin of their parent, in order to pass in as
the target origin argument to postMessage. They do this with
SpecialPowers.wrap(parent.window.location); which does not work with
Fission enabled, because the parent is often in another process. Every
web page in this directory is only ever has a single origin as a
parent, so we can hard code it as an argument to the first call to the
postMsg() helper method.

Differential Revision: https://phabricator.services.mozilla.com/D43307

UltraBlame original commit: d1615ca96d8f9599d78682439af3b86c011387ea