Bug 1578665 - Create a dedicated pref for Browser Console editor state. r=Honza.
We want the Browser Console and WebConsole to have
distinct preferences, so you might have one in
editor mode, and the other one in inline mode if
you want.
This is as simple as adding the preference, and
stripping the devtools.webconsole. prefix in
the constant we use for retrieving the preference
value.

Differential Revision: https://phabricator.services.mozilla.com/D44631

UltraBlame original commit: 3a06725cd4a2446b0df12dd2fb66635d38e3879a