Bug 1578986 - Refactor tile comparisons to enable dirty rect checking. r=nical
Previously, primitive sub-dependencies (such as transforms, opacity
bindings etc) were invalidated by comparing the dependency array
only. This means that it's not possible to correlate an area of
a tile that is affected by those sub-dependencies.

Now, comparisons on sub-dependencies are handled as part of the
dependency check for each primitive. This means we have the
affected rectangle of the tile available, which will allow
dirty regions within a tile to be built correctly.

This patch is only preparation work - the actual dirty rect
calculation will be done as a follow up patch.

Differential Revision: https://phabricator.services.mozilla.com/D44771

UltraBlame original commit: 7a46b21394232c1e7a7c1466ccf3c270ca812960