Bug 1577660 - Rename `HTMLOrSVGOrXULElementMixin` to `HTMLOrForeignElement`. r=bzbarsky
This commit renames `HTMLOrSVGOrXULElementMixin` to `HTMLOrForeignElement` and
actually makes it a mixin. dataset, tabIndex and blur() are moved from
HTMLElement, SVGElement and XULElement into this new mixin. As a consequence,
a few extended attributes for these interfaces are modified and the dataset
attribute is added to the XULElement one. However, none of these changes are
web-observable.

Differential Revision: https://phabricator.services.mozilla.com/D44879

UltraBlame original commit: 85efdc31cf556ed9cda6a4d831177b11779b53ca