Bug 1579862 - Introducing HuffmanTableImplementationSaturated;r=arai
We introduce a new implementation of Huffman Tables that trades
space for a fast lookup. For the time being, this implementation
is reserved for tables with a max bitlength of 8 or less.

Depends on D45643

Differential Revision: https://phabricator.services.mozilla.com/D45644

UltraBlame original commit: 9e0541d8cea7b62c22b3e8a5a3847e7d4162867a