Bug 1580711 - Refactor storage of primitive instances into primitive clusters. r=nical
Previously, primitive lists were stored as:

  PicturePrimitive
    PrimitiveList
      [PrimitiveInstance]
      [PrimitiveCluster]
      [PictureIndices]

Each primitive instance contained a spatial node index and an
index into the primitive cluster it belongs to.

Now, the instances in a primitive list are stored as:

  PicturePrimitive
    PrimitiveList
      [PrimitiveCluster]
        [PrimitiveInstance]

This provides a number of advantages:
 * Size of the PrimitiveInstance struct is smaller.
 * No need to maintain a separate PictureIndices list.
 * Easy and fast to skip the array, finding pictures or scroll root changes.
 * Much faster to split and reorder PrimitiveList structures.

This patch is refactoring only, it doesn't contain any functional
changes. As we enable multiple picture caching slices, we need to
be able to split and reorder PrimitiveLists. Storing the primitive
instances in this way makes that process much more efficient than
it currently is.

Differential Revision: https://phabricator.services.mozilla.com/D45636

UltraBlame original commit: da7b63b2748082dde8b6d45f3c68ee05deec2afc