Bug 1570435. Handle merging blobs with different bounds. r=nical
We compute the preserved bounds for each item and see if it's
contained in the invalid area. The gecko side will do the same
computation.

Appologies for duplicating euclid with Box2d type. We can deal with
that in a followup.

The asserts are disabled for now. I tried reenabling them and they
triggered. I haven't had a chance to figure out if they are still
valid.

Differential Revision: https://phabricator.services.mozilla.com/D45283

UltraBlame original commit: ad1243b93a1a765fbb6ee99ac791c7b28a13d4fd