Bug 1568857 - Get flexbox information from the right target; r=ochameau
The flexbox inspector is different from the grid inspector in that it
does not list all of the flexbox layouts on the page. In D45524 we needed
to list all of the LayoutFronts to retrieve all of the GridFronts in the
page. However here, we do not need to do this. We only need to get the
LayoutFront contextual to the NodeFront currently selected.

I also took this opportunity to remove an old backward-compatibility
check called hasGetCurrentFlexbox which was no longer required since it
guarded a feature added in Fx60.

Differential Revision: https://phabricator.services.mozilla.com/D46299

UltraBlame original commit: 4809391c9f301becf291e202448e07a273685edd