Bug 1405031 - Add middle-click and click modifiers to browserAction r=robwu and pageActions.
Before this change, browserActions and pageActions did not trigger
onClick events when middle-clicked, and no information on the button or
any modifiers were passed in the onClick event. With this change, middle
clicking triggers an event, and a clickData object is passed in the
onClick event, with the button and a list of modifiers.

Differential Revision: https://phabricator.services.mozilla.com/D41492

UltraBlame original commit: 6c464eede88af150a8c9f1e1aecfa9388831817b