Bug 1582150 - Don't rely on page stop in aboutConfig test. r=snorp
Trying to access `about:config` will fail even before the page has started loading so we don't get the usual flow.

For a normal page (that fails to load) it is:

- OnLoadRequest
- OnPageStart
- OnLoadError
- OnPageStop

but for about:config it's just

- OnLoadRequest
- OnLoadError

So we just wait for the OnLoadError message instead.

Differential Revision: https://phabricator.services.mozilla.com/D46369

UltraBlame original commit: 85248a084397f1823cb4da7f0b10eb351973674e