Bug 1578125 - Fix flakiness in common animation interpolation testing functions; r=hiro
This patch fixes a number of issues with the common interpolation testing
functions upstreamed from Blink.

Firstly this test file sets an animation duration of 2,000,000,000s and a delay
of -1,000,000,000s but that does not appear to be necessary since no time
elapses between when the animation is established and when the result is
checked (measure() is called immediately after interpolate() without any
interleaving call to requestAnimationFrame, for example).

Furthermore, this long time will cause some configurations of Firefox on
Windows to treat the time as an infinite time and end up producing discrete
animation. There is nothing in CSS values that requires supporting times of
this magnitude (~63 years) so it seems best not to make the tests rely on it
being supported (the tests are intended to cover interpolation after all, not
timing range) and instead to use a shorter time.

Similarly, for the Web Animations interpolation tests, this utility function
uses a duration of 1ms and seeks to 0.5ms.  This can likewise lead to precision
issues, particularly when testing interpolation pairs that fall back to
discrete animation since in that case, the test will seek to precisely 0.5ms
and test that the result is at the top of the step that occurs at that time.
Floating-point error can cause this to fail so this patch changes the duration
and seek time to 100s and 50s respectively.

This patch also includes a few additional tweaks to this file including:

* Replacing the "steps(2, end)" timing function with "linear" since this
  avoids floating-point errors at the 0.5 mark.
* Adding translation for "float" to "cssFloat".
* Triggering transitions by querying the property being transitioned.
  (In theory a UA could skip flushing all styles when accessing the 'left'
  property but in practice that is probably not Web-compatible.)

Differential Revision: https://phabricator.services.mozilla.com/D46698

UltraBlame original commit: 0626a86d78b12627264c0d279892e9ff0f98f42a