Bug 1578408 - Attach content process thread actor from the toolbox. r=jlast,jdescottes,yulia
We need to attach the thread actor as soon as the toolbox start.
To:
* know when the thread is paused, even if the debugger isn't opened
* report paused state by changing the debugger icon
* switch to the debugger on pause
This will also fix some other code, like in the console, which
expect the Thread actor to be attached and already inspecting
the debugged target's globals.

Differential Revision: https://phabricator.services.mozilla.com/D44638

UltraBlame original commit: b4875ea160da606061e9767f0e121d61f60678c3