Bug 1578615 - part7 : rename tests 'browser_block_*' to 'browser_delay_autoplay_*'. r=chunmin
As the things we do in those tests are not really "block" autoplay, which is not to allow autoplay starts, what we do is actually delaying them until tab goes to foreground or user clicking `play tab` icon on the tab.

In order to distinguish them from the real blocking-autoplay, and reflect what we really do in those tests, rename all of them with prefix "browser_delay_autoplay_*".

Differential Revision: https://phabricator.services.mozilla.com/D44921

UltraBlame original commit: 263512ebd456125cdab2dab1a90bdaba65ee0137