Bug 1582189 - Include submodules in git fetch tasks; r=froydnj
Using git-archive for the fetch task means that we don't get the
submodules of a git repository included in the archive. There isn't a
straightforward way to get submodules from a bare repo included with
git-archive, so instead we can simply clone & checkout with
--recurse-submodules and then use a standard tar command to bundle up
the tree.

Adding --recurse-submodules to the commands has no effect on a repo
without submodules, so we can add it to all invocations for simplicity.

Differential Revision: https://phabricator.services.mozilla.com/D46827

UltraBlame original commit: 3bd01ec4a9e967abd1473b14c6a970eedadc4b42