Bug 1583444 - Stop using win.top in DevTools tests r=nchevobbe
Depends on D46959

For those tests, using win.top was the same as using win, and did not have any impact on the asserts performed in the test.
In particular, the AutocompletePopup constructor calls for a "toolboxDoc", so passing the host document is actually more correct.

Differential Revision: https://phabricator.services.mozilla.com/D46960

UltraBlame original commit: c62a28520dd66c0697afbf8d809db55778ea66f8