Bug 1580922. Partially revert blob recoord fallback changes. r=nical
This moves the origin of fallback blobs back to the top left of their display
item bounds. This is what they were before the patches in bug 1568227 and makes
more sense because there's not necessarily a reference frame above the fallback
frame which means that the coordinates of the display item can change without
us wanting to invalidate the interior.

Differential Revision: https://phabricator.services.mozilla.com/D47275

UltraBlame original commit: 88a32122785ba7f6269a82442a0a46f971313a0e