Bug 1567201 - Support multiple queries in a single packet; r=ng
If there are multiple pending queries, this allows up to five of them to be
sent in a single packet. This also adds a check to ensure we don't already have
a pending query for a hostname prior to sending a new query.

Differential Revision: https://phabricator.services.mozilla.com/D46983

UltraBlame original commit: c45a8001b0a113944d04e6e2b15bb5c759530de1