Bug 1584009 part 1.  Loosen up the check for dictionary-containing-union-containing-dictionary.  r=peterv
Since unions can now end up a in binding header, it's only a problem when the
two dictionaries are in one header and the union is in a different one.  If all
three are in the same header, for example, there is no issue.

Differential Revision: https://phabricator.services.mozilla.com/D47195

UltraBlame original commit: 86d33a025dd4475bddd9236e98ba24f5bc98c6a8