Bug 1585898 - Fix broken tab bar with picture caching on mac. r=mstange
Promote clear primitives to be picture cache slices that can
be drawn during the composite step. Without this, the clear
primitive is not correct since it only operates on the slice
it is assigned to, not the entire background before it.

Differential Revision: https://phabricator.services.mozilla.com/D48139

UltraBlame original commit: 7a6d4a2a715a265c55b0493b6f3166749753df51