Bug 1512175 - Rename and adapt test_mr_unsupported_src.html. r=jib
With spec compliant mime type handling we no longer fire "error" on start for a
audio-only recording of a (currently) video-only MediaStream.

This patch adapts the test to more accurately name what we originally wanted to
test - that pause() after an error doesn't crash but throws. It does this by
triggering another kind of error, one that happens because we try to remove
the recorded track from the recorded stream. We still keep the behavior of
start()ing the recording before any supported tracks are available, because we
support that.

To clarify what this test is actually testing, this patch also modernizes the
code with async/await to linearize the testing of events and their order.

Differential Revision: https://phabricator.services.mozilla.com/D17811

UltraBlame original commit: dca395faee362d831f36cda88562d346d441fb3a