Bug 1549176 - Part 2: Update test262 tests. r=jwalden
The previous changeset made a change to test262-update.py. This changeset is
the result of running the script, like this:

    cd js/src/tests
    python test262-update.py --revision 59a1a016b7cf5cf43f66b274c7d1db4ec6066935

The revision is taken from the first line of js/src/tests/test262/GIT-INFO; the
purpose of specifying that revision is to run the script again using the same
input, so that the output reflects only changes to the script itself, not new
tests.

Depends on D46991

Differential Revision: https://phabricator.services.mozilla.com/D46992

UltraBlame original commit: b4b538b20b4178c4bc5b3760b7a164b90430d1f7