Bug 1552766: Handle optimised out `this` when reporting missing bindings. r=jimb
In addition to normal bindings, which always use a proper identifier name, the
|this| binding can also be optimised out for arrow functions. The |this|
binding is stored in a binding named ".this", which, because it doesn't match
the IdentifierName syntax, triggered an assertion when generating its printable
representation. Workaround this assertion by special-casing ".this" when
reporting optimised out bindings.

Differential Revision: https://phabricator.services.mozilla.com/D47238

UltraBlame original commit: 0fc012d870ea47c1880eb96635f5124c537872f7