Bug 1571446 part 3 - Combine JSScript's jitScript_ and warmUpCount_ fields in a single warmUpData_ field. r=tcampbell
The warm-up count is stored in ScriptWarmUpData until the script is warm
enough for the Baseline Interpreter and the JitScript is created. At that point
we use the warm-up count stored in JitScript.

ScriptWarmUpData uses pointer tagging. This should make it easy to add new
types for LazyScript data in the future.

Differential Revision: https://phabricator.services.mozilla.com/D42319

UltraBlame original commit: e753d23c1237313b3cc94ee509cbba9a5b6246d8