Bug 1585120 - Add tier 3 Android browsertime smoketests. r=perftest-reviewers,sparky
The intention is to run `tp6m-1` and `tp6m-cold-1` using Browsertime
against the GVE vehicle on integration and try, so as to smoketest the
Browsertime harness as we stabilize it.  At the same time, this also
allows try runs against the Fenix and refbrow vehicles.

Differential Revision: https://phabricator.services.mozilla.com/D48772

UltraBlame original commit: 64b1136cba371984c15ca507e7127a0ad112bae0