Bug 1580830 - Add a display to the private browsing page to inform of the new setting for search engines in private mode. r=mikedeboer,fluent-reviewers,flod
This implements the introduction display for the new search engine preference.

The decision for whether or not to display the introduction is handled in AboutPrivateBrowsingHandler.jsm, as we need
to limit it to being displayed once per session.

Differential Revision: https://phabricator.services.mozilla.com/D48722

UltraBlame original commit: b95dac78d333fc3f5c94f246f7145d779f66ceff