Bug 1587741 - Captive Portal Service should use a new detector when we start again r=mayhemer
If the captive portal service is disabled then enabled again when in a captive
portal, we sometimes don't send the captive-portal-login observer notification
because we're reusing the same mCaptiveDetector object, which believes it's
already sent it.

We should use do_CreateInstance instead of do_GetService for this.

Differential Revision: https://phabricator.services.mozilla.com/D48819

UltraBlame original commit: b085b41c10ce17ce0315234a48292e14513bca6f