Bug 584141 Added WBR frame class implementing nsFrame and overriding thepeekOffset* methods. r=dbaron,Ehsan
The wbr element represents a line break opportunity. Before the change double clicking on a text that contains a wbr element did not select the whole sentence.
After the change, the presence of a wbr element in the selection does not affect the selection functionality.

Differential Revision: https://phabricator.services.mozilla.com/D47558

UltraBlame original commit: 80ffb5ef8efb71442034e56a093a4367ac06cdf6