Bug 1583654 - Use the subgridToParentMap to check that the node was previously a subgrid. r=pbro
The previous condition didn't fully check that the grid node was previously a subgrid.
So, we run into a scenario where we refresh the page and a "display-change" event is hit
after a new root is loaded, and the grid highlighter is restored and hidden because
the check will pass as long as the node is a grid container.

Differential Revision: https://phabricator.services.mozilla.com/D49659

UltraBlame original commit: abfee71f01b0122064a93c0f3c8cf56b887573fd