Bug 1588737 - Hide -moz-touch-enabled media query in Nightly and Early Beta. r=xidorn
This is effectively superseded by the hover / any-hover media queries, which
actually are standard, and is also causing trouble in the wild.

Not even the browser fronted uses it, so we should be able to just remove it
everywhere at once.

Differential Revision: https://phabricator.services.mozilla.com/D49506

UltraBlame original commit: 673c2d3025c58a5307d3ecf2da7d500dbe2b16b2