Bug 1585974 - Increase gecko_profile_entries to 10 million in talos DAMP configuration r=rwood
The current limit of 2 million entries only allows
to record part of a DAMP run. The profiler team will investigate
if the number of entries can/should be reduced, but for the time
being we bump the limit to 10 million in order to record full
profiles when running DAMP on try.

Differential Revision: https://phabricator.services.mozilla.com/D49555

UltraBlame original commit: 3b1f06fb2cb28d6b392889a1be3e6f7373275ee0