bug 1589824 - pass a typed array to OS.File.writeAtomic in certificate export r=Gijs
OS.File.writeAtomic expects either a utf-8 string or a typed array. This patch
fixes instances in pippki.js in certificate export where this was not
guaranteed to be the case. It also extends the test for this functionality to
cover more cases.

Differential Revision: https://phabricator.services.mozilla.com/D50117

UltraBlame original commit: 94f423368ebdbab66f8622ed32a02dcf88b66168