Bug 1586381 [wpt PR 19525] - [ElementTiming] Report text paints of size 0, a=testonly
Automatic update from web-platform-tests
[ElementTiming] Report text paints of size 0

Currently, TextPaintTimingDetector uses |invisible_objects_| to keep
track of text paints which occupy a size of 0. It is an optimization:
LCP does not care about these paints, so it does not need to queue
them to obtain paint times nor store a full TextRecord for them. However
ElementTiming wants to expose all of the text paints for elements that
are annotated with the elementtiming attribute. It is confusing for an
element that is painted when outside the viewport to never be reported.

To solve this problem, there are two alternatives. One is to use a
single |painted_objects_| map which would contain both the visible and
the invisible objects, and hence no extra logic would be needed for
ElementTimng to receive these objects (but some care might be needed so
LCP does not report objects of size 0). This is simpler to code but is
less efficient as it adds extra work and memory for the invisible
objects.

The second alternative is to use a new data structure for text nodes
that are of size 0 but need to be reported to ElementTiming. Paint times
for these are assigned in AssignPaintTimeToQueuedRecords() and then the
TextRecords are deleted (the fact that the element has been painted is
still being handled by |invisible_objects_| so there is no need to keep
the records for longer). This is the alternative implemented in this CL.

Bug: 1011009
Change-Id: I2d15393fc61134b08a5c15bd81d062d42dfb29e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1838260
Reviewed-by: Steve Kobes <skobeschromium.org>
Commit-Queue: Nicolás Peña Moreno <npmchromium.org>
Cr-Commit-Position: refs/heads/master{#704002}

--

wpt-commits: 9931f3bcb6077db1dc7e90db4daf333c9350c73b
wpt-pr: 19525

UltraBlame original commit: 5b28db890a4d02e54e0330b6f69f4d1c9da0d485