Bug 1588038 [wpt PR 19637] - Propagate writing-mode and direction from body to html., a=testonly
Automatic update from web-platform-tests
Propagate writing-mode and direction from body to html.

Direction and writing-mode is propagated from body to the viewport, but
when the computed value of those properties are different between the
root and body elements, their used values are the computed values of the
body element. The computed values are not affected by this.

We implement this propagation by using a ComputedStyle on the
HTMLHtmlElement's LayoutObject which has the propagated values for
direction and writing-mode.

The modified tests relied on direction on #container being different
from the #container parent. That is no longer possible with body being
container and the used direction value of body propagated to html.
The tests were modified to use a child of body as container instead,
the removed 8px margin was just replicating the body margin and we do
not want that on the #container div.

Spec: https://drafts.csswg.org/css-writing-modes-3/#principal-flow

TEST=external/wpt/css/css-writing-modes/wm-propagation-002.html
TEST=external/wpt/css/css-writing-modes/wm-propagation-body-033.html
TEST=external/wpt/css/css-writing-modes/wm-propagation-body-037.html

Bug: 988585
Change-Id: I2879c8fdc082940d2baaf67e126b28915f9f1eca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1852786
Commit-Queue: Rune Lillesveen <futharkchromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrickchromium.org>
Cr-Commit-Position: refs/heads/master{#705218}

--

wpt-commits: c9ddd877ad6714aabd1509438a8648ed62e38035
wpt-pr: 19637

UltraBlame original commit: 25089fc0275d7fa86e70e3b1a2042b3e14949c4c