Bug 1591692 - Ensure the iframe gets loaded before we proceed the test in 'browser_firstPartyIsolation_aboutPages.js'. r=baku
The patch adds 'wantLoad' to the BrowserTestUtils.browserLoaded() to
ensure the iframe gets loaded before we proceed the test.

In addition, we need to disable this test in Fission since the
'wantLoad' feature is broken for iframes in Fission.

Differential Revision: https://phabricator.services.mozilla.com/D50846

UltraBlame original commit: f294a5d018093dd43fe47933ceea83d9433bd72f