Bug 1592808: Annotate ContentDelegate overrides for onFullScreen with AssertCalled; r=geckoview-reviewers,agi
The changes in bug 1583575 cause an additonal event to call the `ContentDelegate`,
so the test proceeds without waiting specifically for `onFullScreen` to be called.

This patch annotates the `onFullScreen` overrides to ensure that we're waiting
on the correct event for these test cases.

Differential Revision: https://phabricator.services.mozilla.com/D51195

UltraBlame original commit: a98de0432770ae5b469a5f8b5ae84027459588c5