Bug 1587385 - Give a better info (not error) message if a webfont is not used because of font-display timeout. r=heycam
This is not an error in the same way as a download failure, and should not be reported as one.
An Info message is sufficient.

Also suppress "unknown" location in messages about font-face rules, as it is not useful,
pending a proper fix (bug 1450903).

Differential Revision: https://phabricator.services.mozilla.com/D50346

UltraBlame original commit: ca10c746267a543aaa63567024c610d5b39e542d