Bug 1545123 - move plugin finding into its own class to clarify dependencies and data flows, r=handyman
This moves plugin finding logic into a separate class (PluginFinder).

PluginFinder is a runnable. After this commit, there are two ways in which it
can be used:
1. to actually find and load plugins.
2. to check if there have been any changes to plugins.

Although it is a runnable, at this point we still invoke its Run method on the
main thread, so all that's happening is we're separating the "look for plugins
on disk" bits out from everything else.

The goal is to be able to run the IO-intensive FindPlugins (including reading
and writing pluginreg) away from the main thread.

Depends on D48330

Differential Revision: https://phabricator.services.mozilla.com/D48331

UltraBlame original commit: d91549e68229531659bf744f34feca3bc02f5ea5