Bug 1593354 - Add BaseScript::IsFunction flag. r=jandem
Set this flag for scripts that are for functions. This is equivalent to
checking the kind of the bodyScope() but is safe to call in GC contexts such
as finalization.
Also use an isFunction() accessor instead of checking if JSScript::function()
is nullptr throughout the codebase. Similarly, cleanup similar uses of
JSScript::module().

Differential Revision: https://phabricator.services.mozilla.com/D51472

UltraBlame original commit: 821b104d357b899ba03415a82066e96db6aa269b