Bug 1590864 - part1 : get the user activation state from top-level browsing context. r=edgar
As `ApproverDocOf()` uses `docshell::GetInProcessSameTypeRootTreeItem()`, which is not compatible in Fission because we are not able to get the top document if we are in the remote process. Therefore, we should check browsering content directly instead, because the user activation state would be sync automatically for each top-level browsing context among different processes.

Differential Revision: https://phabricator.services.mozilla.com/D51427

UltraBlame original commit: f66335a878bac85230b456372eddc4fd191c1473