Bug 1590864 - part3 : use 'notifyUserGestureActivation()' to activate document directly. r=edgar
I found that this test sometime would fail on the try server, and the reason is that the child document doesn't be activated correctly even if we have called `synthesizeMouseAtCenter()`.

In order to reduce the dependency of other modules, which might affect synthesizing event, we should use `notifyUserGestureActivation` to ensure that we indeed activate the document.

Differential Revision: https://phabricator.services.mozilla.com/D51685

UltraBlame original commit: b860f82259d051e790ab53329a10f9408dbad807