Bug 1484780 - Cache gfx::Path to avoid building and flattening path at each restyle cycle. r=heycam
In most cases, we run an animation on an object by changing its
offset-distance/offset-rotate, but keep its offset-path the same.
Building and flattening the path is sometime expensive, especially for
large path, so caching it makes sense to us and have a significant
performance improvement. This is for the main thread motion path
animations.

Note: Even though we support compositor animations for motion path,
nsIFrame::GetTransformMatrix() is still called during the animations for
other usages, so we may still build the gfx::Path on the main thread
without this patch, so this improvement becomes necessary for most cases.

Differential Revision: https://phabricator.services.mozilla.com/D46667

UltraBlame original commit: c0ddb98d498d20f29acf3957e00ff3dc6691b045