Bug 1582779 - Offer to restart the browser when JS Tracer is enabled; r=julienw
To test this feature, check the "JSTracer" feature, and the user should be
prompted with a message bar to restart the browser. This will then set the
proper environment variable. Unfortunately, currently the user can still do
this even if a feature is not available in the build. See Bug 1585659.

Differential Revision: https://phabricator.services.mozilla.com/D50085

UltraBlame original commit: 76cd5919691eeab197d66c9e70944003a6c0c207