Bug 1595965 - Fix tile draw order with native compositor mode. r=sotaro
In Draw compositor mode, tiles are placed into a bucket based on
the tile kind and/or opacity. This simplifies the renderer code
for this mode, which uses the z-buffer to reject opaque tiles.

In native compositor mode, the tiles need to be submitted in the
correct draw order. Opacity is a property of the native surface,
and should not affect the order the tiles are added to the visual
tree of the native compositor.

Differential Revision: https://phabricator.services.mozilla.com/D52933

UltraBlame original commit: e7837d7a78c67fb122f6f4685eb0e2f06841b621