Bug 1595885 - Ensure we have a valid window after wpt reftest timeout, r=webdriver-reviewers,maja_zf
Typically we can close the wpt reftest window at the end of a test and
open a new one for the follwing test. But that interacts badly with
the harness trying to run marionette commands before the test starts;
for example when it tries to set a pref. So we were hitting a
situation where tests that TIMEOUT followed by tests that require a
pref end up erroring. To avoid this make sure to set the original
opener window as the marionette window when we close the reftest window.

Differential Revision: https://phabricator.services.mozilla.com/D53014

UltraBlame original commit: cc9330b5decd89fe8bc422957398d61d89041bb1