Bug 1588241 - P2. Add ability to use lambdas for channel event handlers. r=mayhemer
Remove the need to create separate runnable classes. It's far more readable and remove the need to duplicate lots of code.

We unfortunately need to capture "this" in a ref counter to get around the static analyzer complaining about capturing this by value, even thouch the ChannelEventQueue guarantees that this will outlive the event.

Differential Revision: https://phabricator.services.mozilla.com/D52261

UltraBlame original commit: a00f164555ff94b19f25485908288dcadc563624