Bug 1596052. If the SendUpdateCanvasForEvent call is for a previous test then ignore it. r=mattwoodrow
This is basically what we did in bug 1593171 (Protect against the same test from calling RecordResult more than once in the reftest harness) where we early exit in SendInitCanvasWithSnapshot.

But now we do it in SendUpdateCanvasForEvent too because SendUpdateCanvasForEvent calls SynchronizeForSnapshot which calls setupAsyncScrollOffsets and setupAsyncZoom, both of which get the documentElement of the current doc and operate on it. The problem is that this could be SendUpdateCanvasForEvent call from the previous test operating on the dom of the current test.

I haven't actually observed this, just noticed it while implementing checking of contentRootElement to make sure all cases are covered.

Differential Revision: https://phabricator.services.mozilla.com/D52827

UltraBlame original commit: b9737a17de366f287072fec20ae5dcbc14bfad50