Bug 1593170. Check if contentRootElement is a dead wrapper before using it. r=mattwoodrow,dbaron
With the fission changes everything is more async, meaning some tests can run longer. Some crashtests navigate to a new location meaning the original root element we have a variable for is no longer around, and chrome isn't allowed to keep content nodes alive, so we are left holding a dead wrapper that throws if we try to access anything on it.

So we check if contentRootElement has become a dead wrapper (and null it out) any time we try to access it and it could have become dead. Generally the existing code already handled a null contentRootElement.

Differential Revision: https://phabricator.services.mozilla.com/D52829

UltraBlame original commit: 0dc7546c4a8a75dda043edc312305a6ceba0ad53