Bug 1581572 - Allow AnyRef on JS/wasm fast paths, part 2. r=rhunt,bbouvier
An AnyRef is an opaque pointer type to wasm but can hold any JS Value;
the Value is boxed when it enters wasm and unboxed when it enters JS.
This boxing and unboxing must also happen when wasm and JS interact
along the fast ("Jit") path, not just through the slow ("Interpreter")
path.

This patch allows an imported function (an exit) that returns anyref
to have a fast path; the Value that we receive from the import is
boxed into the wasm representation, which means allocating an object
for it if the wasm representation can't accomodate it directly
(currently only object or null).

It also allows an exported function (an entry) that accept an anyref
parameter to have a fast path; the value is similarly boxed in the
stub layer.

This patch does not handle argument passing for the inlined JS->wasm
path; that is handled by a separate patch.

Differential Revision: https://phabricator.services.mozilla.com/D49728

UltraBlame original commit: b1796f474af442447cfdfd8676d93d540293cecb