Bug 1507674 - Refactor svg observer setup to not do silly things when observing an ancestor. r=longsonr
Right now we post updates and it "works" because we prevent the UpdateOverflow
call if we're during reflow.

If this happens during styling however this is not sound (and it is not sound
in general and has caused badness in the past, as noted by the other
workarounds).

Make it sound by preventing to observe ancestors, and do it everywhere, removing
various ad-hoc hacks that were spread around elsewhere.

This changes expectations of two tests:

 * clip-path-recursion-002.svg: Now we consider the inner clip-path reference
   invalid. This matches WebKit and Blink, and I don't see any spec text
   explicitly asking for our old behavior, so I just changed the test.

 * element-paint-recursion.html: Changes the expectations of elements
   referencing themselves via -moz-element(). Now it is invalid, instead of
   painting ourselves once inside ourselves, which was a bit wild on its own.

Differential Revision: https://phabricator.services.mozilla.com/D53890

UltraBlame original commit: 0fd8713b9050483915cfc784b124537cb5ba5c75