Bug 1586986 - Introduce 'fixed margins' on the main-thread to omit the gap between 'fixed margins' on the compositor and the last dynamic toolbar position on the main-thread. r=botond
The gap is caused by the difference between 'fixed margin' on the compositor
based on the state where the dynamic toolbar is completely visible and
position:fixed elements positioned by the time where is partially visible or
completely hidden.  That's because the 'fixed margin' is computed based on the
bottom of the ICB whereas position:fixed elements are slightly shifted from the
ICB edge during the toolbar transition.

Differential Revision: https://phabricator.services.mozilla.com/D52337

UltraBlame original commit: 0ff21dbde947ed6de9e220279093a18ca524814f