Bug 1597036 - Add fission.autostart as an optional locked pref in the troubleshooting schema. r=mconley
I think that Bug 1572925 made it an error to have any locked prefs
reported in about:support, because they had a pref that accidentally
was getting reported, and I guess we don't have any by
default. fission.autostart is locked on beta and release, but not on
Nightly. This patch makes it so it is okay, but not necessary, to have
fission.autostart reported as a locked preference, so that the test
will pass in either configuration.

Differential Revision: https://phabricator.services.mozilla.com/D53888

UltraBlame original commit: 5d3bdae26c6463df3f3c360564d949581ef85743