Bug 1326182 - Send all the console.table items in the console message. r=Honza.
This allows us to not have to reach for the server
again from the client to get all the data we need
to render the table.

We still need to keep the old function to handle
connections to older servers. We'll use Bug 1597955
in Firefox 74 to clean the bits we don't need anymore.

A test is added to ensure the bug is fixed and we don't
regress; stubs are updated.

Differential Revision: https://phabricator.services.mozilla.com/D53962

UltraBlame original commit: e06f2a4de96908b64f414b57bb1c787f3428b2d1