Bug 1597943 part 4 - Stop using SRC_TABLESWITCH source note in code coverage. r=nbp
This will allow us to remove the source note.

Instead of handling default-clauses different from the switch cases, this
changes the loops iterating over the switch cases to handle the "default"
as well.

The number of hits for the default-target is computed from subtracting the hits
for the cases from the number of hits for the switch-statement

Differential Revision: https://phabricator.services.mozilla.com/D54274

UltraBlame original commit: 2f229ecb29b1df01d0fcc20a349598cc0ecdb8be