Bug 1598635 - nsHttpChannel::ComputeCOOPMismatch should not crash in xpcshell-tests r=ttung
Not having a responseHead or browsingContext should not cause a crash or exception

In xpcshell-tests we don't really have a browsingContext, so we want to make
sure we don't crash when that happens.

If the channel failed, we may not have a response head. In that case we don't
want the call to ComputeCrossOriginOpenerPolicyMismatch to be what throws
an exception.

Differential Revision: https://phabricator.services.mozilla.com/D54408

UltraBlame original commit: bea0b34db7da607092e33adcc37ff5928d4f061e