Bug 1598992 - More precise way to decide when to delete expired Base Profiler data - r=gregtatum
In practice the previous test that was deleting Base Profiler data when the
index became greater that 1 was correct for Firefox, because the Base Profiler
*always* starts before the very first Gecko Profiler active instance.

However in tests (like the one in the following patch) this may not be true,
because each test may start and stop the profiler, and the recent storage update
means that the index doesn't go back to 1. So when a test (apart from the first
test to use the profiler) attemps to use the Base Profiler, that Base Profiler
data will be immediately discarded because the index is already greater than 1
(from previous tests).

This change is more future-proof as well, in case we later want to use the Base
Profiler more than once in a Firefox instance.

Differential Revision: https://phabricator.services.mozilla.com/D54445

UltraBlame original commit: 53bc356053b047a01d77b1d63985fd562f54f936