Bug 1599006 - P4. Properly carry history and frecency data when switching process. r=kmag,mayhemer.
This allows test toolkit/components/places/tests/browser/browser_multi_redirect_frecency.js and others to pass when fission is enabled.

The content process expects to know the chain of redirects encountered while opening a URI. The DocumentChannelParent gather that information and sends it to the new ContentChild which can then propagate the information to the new nsDocShell.

The data used to only be passed around during same-origin redirects when fission mode was enabled.

In order to allow for move semantics and preventing unnecessary copy of the DocumentChannelRedirect array, we make the nsIChildProcessChannelListener::onChannelReady property C++ only (noscript).
As we have only one concrete nsIChildProcessChannelListener class (ChildProcessListener) and that the unique OnChannelReady implementation is infallible; we really don't need for the method to return nsresult (so we declare it nostdcall). This helps simplify that call.

Differential Revision: https://phabricator.services.mozilla.com/D54909

UltraBlame original commit: b87b36262b0116092ebd375ca1acf01603563aab