Bug 1592991 [wpt PR 20018] - Add Web Platform Tests for the Split Cache, a=testonly
Automatic update from web-platform-tests
Add Web Platform Tests for the Split Cache (#20018)

We add a web platform test for the cache when it is partitioned by top-
frame origin. The test caches a resource, then attempts to load it in
a popup window. With the split cache, if the popup window is
cross-origin, we would expect that request to be a cache miss. We
compare the number of requests that hit the server to our expectations.
To enable these tests, the existing framework is slightly refactored,
including to respond to CORS preflight requests.

The virtual test suites virtual/split-http-cache and
virtual/not-split-http-cache explicitly test the cases of a split and
not-split cache, respectively. The default test is disabled as its
result depends on whether the split cache feature is enabled.

Bug: 981970
Change-Id: I049fdaec4aebcbdeebdd3c77d13d6ca420d1bd62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1890952
Reviewed-by: Yutaka Hirano <yhiranochromium.org>
Reviewed-by: Shivani Sharma <shivanishachromium.org>
Commit-Queue: Alex Turner <alexmtchromium.org>
Cr-Commit-Position: refs/heads/master{#714236}
--

wpt-commits: a1be16d3d955a0dc8f71e259c54b2afd54725355
wpt-pr: 20018

UltraBlame original commit: 85b0b1f0e7d819f9672392bbe932bfccc489f12f