Bug 1600195 - rip out unused 'nice' handling for child processes on Linux/Android, r=mconley
This was added for XUL fennec, which was dropped at the time of Firefox 7.
But we're still reading this pref and env var and system info data, so let's
stop doing that (also because 'cpucount' is unlikely to be 1 on modern
desktop machines, we're dropping arbitrary user pref sets for geckoview,
and I don't think it's even possible to set an env var for app execution
on android as a user - certainly not easy).

Depends on D55265

Differential Revision: https://phabricator.services.mozilla.com/D55266

UltraBlame original commit: b45f47bb9911f54c3ffce69ed90d5f6a65dae61d