Bug 1600439 - Replace LazyScriptData with PrivateScriptData. r=jandem
It it straight-foward for users of LazyScriptData to support a single array
for closedOverBindings and innerFunctions. As a result, we can use
PrivateScriptData as the implementation and eliminate the LazyScriptData type
altogether.

Depends on D55035

Differential Revision: https://phabricator.services.mozilla.com/D55360

UltraBlame original commit: 7c27b4f099e875e6aff5e9f2f6f9876ace0ea027