bug 1600121: netwerk: error on empty path in nsIHttpServer.registerPathHandler; r=mayhemer
It should be illegal to add paths that cannot be handled/accessed
or later referenced.  Without a path, it is for example later
impossible to delete the handler.

To address this we return an NS_ERROR_INVALID_ARG when
nsIHttpServer.registerPathHandler is called with an empty string.

Differential Revision: https://phabricator.services.mozilla.com/D55160

UltraBlame original commit: 40b6d9ae060135379c2e9379fbc038f1ed99f548