Bug 1595557 [wpt PR 20197] - Initial implementation of the TestRendered event for reftests, a=testonly
Automatic update from web-platform-tests
Implement TestRendered event for reftests (#20197)

This is described in RFC #34. The TestRendered event is emitted for
reftests with a `reftest-wait` class on the root at the point at which
the screenshot would occurs if that class was not present. It provides
a hook for test authors to dynamically modify the DOM after the
initial layout and paint are complete.

--

wpt-commits: 7b20b28674f71176bf1ad44157eb7287ffcb2c49
wpt-pr: 20197

UltraBlame original commit: 3749aa97c29a618a9a2739490babec8f8deef698