Bug 1601074: Fix preliminary-objects-set issue wrt NEWOBJECT_WITHGROUP. r=iain
The recent addition of the JSOP_NEWOBJECT_WITHGROUP opcode for bug
1598347 (itself a regression fix for 1580246) has led to an issue when
more than a certain number of array elements with the same group are
created within an array literal. In particular, when too many objects
are created, the preliminary-objects-set for the ObjectGroup becomes
full and hits a MOZ_CRASH. This patch avoids trying to add to the
preliminary object set in the _WITHGROUP case.

Differential Revision: https://phabricator.services.mozilla.com/D55736

UltraBlame original commit: e4c1a6acc2b152f69aefa80f46b8c14d999cc385