Bug 1599580 - Set useAddonManager for extension tests that interact with GeckoView. r=snorp
Non-AddonManager extensions are invisible to GeckoView, so when a message like
"open a new tab" comes to GeckoView we will ignore them unless the extension is
properly registered.

Because of this, we need to register them in tests too.

Differential Revision: https://phabricator.services.mozilla.com/D56035

UltraBlame original commit: 7dc53077d58af4d8053adcc706e3ee0effe69555