Bug 1601630 - Remove WebConsoleUI evaluateJSAsync. r=jlast.
This method was directly using the main target console front
to evaluate something, which could cause issue in a fission
world.
The function is moved to the WebConsole class, and calls commands
evaluateJSAsync, in which we retrieve the appropriate front
given the given options object (e.g. if there's a selecteObjectActor,
we'll talk to the target that hold the actor).

Differential Revision: https://phabricator.services.mozilla.com/D55986

UltraBlame original commit: deda1cad1f6534824ff2f316c89f6f3ca057dc47