Bug 1601244 - Fix stub generator intermittent. r=jdescottes.
Due to the caching mechanism we had, running tests more than
once would fail because the Fronts of the first run were released,
which means the actorID of the cached object was set to null.
We simply remove this cache mechanism as it doesn't appear to
be needed at all for the test to pass.
We also freeze the console once we have all the packets
we need, so there won't be any further communication with
the server.
Finally we add test-console-api.html to the support files so
we don't get a 404 anymore.

Differential Revision: https://phabricator.services.mozilla.com/D55938

UltraBlame original commit: 8d6be96741bc68e7756b95af308ca84986c2a227