Bug 1598710 - Remove code that overwrites anchor class names in PopupNotifications.jsm. r=MattN
Looking at
https://searchfox.org/mozilla-central/search?q=notification-anchor-icon&case=false&regexp=false&path=
this code doesn't seem to be needed and it was a little weird for any anchor to rely on this
sort of implicit class rewriting. It certainly breaks any anchor that uses custom class names.

Note that in addition to this, there doesn't seem to be any need for supporting multiple
anchors for one notification anymore, so I'm going to open a separate bug for removing that.

Differential Revision: https://phabricator.services.mozilla.com/D56566

UltraBlame original commit: 98b4029eeaf3a9d03e158701b5d57f474a27c76d