Bug 1603627 - Write out a minimal target.json file for spidermonkey shell builds. r=froydnj
The browser build generates this file during the packaging step, which does not exist for JS shell builds. I made a brief attempt to call informulate.py as it does, but that seemed to require some things that are not available during the shell build. So I went with the cheap and easy approach to provide just the fields :jkratzer uses.

Differential Revision: https://phabricator.services.mozilla.com/D57028

UltraBlame original commit: 96e66af66715ac3914d4adca8d661d409273718f