Bug 1428705 - fix manifestparser/manifestparser for python3 and enable test_convert_directory.py r=ahal
Changes:

Import the appropriate version of `StringIO` instead of `BytesIO` depending on the version of python, and use it in `manifestparser.py`.

This is required for `test_convert_directory.py` to pass on both python versions. Changes to the test was not required.

Differential Revision: https://phabricator.services.mozilla.com/D56865

UltraBlame original commit: 9ffd4ecbfc2b3b9255e66588111a0af2140d0559