Bug 1592356 - Remove LazyScript::hasNonSyntacticScope. r=jandem
Update js::CanReuseScriptForClone to use baseScript() to determine if
non-syntactic. The HasNonSyntacticScope is not computed for lazy scripts so
fall back to checking the scope. Technically this is a fraction slower when
cloning a script with a syntactic global into a function with a non-syntactic
global, but in that case we are just about to clone the entire script
anyways. In practice this rarely comes up.

Differential Revision: https://phabricator.services.mozilla.com/D57209

UltraBlame original commit: 709cb020572064b32f88c24a644478572e1d0009