Bug 1553354 - Store messages when either the port or the delegate cannot be found. r=snorp
`port.connectNative` is synchronous, so there is the slim possibility that a
message could be sent before the app has been notified that the port has been
created. Similarly, if the app has not attached a listener yet, the port
message gets lost.

In either case, we now store this message in a list that we check when a
delegate is attached to the port object.

Differential Revision: https://phabricator.services.mozilla.com/D57511

UltraBlame original commit: 7ae6e38ed776eac9bdadf7acb5a157e01ad1d095