bug 1603833 - re-purpose unused intermediate preloading failure telemetry categories to investigate failures r=kjacobs,leplatrem
Intermediate preloading telemetry is overwhelmingly "failedToObserve", which at
the moment is a catch-all indicating that something in
updatePreloadedIntermediates failed. We need to figure out why, so this patch
re-purposes the categories "emptyAttachment", "failedToFetch", and
"unexpectedLength", which are currently not used, to indicate failures in
specific sub-operations in that function.

Differential Revision: https://phabricator.services.mozilla.com/D57676

UltraBlame original commit: d66517bc41f0fc49451d02cdc6b616ae8db2a436