Bug 1605207 - Remove unnecessary roots in XDRScript. r=caroline
We had a few extra RootedFunction when we only cared if about if the pointer
was null or not, so lets remove them. Also rename to isFunctionScript since
this is called for stand-alone functions as well. Semantics should be
unchanged.

Depends on D57862

Differential Revision: https://phabricator.services.mozilla.com/D57917

UltraBlame original commit: 0f37ad6a815e2508cc149d1d503e8fcc24645cf8