Bug 1604467 - Replace SRC_TRY source note with a JSOP_TRY bytecode operand. r=tcampbell
JIT compilation and bytecode analysis now no longer depend on any source notes.

The deleted code to print the offset-to-catch in Disassemble1 goes back to
bug 680228, before the newer code coverage implementation. It's unlikely that
anyone is still depending on this and it's also confusing to print an offset
in the bytecode disassembly that's not actually an operand.

Differential Revision: https://phabricator.services.mozilla.com/D58430

UltraBlame original commit: a590ca97f13989c17e649988099e1d4b8aa6c397