Bug 1605348 - Only trace the target of a WeakRef if the tracer requests weak edges r=sfink
The problem is the compartment checking tracer currently gets to see the target of a WeakRef, which may be in another compartment. I updated the trace method to check the traceWeakEdges() flag and set this to false for the GCMarking tracer. I was surprised that this wasn't the case before but I did a try run and everything looks fine.

Differential Revision: https://phabricator.services.mozilla.com/D57962

UltraBlame original commit: 7060854d0302a4a2cbb386c755820d89cfbf4262