bug 1606613: remote: stop logging handled errors r=remote-protocol-reviewers,whimboo
Errors that arise from use of the protocol, for example by the client
passing in the wrong argument or an operationg being unsupported, are
returned to the client.  Since these are not inherent implementation
problems there is no point to double-logging them in the server.

This is a left-over from when the remote agent was in a prototype
stage and we couldn't always rely on the Puppeteer client to
interact with Firefox correctly.

Differential Revision: https://phabricator.services.mozilla.com/D58508

UltraBlame original commit: ca9d646f56bad3537ece098dec3773b4fe05e1e5