Bug 1606960 - Use CollectScriptCoverage for getLCovInfo. r=nbp
This changes the implementation of getLCovInfo to use CollectScriptCoverage.
This uses the filename/funname information that was captured when the script
was created. This is what the browser ccov tasks use and this change allows
our tests to better test that mode.

Differential Revision: https://phabricator.services.mozilla.com/D58668

UltraBlame original commit: 3184b2da2c3c1552f087b0707645caa18266da38