Bug 1603140 - Multi-value returns in baseline wasm compiler r=lth
This patch implements one half of support for multi-value returns in the
baseline compiler.

The idea is that when you have a function that has stack return values,
it will receive an additional synthetic argument: a pointer to the stack
results area.  The layout of this area is the same as that understood by
BaseCompiler::pushBlockResults.

When compiling a function that has stack results, the stack result area
pointer will come in either in a register or on the stack.  If it's on
the stack we record that offset and will load a pointer at that offset
when returning values.  If the pointer is in a register, it gets spilled
to the stack after the spilled copies of previous register arguments.

When calling a function that has multiple return values, we bump the
stack pointer and pass an additional argument to the callee: the pointer
to that stack area.  After the call, in the baseline compiler, it may be
that we need to shuffle any stack results down, because incoming
arguments are consumed by the call.

Note that support for actually returning multiple values is not yet
implemented in this patch; we're just adding infra.  Also note that Ion
support isn't there yet either.  However all tests pass, because no
FuncType that has more than one result gets past the validator.

Differential Revision: https://phabricator.services.mozilla.com/D56756

UltraBlame original commit: 63f89c30085907e524cc8694cba029b9012f32e9