Bug 1604412 - Duplicate WinCompositorWidget logic into subclasses r=sotaro
This looks like a large change, but it's really just moving stuff
around.

It takes the logic in WinCompositorWidget and duplicates it into
its only 2 subclasses: InProcessWinCompositorWidget and
CompositorWidgetParent.

This is because CompositorWidgetParent is about to change *a lot*, but
InProcessWinCompositorWidget will basically stay the same. This is an
easy way to verify that I don't accidently break
InProcessWinCompositorWidget.

Differential Revision: https://phabricator.services.mozilla.com/D57428

UltraBlame original commit: 5bef477e99caae479472e79c5fd4cc9f6908826c