Bug 1605724 - Call sizeToContent() again when the icon loads if it hasn't loaded yet. r=dao
https://hg.mozilla.org/mozilla-central/rev/234701139a2a61d1262e609c9d8ac42384ecafda

Removed the following CSS rule:

  #iconContainer {
    -moz-box-pack: center;
    min-height: 55px; /* maximum icon height + icon margin */
    min-width: 58px; /* maximum icon width + icon margin */
  }

Which enforced the size of the icon row.

The icon loads asynchronously, so by the first time we fire DOMContentLoaded it
may not have loaded yet. This means that sizeToContent() will size the window to
an smaller size and stuff will wrap around when it loads.

<image> doesn't block onload so even delaying this wouldn't work.

Instead, call sizeToContent() again when it loads so that we size the window
right again.

Alternatively / on top of this, we could add a wrapper around the <image>
again and fix its width, may be better too...

Differential Revision: https://phabricator.services.mozilla.com/D58704

UltraBlame original commit: 7e410d72f9b01ced9c3ce520ae3bedb9a812ce2f