Bug 1605724 - Ensure loading the dialog icons doesn't change over-all layout. r=dao
https://hg.mozilla.org/mozilla-central/rev/234701139a2a61d1262e609c9d8ac42384ecafda

Removed the following CSS rule:

  #iconContainer {
    -moz-box-pack: center;
    min-height: 55px; /* maximum icon height + icon margin */
    min-width: 58px; /* maximum icon width + icon margin */
  }

Which enforced the size of the icon row.

The icon loads asynchronously, so by the first time we fire DOMContentLoaded it
may not have loaded yet. This means that sizeToContent() will size the window to
an smaller size and stuff will wrap around when it loads.

<image> doesn't block onload so even delaying this wouldn't work.

Restore the wrapper to ensure loading the icon or not doesn't affect layout of
the dialog.

Differential Revision: https://phabricator.services.mozilla.com/D58705

UltraBlame original commit: f81ec1f37a25b7a476091af3a8e75ddb51003f1d