Bug 1603765 - Split console browser.ini into multiple manifest files. r=Honza.
The mochitest-devtools-chrome can exceed its max-run-time
because the webconsole tests are only declared in one big
manifest file.
This patch split the manifest into 3 files:
- one for the browser console and browser toolbox console tests
- one for the jsterm tests
- one for the other console tests

These files are not split eavenly, but that's probably enough
for the mochitest-devtools-chrome to not be too long.

Differential Revision: https://phabricator.services.mozilla.com/D59468

UltraBlame original commit: 9ac818e0f3225d9fabfe965b389152465842ee48