Bug 1607573 - Part 1: Change GetRealFrameElementOuter to use BrowsingContext, r=bzbarsky
This does not fix consumers of GetRealFrameElementOuter, which may still act
incorrectly, as the frame element will not be available in cross-process
situations, but removes a user of the docshell tree.

This has a small behaviour change, in that it causes 'window.realFrameElement'
to return 'null' when run within an <iframe mozbrowser>. A quick inspection of
use-sites suggested that this would not be an issue, and it appears to pass a
try run.

Differential Revision: https://phabricator.services.mozilla.com/D59398

UltraBlame original commit: b7180d7d7db1a556b35e120549cd0a93b8a0c67e