Bug 1609466 - offer 'open with' options even when Windows is clueless about the mimetype, r=bzbarsky
We currently fail to offer 'open with' options for application/zip mimetype
.zip files, because out of the box, the Windows registry does not contain
mimetype entries for application/zip. In fact, it contains comparatively
few entries, and right now we bail out with an empty mime info object if
we cannot find the server-provided mimetype in the registry (or cannot
find an associated extension for that mimetype).

Prior to the changes from bug 1597985, we would first try to find a handler
based on the default extension for the mimetype, and used the file extension
if that failed.

The changes in this commit will cause us to fetch the default file extension
for the mimetype, but continue with the provided extension if no mimetype
information is available, restoring functionality in the case where Windows
has no mimetype information.

Differential Revision: https://phabricator.services.mozilla.com/D60139

UltraBlame original commit: 4c34f33d529c2e2a182b0cced92f9769625a6f5e