Bug 1606782 - part3 : replace assertion with returning result of register/unregister controller r=chunmin,bryce
We should ensure that controllers has been removed from the service when it destroys. Therefore, we remove the assertion in `MediaControlService::RegisterActiveMediaController()` and `MediaControlService::UnregisterActiveMediaController()`, and return the result to allow caller to know if it has been registered or unregistered or not in order to do error handling or add assertions.

Differential Revision: https://phabricator.services.mozilla.com/D59286

UltraBlame original commit: 45684b8d59e1a30731d446ee437ed7050432982c