Bug 1569626 Part 1: Save and restore the presshell resolution when creating and destroying an MVM. r=botond
We need the resolution to be restored when and if the presshell creates
a MobileViewportManager. That decision happens in
PresShell::UpdateViewportOverridden. In order for the resolution to be
saved somewhere it has to exist outside the presshell itself, since
the presshell will be destroyed and recreated during some of the
operations that also destroy the MVM -- such as closing the devtools'
Responsive Design Mode. So Document holds this resolution and gets
accessor methods to allow the presshell to get and set that resolution
as needed.

Differential Revision: https://phabricator.services.mozilla.com/D59821

UltraBlame original commit: 5e1cba38501bcd6cbd25049c056485c30a2b2c65