Bug 1594820: Part 3 - Modify GeckoProcessManager to bind and start child processes asynchronously on the XPCOM launcher thread; r=geckoview-reviewers,agi
The primary purpose of this patch is to convert the internal sequence of
service binding and invoking of `start` to asynchronously run on the XPCOM
launcher thread via `GeckoResult`. Because more of the code now runs on the same
thread, many of these methods no longer need to be `synchronized`.

Disconnecting via `unbind` is also modified to use the launcher thread and
`GeckoResult`.

Note that no changes have been made yet to enable multiple processes of the
same type; those changes will be made in bug 1595834.

Differential Revision: https://phabricator.services.mozilla.com/D57839

UltraBlame original commit: d6d2afaee49da11c1da5d05ec8ed59372ac13438