Bug 1609138 - Clean up float registers on arm64. r=rhunt
Create a clearer distinction between the register's Encoding, which is
its hardware name, and its Code, which is a dense encoding of
bitwidth+Encoding along with a distinguished Invalid value.  These
concepts exist already but it gets out of hand when the FloatRegister
uses a Code to encode the Encoding.

Make FloatRegister contain separate fields for bitwidth, encoding, and
validity, as it does on other platforms.

Add assertions on validity of inputs and on the validity of the
FloatRegister for some operations.  And tidy up some, and rearrange
the file to mirror the x86 file as much as possible.

Expand the register name table so that it covers the possible range of
Code and so that we won't reference the table OOB.

Differential Revision: https://phabricator.services.mozilla.com/D59912

UltraBlame original commit: 8c7371a02fd7f7e6da78289ae3d5c22949490204