Bug 1609199 - Remove Emulation Actor r=mtigley
We still need the front and the spec files to talk to older servers.
However we don't maintain forward compatibility, which means the server is not supposed to work with older clients.
So the emulation actor can be removed.

Differential Revision: https://phabricator.services.mozilla.com/D60294

UltraBlame original commit: 219e155d358e4bdae7ef4eb7ca8b1e853948a8f3