Bug 1155772 Part 9 - Convert FlexItem::mMargin to LogicalMargin. r=mats
Note that I return LogicalMargin by value for GetMargin(). It is
considered a suboptimal practice to return a constant reference to a
member variable. We won't gain performance by returning a reference
because the compiler can optimize the return value. If it were assigned
to a new variable, the copy/move constructor is called anyway.

As noted in FIXME(TYLin), the AxisOrientationType argument in
GetMarginBorderPaddingSizeInMainAxis() and many other methods can be
removed. However, to make the patch size manageable, I'll remove them in
a later part.

Differential Revision: https://phabricator.services.mozilla.com/D59051

UltraBlame original commit: d03d935a66df2a751d11a647b4eb0f865697690c