Bug 1610738 - Add primitive flag to specify a compositor surface is preferred r=sotaro
This allows calling code to specify whether a primitive would prefer
to be promoted to a compositor surface and/or picture cache slice.

This is a performance hint that can be used for large external
primitives, such as videos and canvas elements.

Differential Revision: https://phabricator.services.mozilla.com/D60637

UltraBlame original commit: 71ed65a465d2bcfcf8e02ab8779c6700754ccdb2