Bug 1609992 [wpt PR 21239] - Limit inline negative margin quirks only when line overflows, a=testonly
Automatic update from web-platform-tests
Limit inline negative margin quirks only when line overflows

crbug.com/979894 and crbug.com/1011816 revealed that, boxes
with inline negative margins can be put back to the line even
if they appear beyond the end margin. The behavior is not
defined in the spec, but supported because all browsers do it
and there are good amount of sites relying on the behavior.

The special behavior, however, turned out that it should
apply only when the line overflows.

As described in fixes for the above issues, this behavior is
not well-defined, not really interoperable, and sometimes
the result of combinations of bugs. This patch takes the parts
of observable behaviors that are interoperable and reasonable,
while fixes the reported page.

Bug: 1040461
Change-Id: I3e4883220645a59dd3c33bdad4116f200fc4208d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2000412
Commit-Queue: Koji Ishii <kojiichromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrickchromium.org>
Cr-Commit-Position: refs/heads/master{#733154}

--

wpt-commits: 2f346208c2a115a601b580e7869fe112b97c4246
wpt-pr: 21239

UltraBlame original commit: e9978cee8889d8c0a14f7599fca6bf39122654f0