Bug 1588152 - don't spawn ServiceWorker if it's known it'll be immediately terminated r=dom-workers-and-storage-reviewers,asuth
We can optimize for the case where the target content process that a worker
will be put in is still starting and and we know that we will be terminating
the worker. Instead of starting up the worker and immediately terminating it,
we can just never start it up. RemoteWorkerManager will use the value of
RemoteWorkerController::IsTerminated to determine this.

Differential Revision: https://phabricator.services.mozilla.com/D61391

UltraBlame original commit: a307dba8bddd7ff0c902f18f3c3fc91ffd1f08d0