Bug 1613599 - Implement final eager evaluation design. r=jlast.
This patch adds a dedicated EagerEvaluation.css file so
it's easier to see all the related rules and properties.
The colors of the results are tweaked (mostly to grey)
so eager evaluation doesn't feel as intrusive.
The font-size is lowered for the same reason.
When eager evaluation is enabled, the padding of the console
input is reduced in order to have the input + eager results
to be 40px.

Differential Revision: https://phabricator.services.mozilla.com/D61806

UltraBlame original commit: 37865176d8d00b40c5ab0ecea4c244cd6a62c8ba