Bug 1523853 Part 3: Further relax the mousewheel test to not care about the ending scroll values. r=kats
This is being done to reduce the intermittent failures we see in this test.
There are other non-intermittent tests that measure the specific number of
pixels traveled by mousewheel events. The focus of this test seems to be
"Ensure that mousewheel scrolling and zooming are mutually exclusive" and
this change maintains that.

Depends on D62067

Differential Revision: https://phabricator.services.mozilla.com/D62130

UltraBlame original commit: 37ffe2e2412b1a3c33c41977d3e537161f150e85