Bug 1611439 - Stop AsyncLogger before aborting the audio session. r=padenot
AsyncLogger creates a thread that the log messages are pushed and printed. This thread was not stopped and joined after it was used, which created a shutdown crash. This patch will stop the AsyncLogger thread when the capture session has finished.

Differential Revision: https://phabricator.services.mozilla.com/D60978

UltraBlame original commit: 7cf64e7350acec67a977bd70dacf6d3ddcbf5227