Bug 1607505 - Move StartupCache destructor later in shutdown r=froydnj
Since we've decoupled the writing of the StartupCache from the freeing
of the StartupCache's tables (which takes nontrivial time), this patch
moves the StartupCache's destructor later in shutdown so it will be
skipped in the early _exit(0) efforts. There is a remaining
WaitOnWriteThread in the StartupCache's destructor, but this is a
minor sanity check to avoid use-after-frees in the write thread during
shutdown. In general it won't matter as we already wait for the write
thread in the xpcom-shutdown observer, and if we shut down during write,
the system is set up to be okay with that (because it already happens).

Differential Revision: https://phabricator.services.mozilla.com/D62295

UltraBlame original commit: d79f1d283acde1a971fe0e5e6f1a8db075f8589a