Bug 1562724 - Fix "execute getter" button in Debugger Preview popup. r=nchevobbe.
The 'getter button' wasn't working correctly on root level items in the preview popup.
A check is added on root item to see if it has any corresponding evaluation result,
and if so take that value instead of the original one.
A test is added to ensure that the button now works on those top-level elements in
the preview popup.

Differential Revision: https://phabricator.services.mozilla.com/D61086

UltraBlame original commit: a4d1d2c08331d0965556b4abfa8ebf2c1b9c9dac