Bug 1609881 - Part 2: Move fork server launcher to final-ui-startup. r=gsvelto
Accessing StaticPrefs at xpcom-startup is too early to cause the
initialization of it.  The preferences from user.js and profile are
not load yet.  So, later, it would cause error of changing the value
of static prefs that is marked as mirroring once.

Differential Revision: https://phabricator.services.mozilla.com/D62866

UltraBlame original commit: 19b4dadd16f4e8e53996225c95020576391e776e