Bug 1614472 [wpt PR 21704] - Fix is_alive() usage and always call it as a function., a=testonly
Automatic update from web-platform-tests
Fix is_alive() usage and always called it as a function. (#21704)

There is some confusion in the way the method is_alive is used.
The code seems to assume that this method is accessed like
a variable for Protocol and WebDriverServer classes (by using the
property decorator in the is_alive() definition).
But then some class that inherit from those and re-declare is_alive()
forget to declare the decorator.

For example, the base class Protocol declares is_alive as a property,
but some of the classes that inherit from it like WebDriverProtocol
from executorwebdriver.py, ServoWebDriverProtocol or SeleniumProtocol.

This caused the code that checks for is_alive to be broken for those
classes because "if is_alive" its always true.

To add to the confusion about how this should be used, the browser
classes have also a method named is_alive(), where the current code
assumes it is used as a function.
This has already caused previously confusion (see 8be26a07bc).

Fix this by ensuring is_alive() its always used as a function and
remove the property decorators from the classes that were declaring it.
--

wpt-commits: 5da9ebc80b58ccd9abe5c2022fc659e645466314
wpt-pr: 21704

UltraBlame original commit: bf5900d28a8508adf2897ed8610e4b470103270e