Bug 1615418 - Do not throw when calling watch/unwatchFront on destroyed Fronts r=daisuke,ochameau
Depends on D62893

The issue here is that we are trying to destroy the workers-listener after the target was destroyed,
and calling unwatchFront on a destroyed Front throws.

Most of the fronts monitored in workers-listener are handled by the watchFront API, so properly adding
onDestroyed handlers fixes some issues. However the rootFront cannot be handled with a similar pattern
at the moment.

In general, I think making watchFront/unwatchFront safer to call makes sense, but I could also check
if the rootFront is already destroyed in workers-listener's destroy

Differential Revision: https://phabricator.services.mozilla.com/D62894

UltraBlame original commit: a0cc5557495d8853a74dfbe275d7a46775cd6032