Bug 1615048 [wpt PR 21763] - Update chromium logger to identify subtest failures., a=testonly
Automatic update from web-platform-tests
Update chromium logger to identify subtest failures. (#21763)

This also fixes a bug where a subtest failure would be hidden because
the expected_status was evaluated too late in `test_end`. For a
subtest failurei the test status is Pass. But for Chromium, we flip this
"actual" status to a failure.
But the actual status is also used as a fall-back for the expected status
(see explanation in chromium.py:126). So if "expected" is evaluated after the
"flip", it will just match "actual" and the failure will be marked as
an "expected" failure.
--

wpt-commits: 22a79941ccc023ef301a48734158451f1b8873fb
wpt-pr: 21763

UltraBlame original commit: ec2e37b3e9188a10c7aa650195da6a7024da057d