Bug 1595980 - Move return value adjustment into the debuggee scope. r=jimb
We move adjustment of the return value until the very last step before applying
the value itself, because it is important that any adjustments performed take
place while inside the debuggee's context, and most importantly in the
context of this but, inside the debuggee's microtask queue.

AutoDebuggerJobQueueInterruption swaps out the microtask queue for one
specific to the debugger, and this bug happened because we were queuing
debuggee promise jobs in the debugger's microtask queue.

Differential Revision: https://phabricator.services.mozilla.com/D57727

UltraBlame original commit: 042af67cea49cb6bfab3e9dd7c06e449b2a089c7