Bug 1614295 - Provide a way for apps to know when an extension is installed. r=snorp,ochameau,esawin
This patch adds a `onExtensionListUpdated` method to `DebuggerDelegate` which
is called whenever devtools install a new extension.

This method provides an opportunity for apps to refresh the list of installed
extensions and sets appropriate delegates so that the new extension is
correctly recognized.

Differential Revision: https://phabricator.services.mozilla.com/D62333

UltraBlame original commit: e59914273e5bc8a5e34250b429662ae4c4dffbaa