Bug 1615362 - Registered targets shouldn't keep their FinalizationGroup alive r=sfink
FinalizationGroups are no longer held live by their targets but are still held live when their cleanup callback is queued. According to the spec implementations are allowed to cancel FinalizationGroup callbacks if the FinalizationGroup dies, but I didn't do that because it adds complexity (we would need the browser to have weak pointers to queued groups).

The patch makes FinalizationRecordObject's FinalizationGroup pointer weak by storing it in a private value and sweeping it as necessary. We need a trace method to update the pointer when it's moved by the GC, but only if traceWeakEdges() is true. We also need a read barrier in case we want to queue cleanup callbacks for an otherwise unreferenced finalization group.

Then this adds a set of all active record objects to the FinalizationGroup objects so we can clear their pointer to group when it dies. We can't do this during sweeping because the group may not be in the same sweep group as the target and so may die earlier or later.

Differential Revision: https://phabricator.services.mozilla.com/D62786

UltraBlame original commit: 2575d6f0dca35e6b2bf2925b1aea849883f94801