Bug 1617984 - Fix mach bootstrap failure on Archlinux when PKGDEST is set; r=firefox-build-system-reviewers,rstewart
Makepkg has a commented config in /etc/makepkg.conf to set the path for packages to be built:
```
#PKGDEST=/home/packages
```
When you uncommit this line, packages will be built in the specified path, instead of cwd. This causes the bootstrap script to fail.

This commit will override the PKGDEST config so that package is built in cwd and could be located by the following lines of script.

Differential Revision: https://phabricator.services.mozilla.com/D64195

UltraBlame original commit: a50491a731c2a18ae6728baff5543a93cf9856aa