Bug 1614640 - Improve reftest suite_start data to provide consistent group/manifest names even when not run-by-manifest; r=ahal
Reftest-family test suites can be "run-by-manifest" (reftest-plain, on desktop), or not
(all reftest-family suites on Android, crashtest and jsreftest on desktop). When not
run-by-manifest, code in reftest.jsm is used to format the suite_start data; when
run-by-manifest, code in runreftest.py is used. Currently, the reftest.jsm implementation
submits a simple array of test IDs -- no associated manifests -- possibly consistent with
the concept that tests are not being run/grouped by manifest.

This patch updates the reftest.jsm implementation to generate the same format of
suite_start data as the python implementation: An array of dictionaries mapping
manifest IDs to arrays of associated test IDs.

Differential Revision: https://phabricator.services.mozilla.com/D63875

UltraBlame original commit: 0480c061117c323e661299fa584cbbd307707501