Bug 1611710. Don't restrict the draw target to the visible area.
Sometimes the painting code will look at the clip which is derived
from the intial size of the surface and not draw if things if they
are outside of it. We want to draw the entire item so use dtRect
instead of visibleRect.

Differential Revision: https://phabricator.services.mozilla.com/D64277

UltraBlame original commit: bd13b92ec265ff0927387fda64cb51dcd2993570