Bug 1618564 - Adjust an assertion to check for what it really wants to check. r=tnikkel
There's more stuff that creates nsStyleBorder structs on the stack (like MathML
tables).

Check for what we really mean to check, which is that the image request is the
one of our border-image-source. Otherwise we'd need to setup the association and
so on somehow.

Differential Revision: https://phabricator.services.mozilla.com/D64650

UltraBlame original commit: df0d58ec8020cda10256612b43841563a096c7b1