Bug 1617523 - Ensure remote backbuffer only reused if initialized properly r=jrmuizel
This crash was caused because there was a case where the Provider would
store the SharedImage even if it failed to initialize it or send it to the
Client.

When the Client next requested a borrow, the Provider would see it was
already storing a SharedImage and would tell the Client to just
reuse the one it already had.

Since the Client never actually received it, it would end up dereferencing
a null pointer.

The fix for this is to wait until the SharedImage is fully initialized and
shared with the Client before storing it. That way, we know that if we have
it then so does the Client.

Differential Revision: https://phabricator.services.mozilla.com/D64659

UltraBlame original commit: 992f5734eb4e39b3c76ff35b0f744b129f80b483