Bug 1615931 - Handle invalid bookmark URLs in new bookmark sync. r=markh,LougeniaBailey
This commit changes `Store::local_row_to_item` to validate local URLs,
and flags items with malformed URLs as invalid. These items are either
replaced with valid remote copies, if they exist, or deleted if not.
Additionally, `BaseBookmarksStore#_calculateIndex` no longer throws
when determining the sort index for an item with an invalid URL.

As an aside, we use the `url` crate to parse URLs. This is the same
crate as used by `MozURL`, which, in turn, backs the JS `URL`
constructor...so URLs should be validated the same way in Rust and JS.

Differential Revision: https://phabricator.services.mozilla.com/D64695

UltraBlame original commit: 0c47f5848279e59e45c6acdc7d207db0d648f651