Bug 1557765 - Part 2: Inline through derived class constructors. r=jandem
Allow to inline when callee and new.target don't match, as long as the callee
is a derived class constructor, because for derived class constructors we don't
create the this-object at the call-site. This allows to inline through a chain
of derived class constructors.

Differential Revision: https://phabricator.services.mozilla.com/D64971

UltraBlame original commit: 9e219b83117104abaaf2be4a1da3088c86ff4e70