Bug 1619578 - Make the profiler WebChannel work on beta; r=canaltinova
The WebChannels were not being initialized on beta due to being initialized
after the feature pref was checked. This patch changes it so that
profiler.firefox.com will still have the WebChannel registered, and then
if the menu button is enabled via the WebChannel, the feature flag is flipped.

I did not include a test on this, as this is a temporary measure until we flip
on the mechanism for all users.

Differential Revision: https://phabricator.services.mozilla.com/D65199

UltraBlame original commit: 2e1a978b09d77c7422e972e9c3bfdeb484f6f608