Bug 1617147 - Stop setting check_prog result to : for missing programs. r=nalexander,rstewart
This was cargo culted from the autoconf equivalent, and while it makes a
command that does "$(PROG) foo" work because it becomes ": foo", that
may or may not actually be a desirable outcome.

OTOH, we do have some places where there are some "ifdef PROG" that are
just plain wrong when PROG is always actually set.

One place I do know that does check if the value is not ":" is for
OBJCOPY, which is still set from autoconf.

All in all, looking at all the check_prog(allow_missing=True) we have in
python configure, it doesn't seem anything is checking for ":", and that
doesn't seem like the right status quo.

Differential Revision: https://phabricator.services.mozilla.com/D65276

UltraBlame original commit: b5069c471cc989080299b245d91e1c92c3cac89e