Bug 1616847 - When xulWrapper is visible move it instead of trying to show it again. r=jdescottes.
If the XUL wrapper was opened and a consumer called show again,
the panel wouldn't update its position. This can be fixed by checking the
state of the element, and if it's open move the panel using `_moveXulWrapperTo`.

With this change in place, we can remove the updateContainerBounds methods,
as it has the same arguments and is only used in one place.

Differential Revision: https://phabricator.services.mozilla.com/D63488

UltraBlame original commit: 2a8821347dd80a2b05697b739ada15dcc1dbf300