Bug 1620476 - Fix a test that makes assumptions about the padding of a textarea. r=mstange
The test is duplicated in two places, so four files to change. The padding in
the non-native theme is just big enough so that the textarea overlaps the float.

In other themes the textarea overflows, but not enough to hit the float.

Differential Revision: https://phabricator.services.mozilla.com/D65663

UltraBlame original commit: 296cb55bfa55d21f23b07808eb2e1eea91d19a7d