Bug 1594798 - Store remote prefs on the client under a separate pref branch; r=julienw
Currently the preferences for remote profiling are stored on the debuggee. This leads
to a negative user experience, as oftentimes phones do not persist the preferences.
This patch changes the strategy to store one set of preferences for local profiling,
and a second set of preferences for remote profiling.

Differential Revision: https://phabricator.services.mozilla.com/D65148

UltraBlame original commit: 79f957c57550674192769fe6714bcc24eb70c970