Bug 1617673 - Give the root element in the bookmark properties window an id so that persistence works again r=mak
The ID was removed in Bug 1585482 and moved to the child dialog element which breaks xulstore persistence.
This gives the dialog a new id and uses the original "bookmarkproperties" for the window so any existing
xulstore values will work without a ui migration.

Differential Revision: https://phabricator.services.mozilla.com/D65829

UltraBlame original commit: 1bf266a19339343869b4464f7f909cd032397627