Bug 1621319 - Make test_sanityEventUtils hit the border of the input, not the contents inside. r=smaug
Depending on the widget implementation details, hitting one pixel inside the
input may hit the input itself, or may hit the anonymous scrollable element
instead.

synthesizeMouseExpectEvent checks the originalTarget, for some reason:

  https://searchfox.org/mozilla-central/rev/070a000dd49aac4a26147e137efcd91a728d13b8/testing/mochitest/tests/SimpleTest/EventUtils.js#1419

I'm not sure what the reason for that is, or why test_sanityEventUtils needs to
test it using an <input>, but I don't want to think too hard about it.

This ensures we hit the border on the <input> and thus target the right element.

Differential Revision: https://phabricator.services.mozilla.com/D66241

UltraBlame original commit: 9181798549f23ad1444142ac9d69a5aef90c9f1e