Bug 1614871 part 1: Make non-Windows ProxyAccessible::ChildAtPoint cross into in-process iframes. r=eeejay
Accessible::ChildAtPoint can return an Accessible in a descendant document.
Thus, we must return the result PDocAccessible via IPC, not just the id.
Previously, we only returned the id, but we'd fail when we tried to look it up if it belonged to a descendant document.

Differential Revision: https://phabricator.services.mozilla.com/D66710

UltraBlame original commit: 0f241c48730c6e4cce24bc9970408087045ca2ab