Bug 1618913 [wpt PR 22156] - [Gecko Bug 1618913] Preload Firefox instances in wpt harness, a=testonly
Automatic update from web-platform-tests
Preload Firefox instances in wpt harness

The wpt harness spends a lot of time — about 20% of the total runtime
on Linux64  — between tests. The majority of this is shutting down and
restarting the Firefox instance under test. To reduce this
overhead we add the option to preload a Firefox instance which can be
immediately used when a new session is required. This makes the slow
process of creating a profile and starting the browser
asynchronous. Testing shows that this cuts the between-test time to
between a half and a third of the previous value.

The cost of this change is that we are adding background load at the
time tests are running, which could influence the results and add
randomness. However that seems unlikely unless we are right on the
edge of the system performance.

Differential Revision: https://phabricator.services.mozilla.com/D64954

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1618913
gecko-commit: 1da2a939d9cf39156b38b639f77d777b0821eb4e
gecko-integration-branch: autoland
gecko-reviewers: karlcow

--

wpt-commits: 660e74817e4d0ac21ad21de6764e9ac2902704bd
wpt-pr: 22156

UltraBlame original commit: 95ce2ba147c6bfa5da943e8d190d28c97c038279