Bug 1622893 - Don't use background color for 3d colors on GTK. r=stransky
On high contrast themes, we avoid using the colors from the author to ensure
contrast.

We allow the border-style though, and that unfortunately means that:

  <input type=text style="border: 1px solid black">

Ends up rendering like:

  <input type=text style="border-style: solid; border-width: 1px">

Which for a theme with a white background means that we'll render a white
border which users can't see, and is unfortunate.

Hard-code these colors, as using the background color for the light variant
doesn't seem right anyway, and the dark variant was hard-coded already.

These colors are taken from the cocoa widget. You can see the change in colors
with something like:

  <input type=text style="border-width: 2px">

For default themes it makes the colors a bit less subtle. But I don't think it
matters all that much in practice since the fallback unthemed border is very
rare on the web these days.

The rendering here is still not great, to be clear (we may want to tweak the
high-contrast heuristics here), but it's way better than invisible borders.

Differential Revision: https://phabricator.services.mozilla.com/D67127

UltraBlame original commit: 5497bb642dfd2c6a7f1a9820c77ebe2be9232010