Bug 1623333: part 1) Throw error in Selection's webidl methods, when there's no `nsFrameSelection` instance. r=smaug
Alleviates the possibility to add calls which are unintentionally no-ops.

In practice, these errors should never be propagated to the web, and
signal to other callers (for copy&paste) that the calls are wrong.
Ideally, the other callers wouldn't see these methods, but that requires
a larger refactoring.

Differential Revision: https://phabricator.services.mozilla.com/D67327

UltraBlame original commit: d05172dd2297d4cfe92564146252118055c2552a