Bug 1623602 - Properly guard a test case.  r=rhunt
An earlier change would disable SharedArrayBuffer in the JS shell for beta and release builds.
This has uncovered some broken guards in various test cases, in this case, a test on whether
the shell built-in setSharedObject is available.  However, that is not affected by the change
in the setting, so the guard really should test both setSharedObject (to ensure it's only
tested in the shell) and SharedArrayBuffer (to ensure it's only tested when shared memory
is available.)

Differential Revision: https://phabricator.services.mozilla.com/D67474

UltraBlame original commit: 86267e092be2e6ccca476bec2f5776fcdd643672