Bug 1595244 - Add unit test for LoginAutoComplete._isProbablyANewPasswordField r=MattN
For the 'Basic password change form' case, the confirm password field should also be true, as selecting the password generation in the popup will fill the same generated password as in the new password field before it without the user having to copy it. Passwords are cached per origin.

There is an existing bug on file to change the text of the autocomplete popup in this case to make that more clear that this is what happens for the confirm password field (Bug 1548854), and other bugs are on file to make it possible to have more than one generated password for the same site in various scenarios (Bug 1569568 and Bug 1551723).

Differential Revision: https://phabricator.services.mozilla.com/D67723

UltraBlame original commit: c72482e89c472d7d194bd97fec6e57549a79de90