Bug 1622254 - Match repeating tile sizes more aggresively to their snapped bounds. r=kvark
When we calculate the repeating tile size, errors can be introduced,
such that there are cases it should be the same as the bounds, but is
slightly different. This can cause a new repetition where we did not
expect one. This patch makes the comparison when may force the tile size
to be the same as the snapped bounds more fuzzy.

Differential Revision: https://phabricator.services.mozilla.com/D67620

UltraBlame original commit: 93cf130f59f97acc6ebb84c3505880cdcaca36c0