Bug 1621022 - Change the workflow of enabling the profiler menu button; r=canaltinova
This patch makes it so that the profiler shortcuts work based on the location of
the profiler menu button. It changes it so that if the menu button is in the navbar
or other menus, the shortcuts will work. Otherwise, the shortcuts will be a no-op.

This removes the Tools -> Web Developer - Enable Profiler Menu Button option. By
default on Nightly and Dev Edition the profiler menu button will be available.
On other channels, users must visit profiler.firefox.com.

Differential Revision: https://phabricator.services.mozilla.com/D66785

UltraBlame original commit: bc5fee900c024eadf7cca28e7717aa947ff7ce94