Bug 1624247 part 2: Rename flex item helper-struct from CachedMeasuringReflowResult to CachedBAxisMeasurement. r=TYLin
This patch should not impact behavior; it is purely a rename.

The "reflow result" naming is too generic for this struct, now that I'm
planning to cache other results from flex item reflow (in later patches).

This particular struct is caching block-axis measurements (a BSize and an ascent), so let's rename it to reflect that.

Differential Revision: https://phabricator.services.mozilla.com/D67801

UltraBlame original commit: 7d8eddd3ede1babf661c95a672d0188b4af221c0