Bug 1624655 - Don't backplate over themed backgrounds. r=morgan
data:text/html,<input type=submit> is white-on-white when depressed with
backplating enabled on the GTK light theme, because even though the computed
background-color is white, the actual button background is themed and the color
is not white instead.

This causes problems with the unthemed appearance anyway (<input type=submit
style="-webkit-appearance: none"> and so on), but this is probably worth it
anyway.

Differential Revision: https://phabricator.services.mozilla.com/D68063

UltraBlame original commit: cb8f33fc316772581fc4a03e02571f483c79bdaa