Bug 1620113 - part4 : update main controller when controller changes its playback state to playing. r=chunmin
The main controller is the only controller that can receive media control keys among all other controllers registered to the service, and it should always be the last controller which is playing.

Therefore, we would maintain a list which sorts the controllers based on the order of playing controller, and the last one (last playing controller) would be the main controller.

For example, if the controller A starts first, then B starts, B should be the main controller. But if A restarts later then A should become the main controler again.

Differential Revision: https://phabricator.services.mozilla.com/D66857

UltraBlame original commit: 5241e4aab772ac8c55999c144731313360d85e5e