Bug 1625251 - Stop using MOZ_DIAGNOSTIC_ASSERT when checking blocking decision. r=baku
We turn the MOZ_DIAGNOSTIC_ASSERT() in NotifyBlockingDecision() to
MOZ_ASSERT() to prevent further crashes for users. And we will
investigate why this assertion is hit and check it back to a diagnostic
assertion once we figure out the reason.

Differential Revision: https://phabricator.services.mozilla.com/D68517

UltraBlame original commit: 11de4f35419b126e27ea77003fc09fe8454c19fe