Bug 1615257 - Enable google chrome android browsertime tests. r=perftest-reviewers,AlexandruIonescu
This patch adds the capability to run Google Chrome for Android tests through Raptor-Browsertime.

Chrome must be available on the device being tested, and there are no installation steps being added in this patch since CI already has the latest Chrome version available to test with. That said, this patch adds the ability to find the version of Chrome that is being tested and stores this in the Perfherder artifact. Getting this version is also necessary to be able to run Chrome with the correct chromedriver.

Two tests are initially be enabled for Chrome in this patch: Amazon, and YouTube. They will only run through a cron task three days a week. The other changes done in this patch are required for Chrome to work with Raptor-Browsertime.

Differential Revision: https://phabricator.services.mozilla.com/D65610

UltraBlame original commit: 07d40de55dbbcf4b9ca62c6095a7d75cf4c68c6c