Bug 1616788 - Part 6: Propagate the HasStoragePermission from the loadInfo to the WindowContext. r=dimi,baku
We propagate the HasStoragePermission flag from the loadInfo to the
WindowContext in the patch. We add a flag HasStoragePermission in the
document and this flag will get updated when the
Document::StartDocumentLoad() happens. And then, we would sync this to
the WindowContext in the final stage of the
nsGlobalWindowOuter::SetNewDocument() where the WindowContext is ready.

Differential Revision: https://phabricator.services.mozilla.com/D67471

UltraBlame original commit: 05e0f36e1ef9b5cae5acbf3fd232ee98c8eb6c0d