Bug 1625419 - Don't double-apply non-symmetric vertical padding for number inputs. r=dholbert
Turns out that align-self is enough for sizing to end up like we want (not
getting squashed by the flex container's padding). This prevents double-applying
the padding in cases where the top and bottom paddings don't match.

We also fix another change from my previous patch, which is that we wouldn't
clip stuff if the text editor is much bigger than the flex container. Most
trivial example of this is:

  data:text/html,<input type=number style="font-size: 300px; height: 40px;">

I renamed the test from overflow-clip-box-* to padding-*, as we're not
technically exercising overflow-clip-box anymore, and added a test for the
clipping too.

Differential Revision: https://phabricator.services.mozilla.com/D68535

UltraBlame original commit: 4627b29188165a96ef1a0b245b27d956dc6d1163