Bug 1626198 - Test RTL bookmarks panels better. r=ntim
I can reproduce the same test failure if I add a flush after setting the
direction value in changeSidebarDirection.

This is arguably a XUL tree bug in a way, but it's a tough one to fix. But the
test is testing something else, and is testing it poorly. Make it test RTL
sidebars properly.

Differential Revision: https://phabricator.services.mozilla.com/D68993

UltraBlame original commit: db538dca6ea1b0c08be5272a5ad1c506c6e0b1ae