Bug 1626201 - Support OSR in WarpBuilder. r=iain
The strategy is the same as in IonBuilder: create an OSR entry block and a second
block merging the two predecessor blocks before the loop.

This is a lot simpler than IonBuilder because we don't attempt to type-specialize
OSR values. If this becomes an issue we'll need a strategy for that. Hopefully
this will be good enough for now until we make bigger changes to how OSR and the
Ion backend work.

Differential Revision: https://phabricator.services.mozilla.com/D68975

UltraBlame original commit: ad15dd339067a02616f194f804b5d7e3562c947b