Bug 1615109, part 1 - Get rid of the unused content process share picker registration. r=marcosc
This file has an entry in kWidgetContracts for kNS_SHAREPICKER_CID, but not in
kWidgetCIDs, so there's no implementation. I think this is causing the "Could
not map contract" warning message. We never use the share picker in the content
process, so just remove this.

Differential Revision: https://phabricator.services.mozilla.com/D69146

UltraBlame original commit: baebb5cd31634769b1ccf784393ca571fa97fa5c