Bug 1626615 - Fix forced-return propagation for Ion frames after debugger has been disabled. r=loganfsmyth
With WarpBuilder enabled we failed the Frame-onStep-resumption-05.js jit-test. This
failed because when we are propagating a forced-return and turn off the debugger,
we ignored this forced-return for Ion frames. This patch fixes that test failure.

It also tidies up the code with a ShouldBailoutForDebugger helper function.

Differential Revision: https://phabricator.services.mozilla.com/D69219

UltraBlame original commit: 4d282675fe945959f257075d674ca2f89942db8f