Bug 1626621 - Determine os version and architecture at runtime for FOG. r=chutten
All of this is currently done in C++ and then passed in.
Eventually we might want to do that in Rust, but the services are not
yet exposed to Rust, so we opt for the short way of passing in data.

Depends on D69366

Differential Revision: https://phabricator.services.mozilla.com/D69367

UltraBlame original commit: 2e661ffb15f2d0c90deb73d497b4752a94b4168d