Bug 1626859 - Port the loop iterator mechanism from IonBuilder to WarpBuilder. r=iain
This unfortunate mechanism is needed for certain iterator loops. This fixes the
ion/bug1197769.js jit-test to not crash in the exception handler.

A few weeks ago bug 1620960 refactored and cleaned up the IonBuilder code so this
is just a port of that to WarpBuilder with some minor cleanups.

Differential Revision: https://phabricator.services.mozilla.com/D69321

UltraBlame original commit: 10cd10a7f6d1c67f6b151fa891e19de615cb426c