Bug 1625181 Skip over unneeded work when installing system/builtin addons r=zombie
We don't put system addons on the blocklist so there's no need to compute
blocklist state for them.  The immediate motivation for this is to avoid
initializing the blocklist early in startup for the initial startup in a
new profile or after a browser upgrade when we're installing/udpating
system and builtin addons.

Differential Revision: https://phabricator.services.mozilla.com/D68443

UltraBlame original commit: a62adf9c09cc037382a7abb1549fcab85ca3ca90