Bug 1625892 - [raptor] Refactor webextension code for opening and closing tabs. r=perftest-reviewers,sparky
When awaiting the promise to be resolved, the returned Tab object
will contain the neccessary reference to the id. As such no separate
listeners have to be defined. Side-effect with listeners is that those
are async and could cause race conditions in the sync program flow.

Differential Revision: https://phabricator.services.mozilla.com/D69562

UltraBlame original commit: 57ad041ef4169609db138d427ab389a3dab6330b