Bug 1626893 part 6 - Tweak Memory-drainAllocationsLog-01.js. r=iain
This fixes the last test failure. The debugger can collect "allocation metadata"
for allocated objects. This test is checking that the objects in the array have
different allocation sites associated with them.

Ion can fail this test because MNewArray is marked as non-effectful so it doesn't
get its own resume point, resulting in the allocation site matching that of the
object literal before it.

We could fix this by making MNewArray* and similar instructions effectful, but
I'm concerned about the perf impact of changing that just for the unlikely event
the memory profiler is being used. Since this is a pre-existing issue, this tweaks
the test to work around it.

Differential Revision: https://phabricator.services.mozilla.com/D69506

UltraBlame original commit: a3bc77baa5196e70bc4d1b2ed5db1024a528c03a