Bug 1627174 - Add a Treeherder job for running jit-tests with WarpBuilder enabled. r=sfink,iain,jmaher
Note that it only runs jit-tests (not jsapi-tests and the slower jstests) and
only two jit-test configurations. This makes the job very fast, I see it finish
in just 12-13 minutes on Try.

When this lands it should be green. For now it's tier 2 and only runs on central.
When WarpBuilder gets closer to shipping we can consider changing that (and probably
run more tests).

Differential Revision: https://phabricator.services.mozilla.com/D69512

UltraBlame original commit: 0bd0baa8a6775af4be4afeb829b99704dc0b7853