Bug 1620145: Part 2 - Add AndroidProcessPriority.cpp to hal; r=gsvelto,geckoview-reviewers,snorp
This patch is pretty straightforward: it translates Gecko priority levels
into GeckoView priority levels and then sends it up to GV's
`GeckoProcessManager` via JNI.

We do assume that the process is content, but if we try to do that on a
non-content process, it's just a no-op. We can expand this coverage to other
process types later as necessary.

Differential Revision: https://phabricator.services.mozilla.com/D68418

UltraBlame original commit: 5390cddc67103d922061a5257c0d5c215af03660