Bug 1627734 - Make ZoomToFocusedInput skip fixed frames again. r=botond
We lifted this restriction because with bug 656036 it makes sense to zoom into a
focused input in a fixed subtree. But this causes undesired panning up.

We could conceivably fix it if needed, but for now preserve the previous
behavior of not panning nor zooming.

Differential Revision: https://phabricator.services.mozilla.com/D69946

UltraBlame original commit: b70c2c720c4d7aa6572f72fd62112d52c2dd2d71