Bug 1626286 - Assign the same timestamp to all WebConsole stubs. r=Honza.
Since we're going to re-order the packets in the webconsole messages
reducer based on the messages timestamps, the node tests are going
to fail, as stubs timeStamps are generated and we don't have much
control on them.
Assigning the same timestamp means we'll defer to order based on the
message id, which should follow the order of insertion. The only exception
is network messages, where the id needs to be the actor, which is a string.
For the tests that are using network messages and assert some order, we
simply reassign the timestamps so we can guarantee the order.

Differential Revision: https://phabricator.services.mozilla.com/D69966

UltraBlame original commit: cc342696a392e4d354daaf18419dd85b31ae42e8