Bug 1627634 [wpt PR 22716] - Improve the logging when a server fails to start., a=testonly
Automatic update from web-platform-tests
Improve the logging when a server fails to start.

Make sure we have a useful error message when the servers don't start,
and also update wptrunner to fail early in the case that the process
has exited rather than trying to reconnect to a server that isn't running
until we reach the timeout.

The patch contains a little blemish in that we always print to stderr from
the Websockets processes; the logger isn't correctly configured there and
fixing it is outside the scope of this patch (we've had regressions in
the past from that code, so I want to keep the scope as small as possible.

--

wpt-commits: 1383186a5d6d3313547f7105a8a57adf5f82d03d
wpt-pr: 22716

UltraBlame original commit: e9643eb3853f5484dddd758098d105792ebe3bf5