Bug 1627838: Factor out RegExpShared::executeAtom r=mgaudet
This code is the same for both the new and the old version.

We use `patternAtom` because the new version will add the ability to search for an atom that is not identical to the pattern source (for example, a pattern containing a unicode escape).

Differential Revision: https://phabricator.services.mozilla.com/D69895

UltraBlame original commit: 8ce2dbfa9755ac1fdde3366a9de2c3b4d31aa972