Bug 1513603 - Remove SVG tearoff objects from the table when they are unlinked. r=longsonr
The only three types that can hit the PreservedWrapper() assertion
currently, and which this patch avoids, are DOMSVGPathSegList,
DOMSVGPointList, and DOMSVGStringList.

This is because these are the only types that are implemented as proxies
(due to having e.g. indexed getters on them, and thus have the offending
DOMProxyHandler::GetExpandoObject() call in their bindings) and which we
store in tearoff tables.

Differential Revision: https://phabricator.services.mozilla.com/D70314

UltraBlame original commit: 39f8ceb567ec90338c4daea1d9dad48043206dad