Bug 1628379 - Make browsertime fail early if chromedriver doesn't exist. r=perftest-reviewers,AlexandruIonescu
Some phones aren't updating their google-chrome version so we are hitting issues where the chromedriver is not being found because it's looking for an older version. This patch makes the error occur earlier and prevents browsertime from running if we hit it.

Differential Revision: https://phabricator.services.mozilla.com/D70210

UltraBlame original commit: 91466c558e0498dfeeef1f5715e6e98b7b009458