Bug 1628994 - Don't render input related elements in browser console when devtools.chrome.enabled is not true. r=jlast.
When the pref is not set to true, we should not display the input,
but also the editor toolbar, the instant evaluation result and
the editor resizer.
The existing test is modified to ensure we cover all these elements.

Differential Revision: https://phabricator.services.mozilla.com/D70843

UltraBlame original commit: c00da8e9e19a68cf250a3fd0b8ca5843f5e3ee0f