Bug 1628626 - Fix issue when clearing output with messages from destroyed targets. r=jdescottes.
With the new architecture, it might happen that a message (and the ObjectFronts
it holds), are still displayed in the Browser Console / Browser Toolbox Console,
even if the target of those object fronts was destroyed.
In such case, when the user would legimitely try to clear the console, we'd try
to release the fronts that were already destroyed, which would throw an exception
and leave the console in a bad state.
This patch simply check that the fronts are still alive when we try to release
them, and adds a test (that was failing without that patch, with fission ON) for
the Browser Console.

Differential Revision: https://phabricator.services.mozilla.com/D70398

UltraBlame original commit: 6b974e1bf99af92f590363cdfa028e6dbd4e1118