Bug 1607984 - P11. Don't assume the page will be loaded synchronously. r=zombie
We must wait for the iframe created by extension1 to be fully loaded before we can attempt to send a message to it.
So we send a message when it is ready, and suspend the execution until this message is received.

Differential Revision: https://phabricator.services.mozilla.com/D70008

UltraBlame original commit: 9f216e9bd32ec708c8763e1517baba6ca241d226