Bug 1482147 - Simplify nsTypeAheadFind visibility code to not lie. r=masayuki
Multiple issues here. The IsRangeVisible code was wrong, it was returning false
for ranges that were perfectly valid, but outside the viewport, because the
following piece of code:

 if (!aMustBeInViewPort) {
   // This is an early exit case because we don't care that that range
   // is out of viewport, so we return that the range is "visible".
   return true;
 }

Was incorrectly after some stuff checking viewport visibility. This code is
pretty complex for no good reason, it wants to do something very
simple: Start from the visible selection if possible.

This patch still achieves this, using IsRangeRendered (which does a proper
hit-test to figure out if a range is in the viewport). Should have no behavior
differences except for non-collapsed ranges that are partially inside the
viewport.

Differential Revision: https://phabricator.services.mozilla.com/D71067

UltraBlame original commit: 6a15a1a873488084dd9fc9be1da317cd4b24cbe9