Bug 1630090: Refactor PrepareAndExecuteRegexp r=jandem
PrepareAndExecuteRegexp has to be updated to call the new engine. In preparation, this patch factors out a couple of helper functions that a) can be shared between the two versions, and b) make sense as standalone functions in the future when we remove the old implementation.

Differential Revision: https://phabricator.services.mozilla.com/D71080

UltraBlame original commit: 7efbecf663a804866d11a696f8f5f53f4f8adc2f