Bug 1630788 - Init FunctionHasExtraBodyVarScope during lazy parse. r=arai
Move the initialization of this flag to `finishFunctionScopes` so that it is
set for both lazy and non-lazy compilations. We add a `VarScopeHasBindings`
helper to mirror the logic of `newVarScopeData` without allocating. We use
the ImmutableFlag to enable consistent initialization for lazy and non-lazy
scripts.

Note that the previous `hasExtraBodyVarScope` helper is only used for
bytecode emission which occurs after the FunctionHasExtraBodyVarScope flag is
properly initialized.

Differential Revision: https://phabricator.services.mozilla.com/D71241

UltraBlame original commit: 983377c0fc2821a46bffbb9a02ef6dfe6a9eb554