Bug 1629791 part 5 - Auto-generate CacheIRWriter and CacheIRCompiler code for a handful of ops. r=iain
For ops with the gen_boilerplate attribute, this auto-generates the writer and compiler
boilerplate code. The attribute has been added to some arbitrary ops to test this.

CacheIRWriter methods that return a new OperandId are not supported yet.

Initially I wanted to generate multiple header files and classes, but the current
approach works fine and is probably a bit easier to reason about. Because the
code generation itself is pretty simple it should be easy to experiment with different
codegen strategies in the future.

Differential Revision: https://phabricator.services.mozilla.com/D71153

UltraBlame original commit: 667744615032982f90e8f4e5d38a0c403371a7d0