Bug 1630961 - Remove unused heurisitic to delaying GCs that may cause resets r=sfink
A while back we added a heuristic to delay a GC if it would cause a reset (bug 1367455), but it was turned off after it caused a performance regression and hasn't been reenabled since.

Telemetry shows that that only 0.2% of GCs get reset anyway so we should probably just remove this.

Depends on D71328

Differential Revision: https://phabricator.services.mozilla.com/D71329

UltraBlame original commit: 0aab8dc843cdb4c97a5ba848cd3d16a894fb66f4