Bug 1623987 - Fix front/back camera flipping on Android; r=ng
We plumb through front/back facing in the device name. When using the Camera1
API this is already present. With Camera2, it is not. Since we later use the
camera name when opening the camera, this prepends the front/back facing
information to the name, so it can easily be removed when we need to open
the camera.

Ideally, upstream would provide a field for this in VideoCaptureCapability and
none of this would be necessary.

Differential Revision: https://phabricator.services.mozilla.com/D71214

UltraBlame original commit: b034f792c13572d20648d895edcf44493c22cc0b