Bug 1631193 - drop Wayland permissions from flatpak r=mtabara
Firefox still does not default to Wayland when it's available (or when
X11 socket is not accessible). This can be changed by MOZ_ENABLE_WAYLAND=1
but as it's not the default, it's safer to just drop access to Wayland.

Differential Revision: https://phabricator.services.mozilla.com/D71448

UltraBlame original commit: fc4ad3fce1b79586adb4a2bfd00cfbf05d4e3634