Bug 1596812 Part 2 - NSIS WebBrowser plugin. r=agashlin,nalexander
This is all the code and build files for an NSIS plugin that enables
rendering a web page as the content of an NSIS dialog.

Documentation and the compiled binary are in later commits in this series.

Depends on D56576

Differential Revision: https://phabricator.services.mozilla.com/D56577

UltraBlame original commit: 1c80f6c6d43bc4e39fe11150ae2a92ac0bd45a1d