Bug 1630048 - Fix visual metrics encoding failures. r=perftest-reviewers,Bebe
This patch fixes a failure with encodings when opening JSONs. It changes the file opening portion to ignore these types of errors - they are occuring because of the different language being used. We have no need for this data so we can safely ignore them. One other change that is being done here is adding additional error logging because it was possible to lose an error stack trace.

Differential Revision: https://phabricator.services.mozilla.com/D70939

UltraBlame original commit: d41099e117db6dad99ccf93c6f9361be0cbaf854