Bug 1632467: Clear eGroupInfoDirty flag when creating AccGroupInfo. r=MarcoZ
When the tree is mutated, we set eGroupInfoDirty regardless of whether there is GroupInfo.
Previously, we cleared eGroupInfoDirty when updating GroupInfo, but not when creating it.
This meant that the flag was often still set, causing the cache to not be used a lot of the time.

Differential Revision: https://phabricator.services.mozilla.com/D72139

UltraBlame original commit: 6998e2f1e2d8cf71a1a72b1231bc6b6fadbf60b5