Bug 1625470 - Part 1: Add --live-sites command line option to Raptor. r=perftest-reviewers,Bebe,AlexandruIonescu
This is the first patch of a multi-part patch to enable live site testing on Google Chrome for Android and Fenix. It adds the `--live-sites` command line option that disables the mozproxy playback (or enables live sites) when it is supplied.

Differential Revision: https://phabricator.services.mozilla.com/D69051

UltraBlame original commit: baadcf329af6b4a21b5b74100cd1a42f02a005fb