Bug 1625864 - Expose unlinked a tags as AXGroup. r=morgan
This is the recommended role in the HTML Accessibility API Mappings draft[1].

Although Chrome and Safari just expose unlinked a tags as AXStaticText.
They also don't update the accessble if it gets an href. So while we
are doing this a bit differently, I think it is probably better. And VO is happy.

1. https://www.w3.org/TR/html-aam-1.0/#html-element-role-mappings

Differential Revision: https://phabricator.services.mozilla.com/D71260

UltraBlame original commit: 7f2fc9ae7d7bca8abb105bed9cbc7a07b5e446db