Bug 1628732 1628501 - disable antitracking mochtiests that fail permanently on window64 opt r=timhuang
Following tests are disabled
- browser_protectionsUI_cookies_subview.js
- browser_protectionsUI_state.js
- browser_allowListNotifications.js
- browser_urlDecorationStripping.js

The reason that this tests are failing in fission mode is because in
window64 opt mode, additional content blocking event is triggered when
loading chrome:://global/skin/icons/resizer.svg

We don't have this problem before because in this case, we couldn't get the
top-level window, so the event was not fired.
But the recent working to make anti-tracking fission compatible (Bug 1624552) removes
the need of getting top-level window, so the event is fired and
testcases don't expect that.

Differential Revision: https://phabricator.services.mozilla.com/D72601

UltraBlame original commit: 65b2d0679b0f44b561684ae0b8a7b659580db465