Bug 353575 - Allow IDTracker to look up elements in <svg:use> shadow trees. r=smaug
Other UAs allow this, and it seems in the general consensus of
https://github.com/w3c/webcomponents/issues/179.

This matches WebKit's behavior. Blink, for some reason shows red on the
test-case, probably because they're not doing quite this, but they
manage to render masks inside the display: none symbol element or such.

Differential Revision: https://phabricator.services.mozilla.com/D72610

UltraBlame original commit: 8efff752175e177108adceddb73a60f98f8cdd27