Bug 1633464 - Use a jump to enter Ion invalidation trampoline. r=jandem
On x86/x64 platforms, we were using a call and then immediately popping and
ignoring the return value, so use a jump instead. For architectures that use
a link-register, we no longer clobber it but it was already ignored.

Differential Revision: https://phabricator.services.mozilla.com/D72697

UltraBlame original commit: 2407418c84764d536b3607aeacc7a5153e8f7dbb